% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMetricToSeObj.R
\name{addMetricToSeObj}
\alias{addMetricToSeObj}
\title{Add results and plots of a metric to SummarizedExperiment object.}
\usage{
addMetricToSeObj(
  se.obj,
  slot = "Metrics",
  assay.names,
  assessment.type = "gene.level",
  assessment,
  method,
  variables,
  file.name,
  results.data
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{slot}{Symbol. A symbol indicating the name of the slots in the 'metadata' of the SummarizedExperiment object.}

\item{assay.names}{Symbol. A symbol or a vector of symbols specifying the name(s) of the assay(s) in the SummarizedExperiment
object. The default is set to 'all'.}

\item{assessment.type}{Symbol. A symbol indicating the type of assessment. Options are 'gene.level' or 'global.level'.}

\item{assessment}{Symbol. A symbol indicating the name of the metric to be checked.}

\item{method}{Symbol. A symbol indicating the method used to calculate the metric.}

\item{variables}{Symbol. A symbol or a vector of symbols indicating the variables used to calculate the metric.}

\item{file.name}{Symbol. A symbol indicating the file name to which the results of the metric are assigned.}

\item{results.data}{Symbol. A symbol indicating the name of the results file for the metric.}
}
\description{
Add results and plots of a metric to SummarizedExperiment object.
}
\author{
Ramyar Molania
}
