% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNCGs.R
\name{addNCGs}
\alias{addNCGs}
\title{Add pre-selected set of negative control genes to SummarizedExperiment object.}
\usage{
addNCGs(
  se.obj,
  ncg,
  assay.name = NULL,
  assess.ncg = FALSE,
  variables.to.assess.ncg = NULL,
  apply.log = TRUE,
  pseudo.count = 1,
  nb.pcs = 5,
  center = TRUE,
  scale = FALSE,
  assess.se.obj = TRUE,
  remove.na = "none",
  save.se.obj = TRUE,
  output.name = NULL,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{ncg}{A logical or vector of gene names of pre-selected genes as NCGs.}

\item{assay.name}{Symbol. Indicates a name of an assay in the SummarizedExperiment object.}

\item{assess.ncg}{Logical. Indicates whether to assess the performance of selected NCG or not. This involves principal
component analysis on the selected NCG and then explore the R^2 or vector correlation between the 'nb.pcs' first principal
components and with biological and unwanted variables.}

\item{variables.to.assess.ncg}{Symbols. Indicates the column names of the SummarizedExperiment object that
contain variables whose association with the selected genes as NCG needs to be evaluated. The default is NULL. This
means all the variables in the 'uv.variables' will be assessed.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data, by default it is set to TRUE.}

\item{pseudo.count}{Numeric. A value as a pseudo count to be added to all measurements before log transformation.}

\item{nb.pcs}{Numeric. Indicates the number of the first principal components on selected NCG to be used to assess the
performance of NCGs. The defaut is 5.}

\item{center}{Numeric. Indicates the number of the first principal components on selected NCG to be used to assess the
performance of NCGs. The defaut is 5.}

\item{scale}{Logical. Indicates whether to scale the data or not before applying SVD.  If scale is TRUE, then scaling
is done by dividing the (centered) columns of the assays by their standard deviations if center is TRUE, and the root
mean square otherwise. The default is FALSE.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment class object or not.}

\item{remove.na}{Symbol. Indicates whether to remove NA or missing values from either the 'assays', the 'sample.annotation',
'both' or 'none'. If 'assays' is selected, the genes that contains NA or missing values will be excluded. If 'sample.annotation' is selected, the
samples that contains NA or missing values for any 'uv.variables' will be excluded. By default, it is set to both'.}

\item{save.se.obj}{Symbol. Indicates}

\item{output.name}{Symbol. Indicates the name of the output in the meta data of the SummarizedExperiment object. The
default is 'NULL'. This means the function will create a name based the specified argument.}

\item{verbose}{Logical. If TRUE, displaying process messages is enabled.}
}
\description{
This function adds a pre-selected set of negative control genes to SummarizedExperiment object.
}
\author{
Ramyar Molania
}
