% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addOverallPlotToSeObj.R
\name{addOverallPlotToSeObj}
\alias{addOverallPlotToSeObj}
\title{Add overall plot to the "metadata" of SummarizedExperiment object.}
\usage{
addOverallPlotToSeObj(
  se.obj,
  slot = "Plots",
  assessment.type = "gene.level",
  assessment,
  method,
  variables,
  file.name,
  plot.data
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{slot}{Symbol. A symbol indicating the name of the slots in the 'metadata' of the SummarizedExperiment object.}

\item{assessment.type}{Symbol. A symbol indicating the type of assessment. Options are 'gene.level' or 'global.level'.}

\item{assessment}{Symbol. A symbol indicating the name of the metric to be checked.}

\item{method}{Symbol. A symbol indicating the method used to calculate the metric.}

\item{variables}{Symbol. A symbol or a vector of symbols indicating the variables used to calculate the metric.}

\item{file.name}{Symbol. A symbol indicating the file name to which the results of the metric are assigned.}

\item{plot.data}{Symbol. A symbol indicating the name of the plots file for the metric.}
}
\description{
Add overall plot to the "metadata" of SummarizedExperiment object.
}
\author{
Ramyar Molania
}
