% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_gene_exp_catvar.R
\name{anova_gene_exp_contvar}
\alias{anova_gene_exp_contvar}
\title{is used to compute the correlation between the gene expression (assay)
of a SummarizedExperiment class object and a continous variable (i.e. library size).}
\usage{
anova_gene_exp_contvar(
  se,
  cat_var,
  cat_var_label,
  assay_names = NULL,
  apply.log = FALSE,
  plot_output = TRUE,
  top.genes.no = 3
)
}
\arguments{
\item{se}{A SummarizedExperiment object that will be used to compute the correlation}

\item{cat_var}{Vector of a categorical variable such as sample types
(i.e. biological subtypes) or batches.}

\item{cat_var_label}{String or vector of strings of the label of categorical variable(s) such as
sample types or batches from colData(se).}

\item{assay_names}{Optional string or list of strings for selection of the names
of the assays of the SummarizedExperiment class object to compute the correlation.}

\item{apply.log}{Indicates whether to apply a log-transformation to the data. By default
no transformation will be selected.}

\item{plot_output}{Indicates whether to plot the results}

\item{top.genes.no}{Defines the number of top genes, by default is set to 3.}
}
\description{
is used to compute the correlation between the gene expression (assay)
of a SummarizedExperiment class object and a continous variable (i.e. library size).
}
