% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyOtherNormalizations.R
\name{applyOtherNormalizations}
\alias{applyOtherNormalizations}
\title{Perform several normalization methods for RNA-seq data.}
\usage{
applyOtherNormalizations(
  se.obj,
  assay.name,
  method = "CPM",
  apply.log = TRUE,
  pseudo.count = 1,
  remove.na = "assays",
  assess.se.obj = TRUE,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A summarized experiment object.}

\item{assay.name}{Symbol. A symbol that indicates the name of the assay in the SummarizedExperiment object. The specified assay
should be in raw count format.}

\item{method}{Symbol. Indicates the normalization method to apply on the 'assay.name'. Options are:
\itemize{
\item 'CPM': Count Per Million from the edger R package.
\item 'TMM': Trimmed Mean of M-values from the edger R package,
\item 'upper':a scaling normalization that forces the median of each lane to be the same from the EDAseq R package.
\item 'median': a scaling normalization that forces the upper quartile of each lane to be the same from the EDAseq R package.
\item 'full': a non linear full quantile normalization from the EDAseq R package
\item 'VST', variance stabilizing transformation from the DESeq2 R package.
The default is set to 'CPM'.
}}

\item{apply.log}{Logical. A symbole that indicates whether to apply a log-transformation to the data or not after the
normalization is performed. The default is 'TRUE'.}

\item{pseudo.count}{Numeric. A numeric value as a pseudo count to be added to all measurements of the assay(s) after
applying the normalization to avoid -Inf for measurements that are equal to 0. The default is set to 1.}

\item{remove.na}{Symbol. A symbol that indicates whether to remove NA or missing values from the assay or not. Refer
to the 'checkSeObj' function for more details. The defualt is set to assays.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment class object. The default it is set to
TRUE.}

\item{save.se.obj}{Logical. Indicates whether to save the normalized data as new assay(s) to the SummarizedExperiment
object. The default it is set to 'TRUE'. If set to 'FALSE', the normalized data(s) will be outputted as a matrix.}

\item{verbose}{Logical. Indicates whether to show or reduce the level of output or messages displayed
during the execution of the functions, by default it is set to TRUE.}
}
\value{
A SummarizedExperiment object containing an assay normalized by the selected normalization method or just
return a normalized gene expression matrix.
}
\description{
This functions provides different normalization methods including CPM, TMM, upper-quartile, non linear full quantile
normalization, median, and VST for RNA-seq
data.
}
\author{
Ramyar Molania
}
