% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ari_catvar_all_assays.R
\name{ari_catvar_all_assays}
\alias{ari_catvar_all_assays}
\title{is used to compute the adjusted rank index (ARI) from the first PC of all assays
given a categorical variable}
\usage{
ari_catvar_all_assays(pca, normalization, cat_var, nPCs = 3)
}
\arguments{
\item{pca}{PCs of the dataset that will be used}

\item{normalization}{All the available assays for the data (i.e. normalizations methods)}

\item{cat_var}{is a categorical variable such as sample types or batches}

\item{nPCs}{is the number of PCs used to measure the distance}
}
\value{
list List containing the association plot and the computed ari
}
\description{
is used to compute the adjusted rank index (ARI) from the first PC of all assays
given a categorical variable
}
