% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ari_catvar_single_assay.R
\name{ari_catvar_single_assay}
\alias{ari_catvar_single_assay}
\title{is used to compute the adjusted rank index (ARI) from the first PC of a single assay
given a categorical variable}
\usage{
ari_catvar_single_assay(select_pca, cat_var, nPCs = 3)
}
\arguments{
\item{select_pca}{PCs of the dataset that will be used}

\item{cat_var}{is a categorical variable such as sample types or batches}

\item{nPCs}{is the number of PCs used to measure the distance}
}
\value{
a single value, the ari computed
}
\description{
is used to compute the adjusted rank index (ARI) from the first PC of a single assay
given a categorical variable
}
