% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessNCGs.R
\name{assessNCGs}
\alias{assessNCGs}
\title{Assess the different sets of NCGs.}
\usage{
assessNCGs(
  se.obj,
  assay.name,
  variables,
  bio.variables = NULL,
  uv.variables = NULL,
  ncg.type,
  ncg.group,
  ncg.set.names = "all",
  nb.pcs = 5,
  center = TRUE,
  scale = FALSE,
  apply.log = TRUE,
  pseudo.count = 1,
  plot.output = TRUE,
  output.name = NULL,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A summarized experiment object.}

\item{assay.name}{Symbol. A symbol indicating the name of the assay in the SummarizedExperiment object. This assay
should be the one that will be used as an input data for the RUV-III-PRPS normalization.}

\item{variables}{Symbol. A symbol or a set of symbols indicating the names of the variables in the SummarizedExperiment
object. These will be used to assess the performance of NCGS.}

\item{bio.variables}{Symbol. A symbol or symbols representing the label of biological variable(s), such as cancer
subtypes, tumor purity, ... within the SummarizedExperiment object. This can comprise a vector containing either
categorical, continuous, or a combination of both variables.}

\item{uv.variables}{Symbol. A symbol or symbols representing the label of unwanted variable(s), such as cancer
batch effects, library size, ... within the SummarizedExperiment object. This can comprise a vector containing either
categorical, continuous, or a combination of both variables.}

\item{ncg.type}{Symbol. A symbol indicating the type of NCGs that are obtained either by supervised or un.supervised
approaches. The options are 'supervised' or 'un.supervised'.}

\item{ncg.group}{Symbol. A symbol specifying the name of a group of NCGs under the supervised or un.supervised slots
in the SummarizedExperiment object.}

\item{ncg.set.names}{A symbol specifying th exact name of a NCG set under the 'ncg.group' name in the supervised or
un.supervised slots in the SummarizedExperiment object. The default is set to 'all'. This will select all avaiable
NCGs sets.}

\item{nb.pcs}{Numeric. A numeric value indicating the number of PCs to be computed and used for the assessment analysis.
The default is set to 5.}

\item{center}{Logical. Indicates whether to scale the data or not prior to apply SVD. If center is 'TRUE', then
centering is performed by subtracting the column means of the assay from their corresponding columns. The default is 'TRUE'.}

\item{scale}{Logical. Indicates whether to scale the data or not before applying SVD.  If scale is 'TRUE', then scaling
is done by dividing the (centered) columns of the assays by their standard deviations if center is 'TRUE', and the root
mean square otherwise. The default is FALSE.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data before computing the SVD. The
default is 'TRUE'. The data must be in log scale before computing the SVD.}

\item{pseudo.count}{Numeric. A value as a pseudo count to be added to all measurements before applying log transformation.
The default is 1. This argument cannot be NULL or negative.}

\item{plot.output}{Logical. Indicates whether to plot the assessment plot or not while running the function. The default
is set to 'TRUE'.}

\item{output.name}{Symbol. A symbol indicating the name under which the assessment plot will be saved in the
SummarizedExperiment object. If 'NULL', the function will select the name as follow:
'paste0('comparsion_', paste0(names(ncgs), collapse = '&'))'}

\item{save.se.obj}{Logical. Indicates whether save the assessment plot to the SummarizedExperiment object or not. The
default is set to 'TRUE'. The plot will be save to:}

\item{verbose}{Logical. Indicates whether to plot the output or not. The default is set to 'TRUE'.}
}
\description{
This function assesses the performance of various sets of negative control genes in terms of their ability to capture
unwanted variation and their lack of association with biological variation. First, for each individual set of negative
control genes (NCG), the function applies PCA to the specified data. Then, it calculates the association of the 'nb.pcs'
principal components (PCs) with both specified biological and unwanted variables. The correlation values obtained for
the biological variables will be subtracted from 1. Finally, the average of all these correlations is computed to
determine the final correlation values. A higher correlation indicates better performance.
}
\author{
Ramyar Molania
}
