% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessNormalization.R
\name{assessNormalization}
\alias{assessNormalization}
\title{Assess the performance of normalization methods.}
\usage{
assessNormalization(
  se.obj,
  assay.names = "all",
  bio.variables,
  uv.variables,
  assessments.to.exclude = NULL,
  fast.pca = TRUE,
  sil.dist.measure = "euclidian",
  ari.clustering.method = "hclust",
  ari.hclust.method = "complete",
  ari.hclust.dist.measure = "euclidian",
  corr.method = "spearman",
  corr.cutoff = 0.2,
  anova.method = "aov",
  fvalue.cutoff = 1,
  pcorr.method = "spearman",
  pcorr.cutoff = 0.3,
  bio.weight = 0.6,
  uv.weight = 0.4,
  plot.output = TRUE,
  save.se.obj = TRUE,
  output.name = NULL,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Symbol. A symbol or a vector of symbols for the selection of the name(s) of the assay(s) in the
SummarizedExperiment object. The default is set to 'all', so all the assays in the SummarizedExperiment object will
be selected.}

\item{bio.variables}{Symbols. A symbol or a vector of symbols indicating the column names that contain known biological
variable(s) in the SummarizedExperiment object. These biological variables can be both categorical and continuous.}

\item{uv.variables}{Symbols. A symbol or a vector of symbols indicating the column names that contain unwanted variable(s)
in the SummarizedExperiment object. These unwanted variables can be both categorical and continuous.}

\item{assessments.to.exclude}{Symbol. A symbol or vector of symbols (Code) specifying which assessment metrics to exclude.
These metrics can be acquired using the 'getAssessmentMetrics' function. Default is set to 'NULL'.}

\item{fast.pca}{Logical. Indicates whether to obtain computed fast PCA or not. The default is set to 'TRUE'.}

\item{sil.dist.measure}{Symbol. A symbol indicating which distant measurement has been used to compute silhouette analysis
should be obtained. The options are: 'euclidean', 'maximum', 'manhattan', 'canberra', 'binary' or 'minkowski'.
The default is set to 'euclidean'.}

\item{ari.clustering.method}{Symbol. A symbol indicating which clustering method has been used to compute ARI should
be used. The options are 'mclust' or 'hclust' methods. The default is set to 'hclust'.}

\item{ari.hclust.method}{Symbol. A symbol indicating which hclust method has been used to compute ARI. The options are
'ward.D', ward.D2', 'single', 'complete', 'average' (= UPGMA), 'mcquitty' (= WPGMA), 'median' (= WPGMC) or centroid'
(= UPGMC). The default is set to 'complete'.}

\item{ari.hclust.dist.measure}{Symbol. A symbol indicating which distant measurement has been used to compute 'hclsut'
for the ARI analysis. The options are euclidean', 'maximum', 'manhattan', 'canberra', 'binary' or 'minkowski'. The
default is set to 'euclidean'.}

\item{corr.method}{Symbol. A symbol indicates which correlation methods has been used to compute genes-variable correlation.
Options are 'spearman' or 'pearson'. The default is 'spearman'.}

\item{corr.cutoff}{Numeric. A numeric values that specifies the threshold for selecting genes that have an absolute
correlation coefficient lower than this value with continuous variables. The default is set to 0.2.}

\item{anova.method}{Symbol. A symbol indicating which ANOVA method has been used for compute genes-variable ANOVA. Options
are 'aov' or 'weltch'. The default is set to 'aov'.}

\item{fvalue.cutoff}{Numeric. Specifies the threshold for selecting genes that have a F-statistics lower than this value
with categorical variables. The default value is set to 1.}

\item{pcorr.method}{A symbol indicating which correlation method has been used for compute genes-gene ordinary and
partial correlations. Options are 'spearman' or 'pearson'. The default is set to 'spearman'.}

\item{pcorr.cutoff}{Numeric. A threshold value used to select gene-gene correlations where the difference between
ordinary and partial correlations is less than this value.}

\item{bio.weight}{Numeric. A numeric value to weight the final biological preservation scores. The default is set 0.6.
Refer to the detail for more information.}

\item{uv.weight}{Numeric. A numeric value to weight the final removal of unwanted variation scores. The default is set 0.4.
Refer to the detail for more information.}

\item{plot.output}{Logical. If 'TRUE', the final assessment plot will printed.}

\item{save.se.obj}{Logical. Indicates whether to save all the results to the SummarizedExperiment. The default is set
to 'TRUE'.}

\item{output.name}{Symbol. A symbol indicating the file name for the results in the SummarizedExperiment. If 'NULL',
the function will create "paste0('RUVprps_assessNormalization', length(assays(se.obj)), '_assays.')" as a file name.}

\item{verbose}{Logical. If 'TRUE', shows the messages of different steps of the function.}
}
\value{
A SummarizedExperiment object containing all the assessments matrix, plot and table or a list that contains all
the results.
}
\description{
This function summarizes a range of global and gene level metrics obtained using the 'assessVariation' function for
individual biological and unwanted variables. The functions returns numerical assessments in a table to assess the
performance of difference normalization. Refer to details for more information.
}
\details{
It is essential to assess the performance of a normalization as a remover of unwanted variation and a preserver of
biological variation in the data. An ideal normalization should preserve all known sources of biological variation
and reduce or remove the impact of all known sources of unwanted variation. To rank methods according to their
performance, we aggregate all evaluation metrics into a overall score as follows.
Several assessment will be performed:
For each categorical variable:
\itemize{
\item PCA plot of the categorical variable.
\item Silhouette and ARI computed on the categorical variable.
\item Differential analysis based ANOVA between the gene expression and the categorical variable.
\item Vector correlation between the first cumulative PCs of the gene expression and the categorical variable.
For each continous variable:
\item Linear regression between the first cumulative PC and continuous variable.
\item Correlation between gene expression and continuous variable.
}

It will output the following plots:
\itemize{
\item PCA plot of each categorical variable.
\item Boxplot of the F-test distribution from ANOVA between the gene expression and each categorical variable.
\item Vector correlation between the first cumulative PCs of the gene expression and each categorical variable.
\item Combined Silhouette plot of the combined pair of all categorical variables.
\item Linear regression between the first cumulative PC and continuous variable.
\item Boxplot of the correlation between gene expression and continuous variable.
\item It will also output the RLE plot distribution.
}
}
\references{
**Molania R., ..., Speed, T. P., A new normalization for Nanostring nCounter gene expression data, Nucleic Acids Research,
2019.
**Molania R., ..., Speed, T. P., Removing unwanted variation from large-scale RNA sequencing data with PRPS,
Nature Biotechnology, 2023
}
\author{
Ramyar Molania
}
