% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessVariablesAssociation.R
\name{assessVariablesAssociation}
\alias{assessVariablesAssociation}
\title{assess the association between variables.}
\usage{
assessVariablesAssociation(
  se.obj,
  bio.variables,
  uv.variables,
  cat.cor.coef = c(0.9, 0.9),
  cont.cor.coef = c(0.9, 0.9),
  assess.se.obj = TRUE,
  remove.na = "sample.annotation",
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{bio.variables}{Symbol. A symbol or a vector of symbols specifying the column names of biological variables in
the sample annotation of the SummarizedExperiment object. These 'bio.variables' can be either categorical or continuous
variables.}

\item{uv.variables}{Symbol. A symbol or a vector of symbols specifying the column names of unwanted variables in
the sample annotation of the SummarizedExperiment object. These 'uv.variables' can be either categorical or continuous
variables.}

\item{cat.cor.coef}{Vector. A vector of two numerical values. Indicates the cut-off of the correlation coefficient
between each pair of categorical variables. The first one is between each pair of 'uv.variables' and the second one
is between each pair of 'bio.variables'. The correlation is computed by the function ContCoef from the DescTools
package. If the correlation of a pair of variable is higher than the cut-off, then only the variable that has the
highest number of factor will be kept and the other one will be excluded from the remaining analysis. By default they
are both set to 0.9.}

\item{cont.cor.coef}{Vector of two numerical values. Indicates the cut-off of the Spearman correlation coefficient
between each pair of continuous variables. The first one is between each pair of 'uv.variables' and the second one is
between each pair of 'bio.variables'. If the correlation of a pair of variable is higher than the cut-off, then only
the variable that has the highest variance will be kept and the other one will be excluded from the remaining analysis.
By default they are both set to 0.9.}

\item{assess.se.obj}{Logical. Whether to assess the SummarizedExperiment object or not. If 'TRUE', the function
checkSeobj will be applied.}

\item{remove.na}{Symbol. Indicates whether to remove missing values from the 'bio.variables' and 'uv.variables'. The
options are 'sample.annotation' or 'none'. The default is 'sample.annotation', indicating the missing values from the
variables will be removed.}

\item{verbose}{Logical. If 'TRUE', shows the messages of different steps of the function.}
}
\value{
A list that contains the SummarizedExperiment object and the selected 'uv.variables' and 'bio.variables'.
}
\description{
The function assesses the association between all selected biological and unwanted variation variables separately.
If two categorical variables are highly correlated, the function keeps one that has the highest number of factors.
For two highly correlated continuous variables, the one with higher variance will be kept. This assessment could be
useful for creating PRPS and finding NGCs. If two variables exhibit high correlation, utilizing one of them is
sufficient for PRPS or identifying NCGs.
}
\details{
For each pair of categorical variables from 'uv.variables' and each pair of categorical variables from 'bio.variables',
the correlation is computed using the function 'ContCoef' from the DescTools R package. For each pair of continuous
variables from 'uv.variables' and each pair of continuous variables from 'bio.variables', the correlation is computed
using the Spearman correlation. The user defines a minimum cut-off of the correlation coefficient between each pair of
categorical variables in the 'cat.cor.coef' for the unwanted variables, followed by the minimum cut-off of the correlation
coefficient for the biological variables. If the correlation of a pair of those variables is higher than the minimum
cut-off, only the variable that has the highest number of factor will be kept and the other one will be excluded from
the remaining analysis. The user defines a minimum cut-off of the correlation coefficient between each pair of continuous
variables in the 'cont.cor.coef' for the unwanted variables, followed by the minimum cut-off of the correlation coefficient
for the biological variables. If the correlation of a pair of variable is higher than the minimum cut-off, only the variable
that has the highest variance will be kept and the other one will be excluded from the remaining analysis.
}
\author{
Ramyar Molania
}
