% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ari.R
\name{compute_ari}
\alias{compute_ari}
\title{is used to compute the adjusted rank index (ARI) from the first PC of a SummarizedExperiment class
object given a categorical variable. It can be used to assess how a group of biological samples
are distributed across batches (i.e. example subtypes vs batch), and how batches ### update}
\usage{
compute_ari(pca, cat_var, assay_names = NULL, plot = FALSE, nPCs = 3)
}
\arguments{
\item{pca}{PCA components of a SummarizedExperiment variable.}

\item{cat_var}{Vector of a categorical variable such as sample types (i.e. biological subtypes) or batches.}

\item{assay_names}{Optional string or list of strings for selection of the names
of the assays of the SummarizedExperiment class object to compute the ARI.}

\item{plot}{Optional output of a plot, default set to FALSE.}

\item{nPCs}{is the number of PCs used to measure the distance, default is set to 3.}
}
\value{
list List containing the association plot and the computed ari
}
\description{
is used to compute the adjusted rank index (ARI) from the first PC of a SummarizedExperiment class
object given a categorical variable. It can be used to assess how a group of biological samples
are distributed across batches (i.e. example subtypes vs batch), and how batches ### update
}
