% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pca.R
\name{compute_pca}
\alias{compute_pca}
\title{compute_pca

is used to compute PCA of the data}
\usage{
compute_pca(sce, apply.log = FALSE)
}
\arguments{
\item{sce}{Dataset that will be used to compute the PCA}
}
\value{
pca.all PCA components computed on the provided data
}
\description{
compute_pca

is used to compute PCA of the data
}
