% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pca_single_assay.R
\name{compute_pca_single_assay}
\alias{compute_pca_single_assay}
\title{is used to compute PCA of a single assay}
\usage{
compute_pca_single_assay(sce, apply.log = FALSE)
}
\arguments{
\item{sce}{Dataset that will be used to compute the PCA}

\item{apply.log}{Indicates whether to apply a log-transformation to the data}
}
\value{
pca List containing the svd and the total variation explained by the PCA components computed on the single data
}
\description{
is used to compute PCA of a single assay
}
