% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAri.R
\name{computeAri}
\alias{computeAri}
\title{is used to compute the adjusted rank index (ARI) from the first PC of a SummarizedExperiment class
object given a categorical variable.}
\usage{
computeAri(
  se.obj,
  assay.names = "All",
  variable,
  fast.pca = TRUE,
  nb.pcs = 3,
  save.se.obj = TRUE,
  plot.output = FALSE,
  assess.se.obj = TRUE,
  apply.round = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object that will be used to compute the PCA.}

\item{assay.names}{Optional string or list of strings for the selection of the name(s)
of the assay(s) of the SummarizedExperiment class object to compute the ARI. By default}

\item{variable}{String of the label of a categorical variable such as
sample types or batches from colData(se.obj).}

\item{fast.pca}{TO BE DEFINED.}

\item{nb.pcs}{TO BE DEFINED.}

\item{save.se.obj}{Indicates whether to save the result in the metadata of the SummarizedExperiment class object 'se.obj' or
to output the result. By default it is set to TRUE.}

\item{plot.output}{Indicates whether to plot the ARI, by default it is set to FALSE.}

\item{assess.se.obj}{Indicates whether to assess the SummarizedExperiment class object.}

\item{apply.round}{TO BE DEFINED.}

\item{verbose}{Indicates whether to show or reduce the level of output or messages displayed during the execution
of the functions, by default it is set to TRUE.}
}
\value{
SummarizedExperiment A SummarizedExperiment object containing the computed ARI
on the categorical variable.
}
\description{
It can be used to assess how a group of biological samples
are distributed across batches (i.e. example subtypes vs batch), and how batches ### update
}
