% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeGenesPartialCorrelation.R
\name{computeGenePartialCorrelation}
\alias{computeGenePartialCorrelation}
\title{Compute gene-gene partial correlation.}
\usage{
computeGenePartialCorrelation(
  se.obj,
  assay.names,
  genes = NULL,
  variable,
  apply.log = TRUE,
  pseudo.count = 1,
  method = "spearman",
  apply.round = TRUE,
  plot.output = TRUE,
  assess.se.obj = TRUE,
  remove.na = "none",
  output.name = NULL,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Symbol. A symbol or a vector of symbols for the selection of the name(s) of the assay(s) in the
SummarizedExperiment object to calculate RLE data, medians and interquartile ranges. The default is set to "all, which
indicates all the assays of the SummarizedExperiment object will be selected.}

\item{genes}{vector.}

\item{variable}{Symbol. A symbol that indicates the column name of the SummarizedExperiment object that contains a
continuous variable such as library size, tumor purity, ... .}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data or not. The default is TRUE.}

\item{pseudo.count}{Numeric. A value as a pseudo count to be added to all measurements of the assay(s) before applying
log transformation to avoid -Inf for measurements that are equal to 0. The default is 1.}

\item{method}{Symbol. A symbol that indicates which correlation methods should be used. The options are 'pearson',
'kendall', or "spearman". The default is set to 'spearman'.}

\item{apply.round}{Logical. Indicates whether to round the correlations coefficients or not. The default is set to 'TRUE'.}

\item{plot.output}{Logical. Indicates to print the histogram of partial correlation coefficients or not. The default is
set to 'FLASE'.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment object or not. See the checkSeObj
function for more details.}

\item{remove.na}{Symbol. To remove NA or missing values from the assays or not. The options are 'assays' and 'none'.
The default is "assays", so all the NA or missing values from the assay(s) will be removed before computing RLE. See
the checkSeObj function for more details.}

\item{output.name}{= Symbol.}

\item{save.se.obj}{Logical. Indicates whether to save the RLE results in the metadata of the SummarizedExperiment object or
to output the result as list. By default it is set to TRUE.}

\item{verbose}{Logical. If 'TRUE', shows the messages of different steps of the function.}
}
\description{
This function computes all gene-gene pairwise correlation of the assays in the SummarizedExperiment object.
}
\details{
This function first calculates thte or...
}
\author{
Ramyar Molania
}
