% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeGeneSetScore.R
\name{computeGeneSetScore}
\alias{computeGeneSetScore}
\title{Compute sample wise score of a gene set.}
\usage{
computeGeneSetScore(
  se.obj,
  assay.names = "all",
  upset.genes,
  downset.genes = NULL,
  apply.log = TRUE,
  pseudo.count = 1,
  normalization = NULL,
  regress.out.variables = NULL,
  assess.score = FALSE,
  variables.to.assess = NULL,
  corr.method = "spearman",
  gene.set.name = NULL,
  plot.output = TRUE,
  assess.se.obj = TRUE,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Symbol or a vector of symbols specifying the name(s) of the assay(s) in the SummarizedExperiment
object to calculate sample-wise scores. The default is set to "all," indicating all assays in the SummarizedExperiment
object will be selected.}

\item{upset.genes}{Vector. A character vector of gene names/ids that are up-regulated in the a gen set.}

\item{downset.genes}{Vector. A character vector of gene names/ids that are down-regulated in the a gen set.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data or not. The default is 'TRUE'.}

\item{pseudo.count}{Numeric. A value as a pseudo count to be added to all measurements of the assay(s) before applying
log transformation to avoid -Inf for measurements that are equal to 0. The default is 1.}

\item{normalization}{Symbol. A symbol indicating the name of normalization to be used before computing the scores. The
default is set to 'NULL'. Refer to the 'applyOtherNormalizations' function for more details.}

\item{regress.out.variables}{Symbol or vector of symbols indicating the name(s) of the column(s) in the SummarizedExperiment
object to be regressed out from the data before computing the scores. The default is set to 'NULL'.}

\item{assess.score}{Logical. If 'TRUE', the association between the computed scores and specified variables will be
assessed. The default is set to 'NULL'. See the details for more information.}

\item{variables.to.assess}{Symbol. A symbol or vector of symbols indicating the name(s) of the column(s) in the SummarizedExperiment
object to be used for the assessment of the computed scores. It can comprise of continuous and categorical variables.
The default is set to 'NULL'.}

\item{corr.method}{Symbol. A symbol indicating which correlation method should be used for the correlation analysis of
the computed scores with the specified continuous variable(s). The options are "pearson", "kendall", "spearman". The
default is set to 'spearman'.}

\item{gene.set.name}{Symbol. A symbol indicating the name to be used to save the score in the SummarizedExperiment object.
if 'NULL', the function will select a name as follow :
gene.set.name <- paste0('singscore|',length(c(upset.genes, downset.genes)),'genes')}

\item{plot.output}{Logical. If 'TRUE', the assessment plot will be printed while running the function. The default is
set to 'TRUE'.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment object or not. The default is set
to 'TRUE'. Refer to the 'checkSeObj' function for more details.}

\item{save.se.obj}{Logical. Indicates whether to save the score results in the metadata of the SummarizedExperiment object
or to output the result as list. The default is set to 'TRUE'.}

\item{verbose}{Logical. If 'TRUE', shows the messages of different steps of the function.}
}
\description{
The function uses the singscore function from the R/Bioconductor singscore to calculate sample wise score of
a gene set in for all the assay(s) in the SummarizedExperiment object. Refer to the details for more information.
}
\author{
Ramyar Molania
}
