% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeGenesVariableAnova.R
\name{computeGenesVariableAnova}
\alias{computeGenesVariableAnova}
\title{Computes ANOVA between individual gene expression and a categorical variable.--Finalized}
\usage{
computeGenesVariableAnova(
  se.obj,
  assay.names = "all",
  variable,
  method = "aov",
  apply.log = TRUE,
  pseudo.count = 1,
  plot.top.genes = FALSE,
  nb.top.genes = 3,
  apply.round = TRUE,
  assess.se.obj = TRUE,
  remove.na = "both",
  save.se.obj = TRUE,
  override.check = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Character. A character string or vector of character strings specifying the name(s) of the assay(s)
in the SummarizedExperiment object to compute the ANOVA. By default is set to 'all', which means all assays of the
SummarizedExperiment object will be selected.}

\item{variable}{Character. A character string indicating a column name in the sample annotation of the SummarizedExperiment
object that contains a categorical variable, such as experimental batches, etc.}

\item{method}{Character. A character string specifying the method to be used for the ANOVA analysis. The options are
"aov" and "welch". The default is set to "aov". The function applies a fast implementation of one-way ANOVA using the
row_oneway_equalvar() or the row_oneway_welch() functions from the matrixTests R package. The first function assumes
equal variance between the groups while the second function applied the Welch correction for the groups with unequal
variances.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data before performing ANOVA. The
default is set to 'TRUE'.}

\item{pseudo.count}{Numeric. A numeric value representing a pseudo count to be added to all measurements before applying
the log transformation. The default is set to 1.}

\item{plot.top.genes}{Logical. Indicates whether to plot the gene expression of the "nb.top.genes" genes, as visual
examples, with the highest and lowest F-statistics across the specified variable. The default is set to 'FALSE'.}

\item{nb.top.genes}{Numeric. A numeric value specifying the number of genes to plot from the top or bottom of the ANOVA
F-statistics list. The default is set to 3.}

\item{apply.round}{Logical. Indicates whether to round the values of F-statistics. The default is set to 'TRUE'.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment object. The default is set to 'TRUE'.
This means the function will apply the 'checkSeObj' function.}

\item{remove.na}{Character. A character string specifying whether to eliminate missing values from 'assays', 'sample.annotation',
'both', or 'none'. When 'assays' is chosen, genes with missing values will be omitted. If 'sample.annotation' is selected,
samples with NA or missing values for each 'variable' will be excluded. The default is 'both'.}

\item{save.se.obj}{Logical. Indicates whether to save the results, ANOVA F-statistics, and p-values in the metadata
of the SummarizedExperiment object or to output these results as a list or vector. The default is set to 'TRUE'.}

\item{override.check}{Logical. When set to TRUE, the function checks whether ANOVA has already been computed for the
current parameters on the SummarizedExperiment object. If it has, the metric will not be recalculated. The default is
set to 'FALSE'.}

\item{verbose}{Logical. If TRUE, displays the messages of different steps of the function.}
}
\value{
Either a SummarizedExperiment object containing the log2 F-statistics and p-values of ANOVA for the continuous
variable or a list of these results.
}
\description{
This function calculates the ANOVA between individual gene expression of the assay(s) in a SummarizedExperiment object
and a categorical variable as factor.
}
\details{
ANOVA enables us to assess the effects of a given qualitative variable (which we call a factor) on gene expression
measurements across any set of groups (labeled by the levels of the factor) under study. We use ANOVA F-statistics
to summarize the effects of a qualitative source of unwanted variation (for example, batches) on the expression levels
of individual genes, where genes having large F-statistics are deemed to be affected by the unwanted variation.
We also use ANOVA tests (the aov() function in R) to assign P values to the association between tumor purity and
molecular subtypes.
}
\author{
Ramyar Molania
}
