% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePCA.R
\name{computePCA}
\alias{computePCA}
\title{is used to perform principal component analysis (PCA) using singular value decomposition (SVD).}
\usage{
computePCA(
  se.obj,
  assay.names = "all",
  fast.pca = TRUE,
  nb.pcs = 10,
  center = TRUE,
  scale = FALSE,
  apply.log = TRUE,
  pseudo.count = 1,
  bsparam = bsparam(),
  assess.se.obj = TRUE,
  remove.na = "assays",
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Symbol. A symbol or list of symbols for the selection of the name(s) of the assay(s) in the
SummarizedExperiment object to calculate RLE data, medians and interquartiles. The default is "all, which indicates
all the assays of the SummarizedExperiment object will be selected.}

\item{fast.pca}{Logical. Indicates whether to calculate a specific number of left singular vectors instead of the
full possible vectors to speed up the process. The default is 'TRUE'.}

\item{nb.pcs}{Numeric. The number of first left singular vectors to be calculated for the fast PCA process. The
default is 10.}

\item{center}{Logical. Indicates whether to scale the data or not. If center is TRUE, then centering is done by
subtracting the column means of the assay from their corresponding columns. The default is TRUE.}

\item{scale}{Logical. Indicates whether to scale the data or not before applying SVD.  If scale is TRUE, then scaling
is done by dividing the (centered) columns of the assays by their standard deviations if center is TRUE, and the root
mean square otherwise. The default is FALSE.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data before computing the SVD. The
default is 'TRUE'. The data must be in log scale before computing the SVD.}

\item{pseudo.count}{Numeric. A value as a pseudo count to be added to all measurements before applying log transformation.
The default is 1. This argument cannot be NULL or negative.}

\item{bsparam}{A BiocParallelParam object specifying how parallelization should be performed. The default is bsparam().
We refer to the 'runSVD' function from the BiocSingular R package.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment class object. The default is TRUE.
We refer to the checkSeObj function for more details.}

\item{remove.na}{To remove NA or missing values from the assays or not. The options are 'assays' and 'none'.}

\item{save.se.obj}{Logical. Indicates whether to save the SVD results in the metadata of the SummarizedExperiment object
or to output the results as list. By default it is set to 'TRUE'.}

\item{verbose}{Logical. If TRUE, displaying process messages is enabled.}
}
\value{
A SummarizedExperiment object or a list that containing the singular value decomposition results and the
percentage variation of each PCs.
}
\description{
This function uses singular value decomposition to perform principal component on the assay(s) in a SummarizedExperiment
object. The function provides fast singular value decomposition using the BiocSingular R package.
}
\details{
The PCs (in this context also called singular vectors) of the sample × transcript array of log counts are the linear
combinations of the transcript measurements having the largest, second largest, third largest, etc., variation,
standardized to be of unit length and orthogonal to the preceding components. Each will give a single value for
each sample.
}
\author{
Ramyar Molania
}
