% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePCVariableCorrelation.R
\name{computePCVariableCorrelation}
\alias{computePCVariableCorrelation}
\title{Computes the vector correlation.}
\usage{
computePCVariableCorrelation(
  se.obj,
  assay.names = "all",
  variable,
  fast.pca = TRUE,
  nb.pcs = 10,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Character. A character string or a vector of character strings specifying the name(s) of the assay(s)
in the SummarizedExperiment object for calculating the vector correlation. The default is set to "all", which indicates
that all assays of the SummarizedExperiment object will be selected.}

\item{variable}{Character. A character string indicating the name of the column in the sample annotation of the SummarizedExperiment
object. The variable must be categorical variable such as batches or sample subtypes.}

\item{fast.pca}{Logical. Indicates whether to use the fast PCA or PCA results computed by the 'computePCA' function.
The default is set to 'TRUE'. If 'FALSE', the 'fast.pca' argument in the 'computePCA' function should be set to 'FALSE'.}

\item{nb.pcs}{Numeric. A numeric value indicating the number of the first principal components (PCs) to use for
calculating the vector correlation. The default is set to 10. This number cannot exceed the number of PCs calculated
by the 'computePCA' function.}

\item{save.se.obj}{Logical. Indicates whether to save the results, correlation coefficients, in the metadata of the
SummarizedExperiment object or to output the result as a list. By default, it is set to 'TRUE'.}

\item{verbose}{Logical. If 'TRUE', displays the messages for the different steps of the function.}
}
\value{
A SummarizedExperiment object or a list containing the vector correlation plots of individual assay(s) for
a categorical variable.
}
\description{
This function calculates the the vector correlation between the first cumulative PCs of the gene expression (assay)
of a SummarizedExperiment object and a categorical variable (i.e. batch). Then, the functions generates a line-dot plot
between the first cumulative PCs and the correlation coefficient to see the relationship between different PCs with the
variable. An ideal normalization should results a low correlation with unwanted variation variables and high correlation
with known biology.
}
\details{
We use the Rozeboom squared vector correlation to quantify the strength of (linear) relationships between two sets
of variables, such as the first k PCs (that is 1 ≤ k ≤ 10) and dummy variables representing time, batches, plates and
biological variables. Not only does this quantity summarize the full set of canonical correlations, but it also reduces
to the familiar R2 from multiple regression when one of the variable sets contains just one element.
}
\references{
Molania R., ..., Speed, T. P., Removing unwanted variation from large-scale RNA sequencing data with PRPS,
Nature Biotechnology, 2023
}
\author{
Ramyar Molania
}
