% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePCVariableRegression.R
\name{computePCVariableRegression}
\alias{computePCVariableRegression}
\title{Computes the linear regression. --- Finalized}
\usage{
computePCVariableRegression(
  se.obj,
  assay.names = "all",
  variable,
  fast.pca = TRUE,
  nb.pcs = 10,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Character. A character string or list of character strings specifying the name(s) of the assay(s)
in the SummarizedExperiment object for performing regression analysis. The default is set to "all", which indicates that
all assays of the SummarizedExperiment object will be selected.}

\item{variable}{Character. A character string indicating the name of the column in the sample annotation of the
SummarizedExperiment object. The variable must be continuous, such as library size, tumor purity, etc.}

\item{fast.pca}{Logical. Indicates whether to use the computed fast PCA or PCA results from the 'computePCA' function.
The default is set to 'TRUE'. If 'FALSE', the 'fast.pca' argument in the 'computePCA' function should be set to 'FALSE'.}

\item{nb.pcs}{Numeric. A numeric value specifying the number of the first principal components (PCs) to use for computing
the vector correlation. The default is set to 10.}

\item{save.se.obj}{Logical. Indicates whether to save the results, regression R^2, to the 'metadata' of the
SummarizedExperiment object or output the result as a list. By default, it is set to 'TRUE'.}

\item{verbose}{Logical. If 'TRUE', displays the messages for the different steps of the function.}
}
\value{
A SummarizedExperiment object containing the computed regression R^2 for or output the result as a list.
}
\description{
This function calculates the linear regression between the the first 'nb.pcs' PCs (cumulatively) of the gene expression
data (assays) of a SummarizedExperiment object and a continuous variable (i.e. library size or tumor purity).
}
\details{
R2 values of fitted linear models are used to quantity the strength of the (linear) relationships between a single
quantitative source of unwanted variation, such as sample (log) library size or tumor purity, and global sample
summary statistics, such as the first k PCs (1 ≤ k ≤ 10).
}
\author{
Ramyar Molania
}
