% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSilhouette.R
\name{computeSilhouette}
\alias{computeSilhouette}
\title{Computes the average silhouette width using PCs. -- Finalized}
\usage{
computeSilhouette(
  se.obj,
  assay.names = "all",
  variable,
  dist.measure = "euclidian",
  fast.pca = TRUE,
  nb.pcs = 3,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Character. A character string or a vector of character strings specifying the name(s) of the assay(s)
in the SummarizedExperiment object for which to calculate the Silhouette coefficient. The default is set to 'all', which
indicates that all the assays of the SummarizedExperiment object will be selected.}

\item{variable}{Character. A character string indicating the column name that contains the categorical variable in the
SummarizedExperiment object. The variable can represent a biological or unwanted variable.}

\item{dist.measure}{Character. A character string indicating which distance measure to use for calculating the Silhouette
coefficient. The options are 'euclidean', 'maximum', 'manhattan', 'canberra', 'binary', or 'minkowski'. The default is
set to 'euclidean'. Refer to the 'dist' function from the 'stats' R package for more details.}

\item{fast.pca}{Logical. Indicates whether to use the PCs calculated using fast PCA. The default is set to 'TRUE'.
Fast PCA and ordinary PCA do not affect the Silhouette coefficient calculation.}

\item{nb.pcs}{Numeric. A numeric value indicating the number of principal components (PCs) to use when calculating the
distances between samples. The default is set to 3.}

\item{save.se.obj}{Logical. Indicates whether to save the silhouette coefficient  results in the metadata of the
SummarizedExperiment object or to output the result. The default is set to 'TRUE'.}

\item{verbose}{Logical. If 'TRUE', shows the messages of different steps of the function.}
}
\value{
Either the SummarizedExperiment object containing the computed Silhouette coefficient stored in the metadata
or a list of Silhouette coefficients for all specified assay(s).
}
\description{
This function calculates the average silhouette width for each categorical variables such as sample biological subtypes,
batches, etc.. The distance matrix is based on the principal components.
}
\details{
The Silhouette coefficient analysis is used to assess the separation of categorical samples groups. We use first few
principal components to calculate both the similarity between one sample and the other samples in each cluster and the
separation between samples in different clusters. A better normalization method will lead to higher and lower silhouette
coefficients for biological and batch variables, respectively.
}
\references{
Molania R., ..., Speed, T. P., Removing unwanted variation from large-scale RNA sequencing data with PRPS,
Nature Biotechnology, 2023
}
\author{
Ramyar Molania
}
