% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compuetSurvival.R
\name{computeSurvival}
\alias{computeSurvival}
\title{Computes variable and gene-level Kaplan-Meier survival analysis. Finalized}
\usage{
computeSurvival(
  se.obj,
  assay.names = "all",
  genes = "all",
  variable = NULL,
  survival.time,
  survival.events,
  expr.stratify = 2,
  return.p.value = TRUE,
  return.survival.plot = FALSE,
  plot.output = TRUE,
  assess.se.obj = TRUE,
  remove.na = "none",
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Character. A character or a vector of characters specifying the name(s) of the assay(s) in the
SummarizedExperiment object for gene-level survival analysis. The default is set to "all", which indicates that all
assays in the SummarizedExperiment object will be selected.}

\item{genes}{Character. A character or a vector of gene names/IDs specifying the genes to be used for gene-level survival
analysis. The default is set to 'all'.}

\item{variable}{Character. A character specifying the column name of the SummarizedExperiment object that contains
a categorical variable (e.g., tumor subtypes) for variable-level survival analysis. The default is set to 'NULL'.}

\item{survival.time}{Character. A character specifying the column name of the SummarizedExperiment object that contains
survival time.}

\item{survival.events}{Character. A character specifying the column name of the SummarizedExperiment object that contains
the survival event (0 and 1).}

\item{expr.stratify}{Numeric. A numeric value indicating how the gene expression of each gene should be stratified
for gene-level survival analysis. Stratification is based on the quantiles of gene expression. The default is 2,
meaning the expression level of each specified gene will be divided into two groups based on the 50\% quantile (median).}

\item{return.p.value}{Logical. Indicates whether to calculate p-values in the survival analysis. The default is 'TRUE'.}

\item{return.survival.plot}{Logical. Indicates whether to generate survival plots. The default is set to 'FALSE'.}

\item{plot.output}{Logical. Indicates whether to plot variable-level survival analysis. The default is 'TRUE'.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment object. If 'TRUE', the 'checkSeObj'
function will be applied. The default is 'TRUE'.}

\item{remove.na}{Character. Indicates whether to remove NA or missing values from the assays. Options are 'assays'
, 'sample annotation' or 'none'. The default is set to 'assays', which removes all NA or missing values from the assays
before performing the gene level analysis. See the 'checkSeObj' function for more details.}

\item{save.se.obj}{Logical. Indicates whether to save the results of the survival analysis in the metadata of the
SummarizedExperiment object or output the results as a list. The default is 'TRUE'.}

\item{verbose}{Logical. If 'TRUE', displays messages for each step of the function.}
}
\description{
This function computes gene-level and variable-level survival analysis.
}
\author{
Ramyar Molania
}
