% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlation}
\alias{correlation}
\title{is used to compute the correlation (Spearman or Pearson) between the data and a continuous variable (i.e. library size)}
\usage{
correlation(expr.data, apply.log, variable, method, n.cores)
}
\arguments{
\item{expr.data}{is the gene expression matrix genes by samples}

\item{apply.log}{Indicates whether to apply a log-transformation to the data}

\item{variable}{is a categorical variable such as library size or purity}

\item{method}{to select which method to use either Spearman or Pearson correlation}

\item{n.cores}{is the number of cpus used for mclapply parallelization}
}
\value{
dataframe containing the genes, the rho, the pvalues before and after BH correction
}
\description{
is used to compute the correlation (Spearman or Pearson) between the data and a continuous variable (i.e. library size)
}
