% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_gene_exp_contvar.R
\name{correlation_gene_exp_contvar}
\alias{correlation_gene_exp_contvar}
\title{is used to compute the correlation between the gene expression (assay)
of a SummarizedExperiment class object and a continous variable (i.e. library size).}
\usage{
correlation_gene_exp_contvar(
  se,
  cont_var,
  cont_var_label,
  assay_names = NULL,
  method = "spearman",
  apply.log = FALSE,
  n.cores = 5
)
}
\arguments{
\item{se}{A SummarizedExperiment object that will be used to compute the correlation}

\item{cont_var}{Vector of a categorical variable such as sample types
(i.e. biological subtypes) or batches.}

\item{cont_var_label}{String or vector of strings of the label of categorical variable(s) such as
sample types or batches from colData(se).}

\item{assay_names}{Optional string or list of strings for selection of the names
of the assays of the SummarizedExperiment class object to compute the correlation.}

\item{method}{A character string indicating which correlation coefficient
is to be used for the test: "pearson", "kendall", or "spearman". By default 'spearman will
be selected.}

\item{apply.log}{Indicates whether to apply a log-transformation to the data. By default
no transformation will be selected.}

\item{n.cores}{is the number of cpus used for mclapply parallelization.}
}
\value{
list List containing the association plot and the computed correlation
}
\description{
is used to compute the correlation between the gene expression (assay)
of a SummarizedExperiment class object and a continous variable (i.e. library size).
}
