% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prps.R
\name{create_prps}
\alias{create_prps}
\title{is used to create pseudo-replicates of pseudo samples of a SummarizedExperiment class object
using RUVIII-PRPS method.}
\usage{
create_prps(
  se,
  raw_data_assay_label,
  librarySize,
  biology,
  batch,
  purity,
  include.ls = FALSE,
  include.purity = FALSE,
  minSamplesPerBatchPS = 3,
  minSamplesForPurityPerBiology = 12,
  minSamplesForPurityPS = 3,
  minSamplesForLibrarySizePerBatch = 10,
  minSamplesForLibrarySizePS = 3
)
}
\arguments{
\item{se}{A SummarizedExperiment object that will be used to create PRPS.}

\item{raw_data_assay_label}{String for selection of the name of the raw data assay of the SummarizedExperiment class object.}

\item{librarySize}{Vector containing the library size of the samples.}

\item{biology}{Vector containing the biological subtypes of the samples.}

\item{batch}{Vector containing the batch variable of the samples.}

\item{purity}{Vector containing the purity variable of the samples.}

\item{include.ls}{Create PRPS for library size. By default is set to FALSE.}

\item{include.purity}{Create PRPS for purity. By default is set to FALSE.}

\item{minSamplesPerBatchPS}{Minimum number of samples per batch to create a pseudo-sample.}

\item{minSamplesForPurityPerBiology}{Minimum number of samples with the same biology to create a
pseudo-sample for purity.}

\item{minSamplesForPurityPS}{Minimum number of samples per batch to create a pseudo-sample to correct
for purity.}

\item{minSamplesForLibrarySizePerBatch}{Minimum number of samples per batch to create pseudo-samples
to correct for library size}

\item{minSamplesForLibrarySizePS}{Minimum number of samples with similar library size to create a
pseudo-sample to correct for library size.}
}
\value{
list List containing the PRPS created to correct for batch, for library size and for purity.
}
\description{
We will create distinct group of pseudo-replicates for each source of unwanted variation.
To correct for batch effect defined in the 'batch' argument, several group of pseudo-samples
will be created by averaging the samples of the same biological subtype in each batch. Then those
pseudo-samples will be defined as pseudo-replicates.
To correct for library size defined in the 'librarySize' argument, several group of pseudo-samples
will be created by averaging the top and bottom-ranked samples by library size of the same biological subtype in each batch.
Then those pseudo-samples will be defined as pseudo-replicates.
Similarly to correct for purity defined in the 'purity' argument, several group of pseudo-samples
will be created by averaging the top and bottom-ranked samples by purity of the same biological subtype in each batch.
Then those pseudo-samples will be defined as pseudo-replicates.
}
