% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createHomogeneousBioGroups.R
\name{createHomogeneousBioGroups}
\alias{createHomogeneousBioGroups}
\title{This is used to create all possible homogeneous biological groups in the SummarizedExperiment objct.}
\usage{
createHomogeneousBioGroups(
  se.obj,
  bio.variables,
  nb.clusters,
  clustering.method = "kmeans",
  assess.se.obj = TRUE,
  assess.variables = TRUE,
  cat.cor.coef = c(0.7, 0.7),
  cont.cor.coef = c(0.7, 0.7),
  remove.na = "sample.annotation",
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{bio.variables}{Symbol. Indicates the column names biological variables the SummarizedExperiment object.}

\item{nb.clusters}{A value to specify the number of groups of continuous sources of biological variation..}

\item{clustering.method}{A clustering method to group each continuous sources of biological variation.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment object.}

\item{assess.variables}{Logical. Indicates whether to assess association between the biological variables.}

\item{cat.cor.coef}{Numeric. Indicates correlation coefficients (Cramer's V, Pearson's contingency coefficient)
cut off for assessing association between categorical sources of biological variation.}

\item{cont.cor.coef}{Numeric. Indicates Spearman coefficients cut off for assessing association between continuous sources of biological variation.}

\item{remove.na}{Logical. To remove NA or missing values from sample annotation or not.}

\item{save.se.obj}{Logical. Indicates whether to save the results to SummarizedExperiment object or not.}

\item{verbose}{Logical. Whether to show the messages of the functions or not.}
}
\description{
This is used to create all possible homogeneous biological groups in the SummarizedExperiment objct.
}
