% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLogAssays.R
\name{createLogAssays}
\alias{createLogAssays}
\title{Apply log-transformation to the assay(s) of a SummarizedExperiment object.}
\usage{
createLogAssays(
  se.obj,
  assay.names = "all",
  pseudo.count = 1,
  replace.assays = FALSE,
  new.name = "Log.",
  apply.round = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Character. A single symbol or a vector of symbols specifying the name(s) of the assay(s) in the
SummarizedExperiment object. The default is "all," indicating that all assays in the SummarizedExperiment object will
be selected.}

\item{pseudo.count}{Numeric. A value to be added to all measurements in the assay(s) before log-transformation to prevent
-Inf values for measurements equal to 0. The default is 1.}

\item{replace.assays}{Logical. If 'TRUE', the original assay(s) will be replaced with the log-transformed values.}

\item{new.name}{Character. A symbol to append to the name(s) of the assay(s) for the log-transformed data. The default is 'log.'.}

\item{apply.round}{Logical. If 'TRUE', the measurements in the individual assays will be rounded to two decimal places.
The default is 'TRUE'.}

\item{verbose}{Logical. If 'TRUE', the function will display messages about the different steps being executed.}
}
\description{
Apply log-transformation to the assay(s) of a SummarizedExperiment object.
}
\author{
Marie Trussart
}
