% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPrPsByMnn.R
\name{createPrPsByMnn}
\alias{createPrPsByMnn}
\title{Create PRPS sets using k and mutual nearest neighbors in RNA-seq data.}
\usage{
createPrPsByMnn(
  se.obj,
  assay.name,
  main.uv.variable,
  other.uv.variables = NULL,
  filter.prps.sets = TRUE,
  max.prps.sets = 3,
  clustering.method = "kmeans",
  nb.clusters = 3,
  nb.other.uv.clusters = 2,
  other.uv.clustering.method = "kmeans",
  cont.cor.coef = c(0.9, 0.9),
  cat.cor.coef = c(0.9, 0.9),
  min.sample.for.ps = 3,
  min.mnn = 3,
  min.batch.number = "all",
  hvg = NULL,
  normalization = "CPM",
  apply.cosine.norm = FALSE,
  regress.out.variables = NULL,
  apply.log = TRUE,
  pseudo.count = 1,
  mnn.bpparam = SerialParam(),
  mnn.nbparam = KmknnParam(),
  assess.se.obj = TRUE,
  remove.na = "both",
  save.se.obj = TRUE,
  output.name = NULL,
  prps.group = NULL,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A summarized experiment object.}

\item{assay.name}{Symbol. A symbol indicating the name of the assay in the SummarizedExperiment object. This assay will
be used to first find k nearest neighbors and them mutual nearest neighbors data. This data must the one that will be
used for the RUV-III normalization.}

\item{filter.prps.sets}{Logical. If 'TRUE', the number of PRPS sets across each pair of batches will be filtered if they are
higher than the 'max.prps.sets' value. The default is 'TRUE'. The high number of PRPS sets will just increase the
computational time for the RUV-III normalization.}

\item{max.prps.sets}{Numeric. A numeric value specifying the maximum number for PRPS set across each pair of batches.
The default is set to 10.}

\item{clustering.method}{Symbol.A symbol indicating the choice of clustering method for grouping the 'uv.variable'
if a continuous variable is provided. Options include 'kmeans', 'cut', and 'quantile'. The default is set to 'kmeans'.}

\item{nb.clusters}{Numeric. A numeric value indicating how many clusters should be found if the 'uv.variable' is a
continuous variable. The default is 3.}

\item{hvg}{Vector. A vector of the names of the highly variable genes. These genes will be used to find the anchors
samples across the batches. The default is NULL.}

\item{normalization}{Symbol. Indicates which normalization methods should be applied before finding the knn. The default
is 'cpm'. If is set to NULL, no normalization will be applied.}

\item{apply.cosine.norm}{TTT}

\item{regress.out.variables}{Symbols. Indicates the columns names that contain biological variables in the
SummarizedExperiment object. These variables will be regressed out from the data before finding genes that are highly
affected by unwanted variation variable. The default is NULL, indicates the regression will not be applied.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data or not. The default is TRUE.
Please, note, any RNA-seq data (assays) must be in log scale before computing RLE.}

\item{pseudo.count}{Numeric. A value as a pseudo count to be added to all measurements of the assay(s) before applying
log transformation to avoid -Inf for measurements that are equal to 0. The default is 1.}

\item{mnn.bpparam}{Symbol. A BiocParallelParam object specifying how parallelization should be performed. The default
is SerialParam(). We refer to the 'findMutualNN' function from the BiocNeighbors R package for more details.}

\item{mnn.nbparam}{Symbol. A BiocParallelParam object specifying how parallelization should be performed to find MNN.
. The default is KmknnParam(). We refer to the 'findMutualNN' function from the 'BiocNeighbors' R package.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment object or not. See the checkSeObj
function for more details.}

\item{remove.na}{Symbol. To remove NA or missing values from the assays or not. The options are 'assays' and 'none'.
The default is "assays", so all the NA or missing values from the assay(s) will be removed before computing RLE. See
the checkSeObj function for more details.}

\item{save.se.obj}{Logical. Indicates whether to save the RLE results in the metadata of the SummarizedExperiment object
or to output the result as list. By default it is set to TRUE.}

\item{output.name}{Symbol. A symbol specifying the name of output file. If is 'NULL', the function will select a name
based on "paste0(uv.variable, '|', 'anchor', '|', assay.name))".}

\item{prps.group}{Symbol. A symbol specifying the name of the output file. If is 'NULL', the function will select a name
based on "paste0('prps_mnn_', uv.variable)".}

\item{verbose}{Logical. If 'TRUE', shows the messages of different steps of the function.}

\item{uv.variable}{Symbol. Indicates the name of a column in the sample annotation of the SummarizedExperiment object.
The 'uv.variable' can be either categorical and continuous. If 'uv.variable' is a continuous variable, this will be
divided into 'nb.clusters' groups using the 'clustering.method'.}

\item{mnn}{Numeric. The maximum number of nearest neighbors to compute. The default is set 3.}
}
\value{
The SummarizedExperiment object that contain all the PPRS data, knn, mnn and plot results in the metadata, or
a list of the results.
}
\description{
This function uses the k and mutual nearest neighbors approaches to create PRPS in the RNA-seq data. This function
can be used in situation that the biological variation are entirely unknown. The function applies the 'findKnn' function
to find similar samples per batch and then average them to create pseudo-samples. Then, function uses the 'findMnn' to
match up pseudo samples across batches to create pseudo-replicates.
}
\author{
Ramyar Molania
}
