% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPrPsSupervised.R
\name{createPrPsSupervised}
\alias{createPrPsSupervised}
\title{Create supervised pseudo-replicates of pseudo samples (PRPS).}
\usage{
createPrPsSupervised(
  se.obj,
  assay.name,
  bio.variables,
  uv.variables,
  apply.other.uv.variables = TRUE,
  min.sample.for.prps = 3,
  bio.clustering.method = "kmeans",
  nb.bio.clusters = 2,
  other.uv.clustering.method = "kmeans",
  nb.other.uv.clusters = 2,
  check.prps.connectedness = TRUE,
  apply.log = TRUE,
  pseudo.count = 1,
  assess.se.obj = TRUE,
  assess.variables = FALSE,
  cat.cor.coef = c(0.7, 0.7),
  cont.cor.coef = c(0.7, 0.7),
  remove.na = "both",
  save.se.obj = TRUE,
  plot.output = TRUE,
  prps.name = NULL,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.name}{Symbol. A symbol used to specify the assay name within the SummarizedExperiment object. The selected
assay should be the one that will be used for RUV-III-PRPS normalization.}

\item{bio.variables}{Symbol. A symbol or symbols representing the label of biological variable(s), such as cancer
subtypes, tumour purity, ... within the SummarizedExperiment object. This can comprise a vector containing either
categorical, continuous, or a combination of both variables.}

\item{uv.variables}{Symbol. A symbol or symbols representing the label of unwanted variable(s), such as batch effects,
library size, ... within the SummarizedExperiment object. This can comprise a vector containing either categorical,
continuous, or a combination of both variables.}

\item{apply.other.uv.variables}{Logical. Determines whether to include other specified unwanted variables when generating
PRPS sets for individual ones. The default is set to 'TRUE'.}

\item{min.sample.for.prps}{Numeric. Indicates the minimum number of biologically homogeneous samples to be averaged
to create one pseudo-sample. The default it is set to 3.}

\item{bio.clustering.method}{Symbol. A symbol indicating the clustering method used to group each continuous biological
variable. The option include 'kmeans', cut' and 'quantile'. The default is to 'kmeans'. We refer to the
createHomogeneousBioGroups() function for more details.}

\item{nb.bio.clusters}{Numeric. A numeric value to specify the number of clusters/groups for each continuous biological
variable. The default it set to 3.}

\item{other.uv.clustering.method}{Symbol. A symbol indicating the clustering method used to group each continuous unwanted
variable. The option include 'kmeans', cut' and 'quantile'. The default is to 'kmeans'. We refer to the
createHomogeneousUVGroups() function for more details.}

\item{nb.other.uv.clusters}{Numeric. A numeric value to specify the number of clusters/groups for each continuous biological
variable. The default it set to 3.}

\item{check.prps.connectedness}{Logical. Indicates whether to assess the connectedness between the PRPS sets or not.
The default is set to 'TRUE'. See the details for more information.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data. The default is set to 'TRUE'.}

\item{pseudo.count}{Numeric. A numeric value as a pseudo count to be added to all measurements before log transformation.
The default is set to 1.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment object. If 'TRUE', the checkSeObj
will be applied inside the function. The default is set to 'TRUE'.}

\item{assess.variables}{Logical. Indicates whether to assess the association between the biological or unwanted
variable(s) separately. The default is set to 'FALSE'. We refer to the 'assessVariableAssociation' for more details.}

\item{cat.cor.coef}{Vector of two numerical values. Indicates the cut-off of the correlation coefficient between each
pair of categorical variables. The first one is between each pair of 'uv.variables' and the second one is between each
pair of 'bio.variables'. The correlation is computed by the function ContCoef from the DescTools package. If the correlation
of a pair of variable is higher than the cut-off, then only the variable that has the highest number of factor will be
kept and the other one will be excluded from the remaining analysis. By default they are both set to 0.7.}

\item{cont.cor.coef}{Vector of two numerical values. Indicates the cut-off of the Spearman correlation coefficient
between each pair of continuous variables. The first one is between each pair of 'uv.variables' and the second one is
between each pair of 'bio.variables'. If the correlation of a pair of variable is higher than the cut-off, then only
the variable that has the highest variance will be kept and the other one will be excluded from the remaining analysis.
By default they are both set to 0.7.}

\item{remove.na}{Symbol. A symbol Indicating whether to remove NA or missing values from either the 'assays', the
'sample.annotation', both' or 'none'. If 'assays' is selected, the genes that contains NA or missing values will be
excluded. If sample.annotation' is selected, the samples that contains NA or missing values for any 'bio.variables' and
'uv.variables' will be excluded. The default is set to both'.}

\item{save.se.obj}{Logical. Indicates whether to save the results in the metadata of the SummarizedExperiment object
or to output the result as list. The default by is set to 'TRUE'.}

\item{plot.output}{Logical. Indicates whether to generate the PRPS map plot for individual sources of unwanted variation.
The default is 'TRUE'.}

\item{prps.name}{Symbol. A symbol to specify the name of all PRPS sets that will be created for all specified source(s)
of unwanted variation. The default is set to 'NULL'. The, the function creates a name based on paste0('prps_', uv.variable).}

\item{verbose}{Logical. If 'TRUE', shows the messages of different steps of the function.}
}
\value{
A SummarizedExperiment object that contains all the PRPS data and PPRS map plot in the metadata or a list
that contains all the results.
}
\description{
This function creates different PRPS sets when all sources of unwanted and biological variation are known.
}
\details{
We will create distinct group of pseudo-replicates for each source of unwanted variation defined in the 'uv.variables'
argument. For example to correct for batch effect if defined in the 'uv.variables' argument, several group of
pseudo-samples will be created by averaging the samples of the same biological subtype defined in 'bio.variables' in
each batch. Then those pseudo-samples will be defined as pseudo-replicates which constitutes a PRPS set. For example
to correct for library size if defined in the 'uv.variables' argument, several group of pseudo-samples will be created
by averaging the top and bottom-ranked samples by library size of the same biological subtype in each batch. Then those
pseudo-samples will be defined as pseudo-replicates which constitutes a PRPS set. Similarly to correct for purity if
defined in the uv.variables' argument, several group of pseudo-samples will be created by averaging the top and
bottom-ranked samples by purity of the same biological subtype in each batch. Then those pseudo-samples will be defined
as pseudo-replicates which constitutes a PRPS set.
}
\references{
Molania R., ..., Speed, T. P., A new normalization for Nanostring nCounter gene expression data, Nucleic Acids Research,
2019.
Molania R., ..., Speed, T. P., Removing unwanted variation from large-scale RNA sequencing data with PRPS,
Nature Biotechnology, 2023
}
\author{
Ramyar Molania
}
