% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSeObj.R
\name{createSeObj}
\alias{createSeObj}
\title{is used to create a summarizedExperiment object.}
\usage{
createSeObj(
  assays,
  raw.count.assay.name = NULL,
  remove.lowly.expressed.genes = FALSE,
  count.cutoff = 10,
  biological.group = NULL,
  minimum.proportion = 0.5,
  calculate.library.size = FALSE,
  sample.annotation = NULL,
  create.sample.annotation = FALSE,
  gene.annotation = NULL,
  create.gene.annotation = FALSE,
  add.gene.details = FALSE,
  gene.group = NULL,
  gene.details = NULL,
  add.housekeeping.genes = FALSE,
  add.immunStroma.genes = FALSE,
  metaData = NULL,
  verbose = TRUE
)
}
\arguments{
\item{assays}{A list of assays.}

\item{raw.count.assay.name}{The name of raw counts data in the assays.}

\item{remove.lowly.expressed.genes}{logical, if TRUE then remove lowly expressed genes.}

\item{count.cutoff}{numeric. Minimum count required for at least some samples.}

\item{biological.group}{A column name in sample annotation that specify biological groups. The smallest population is considred for relvling lowly expressed genes.}

\item{minimum.proportion}{numeric. In large sample situations, the minimum proportion of samples in a group that a gene needs to be expressed in. See Details below for the exact formula.}

\item{calculate.library.size}{logical, if TRUE then library size is calculated on the raw.count.assay.name.}

\item{sample.annotation}{a data frame, contains information for samples.}

\item{create.sample.annotation}{logical, if TRUE then a sample annotation the initially contains column names of the assays.}

\item{gene.annotation}{a data frame, contains information for genes.}

\item{create.gene.annotation}{logical, if TRUE then a gene annotation that initially contains row names of the assays.}

\item{add.gene.details}{logical, if TRUE then a pr-set or provided gene details will be added to the gene annotation.}

\item{gene.group}{A name of gene class in the assays, this should be in 'entrezgene_id', 'hgnc_symbol', 'ensembl_gene_id'.}

\item{gene.details}{A list of gene details to be added to the gene annotation.}

\item{add.housekeeping.genes}{logical. if TRUE then several sets of publicly available sets of "housekeeping" will be added to the gene annotation.}

\item{add.immunStroma.genes}{Logical. If TRUE, the immune and stromal genes signature from Kosuke Yoshihara et.al will be added.}

\item{metaData}{Any metadata data for in the data.}

\item{verbose}{logical, if TRUE shows the messages.}
}
\value{
a summarizedExperiment that contains assays, gene annotation, samples annotation and metadata.
}
\description{
is used to create a summarizedExperiment object.
}
