% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSupervisedPRPS.R
\name{createSupervisedPRPS}
\alias{createSupervisedPRPS}
\title{Create supervised pseudo-replicates of pseudo samples (PRPS).}
\usage{
createSupervisedPRPS(
  se.obj,
  assay.name,
  bio.variables,
  uv.variables,
  apply.other.uv.variables = TRUE,
  min.sample.for.prps = 3,
  bio.clustering.method = "kmeans",
  nb.bio.clusters = 2,
  other.uv.clustering.method = "kmeans",
  nb.other.uv.clusters = 2,
  check.prps.connectedness = TRUE,
  apply.log = TRUE,
  pseudo.count = 1,
  assess.se.obj = TRUE,
  assess.variables = FALSE,
  cat.cor.coef = c(0.7, 0.7),
  cont.cor.coef = c(0.7, 0.7),
  remove.na = "both",
  save.se.obj = TRUE,
  plot.output = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.name}{Symbol.  for the selection of the name of the assay of the SummarizedExperiment class object.}

\item{bio.variables}{String of the label of (a) categorical or continuous variable(s) used to define homogeneous
biological groups of samples such as samples types from colData(se).}

\item{uv.variables}{String of the label of a categorical variable
such as samples types or batch from colData(se) that will be used to define PRPS.}

\item{apply.other.uv.variables}{String of the label of (a) categorical or continuous variable(s) used to define
homogeneous groups of samples such as library size, plates from colData(se). By default it is set to 'NULL' meaning the
samples will be assigned to homogeneous biological groups of samples using the 'bio.variables'.}

\item{min.sample.for.prps}{Numeric. Indicates the minimum number of samples to create one pseudo-sample,
by default it is set to 3.
to create a PRPS set for each continuous variable. The minimum should be '2*min.sample.for.prps'. By default it is set
to 6.}

\item{bio.clustering.method}{String of the clustering method to assign each sample to an homogeneous biological
clusters/group using the function 'kmeans', cut' from base or the function 'quantile'. By default it is to 'kmeans'.}

\item{nb.bio.clusters}{Numeric. A value to specify the number of homogeneous biological clusters/groups of samples.
By default it is set to 3.}

\item{other.uv.clustering.method}{String of the clustering method to assign each sample to an homogeneous clusters/group
of samples based on the 'other.uv.variables' using the function 'kmeans','cut' from base or the function 'quantile'.
By default it is to 'kmeans'.}

\item{nb.other.uv.clusters}{Numeric. A value to specify the number of groups for continuous sources of biological
variation. The default is 2. This means each continuous sources will be divided into 2 groups using the 'clustering.method'.}

\item{check.prps.connectedness}{Logical. Indicates to assess the connectedness between the PRPS sets or not. The default
is set to 'TRUE'. We refer to the reference number X for more detail.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data, by default it is set to TRUE.}

\item{pseudo.count}{Numeric. A value as a pseudo count to be added to all measurements before log transformation,
by default it is set to 1.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment class object.
By default it is set to TRUE.}

\item{assess.variables}{Logical. Indicates whether to assess the association between the variable. The default is set
to 'FALSE'. We refer to the 'assessVariableAssociation' for more details.}

\item{cat.cor.coef}{Vector of two numerical values. Indicates the cut-off of the correlation coefficient between each
pair of categorical variables. The first one is between each pair of 'uv.variables' and the second one is between each
pair of 'bio.variables'. The correlation is computed by the function ContCoef from the DescTools package. If the correlation
of a pair of variable is higher than the cut-off, then only the variable that has the highest number of factor will be
kept and the other one will be excluded from the remaining analysis. By default they are both set to 0.7.}

\item{cont.cor.coef}{Vector of two numerical values. Indicates the cut-off of the Spearman correlation coefficient between
each pair of continuous variables. The first one is between each pair of 'uv.variables' and the second one is between
each pair of 'bio.variables'. If the correlation of a pair of variable is higher than the cut-off, then only the variable
that has the highest variance will be kept and the other one will be excluded from the remaining analysis. By default
they are both set to 0.7.}

\item{remove.na}{String. Indicates whether to remove NA or missing values from either the 'assays', the 'sample.annotation',
'both' or 'none'. If 'assays' is selected, the genes that contains NA or missing values will be excluded. If
'sample.annotation' is selected, the samples that contains NA or missing values for any 'bio.variables' and
'uv.variables' will be excluded. By default, it is set to both'.}

\item{save.se.obj}{Logical. Indicates whether to save the result in the metadata of the SummarizedExperiment class
object 'se.obj' or to output the result, by default it is set to TRUE.}

\item{plot.output}{Logical. Indicates the PRPS map for individual sources of unwanted variation. The default is 'TRUE'.}

\item{verbose}{Logical. Indicates whether to show or reduce the level of output or
messages displayed during the execution of the functions, by default it is set to TRUE.}
}
\value{
A SummarizedExperiment object or a list that contains all the PRPS sets.
}
\description{
This function creates different PRPS sets when all sources of unwanted and biological variation are known.
}
\details{
We will create
distinct group of pseudo-replicates for each source of unwanted variation defined in the 'uv.variables' argument. For
example to correct for batch effect if defined in the 'uv.variables' argument, several group of pseudo-samples will be
created by averaging the samples of the same biological subtype defined in 'bio.variables' in each batch. Then those
pseudo-samples will be defined as pseudo-replicates which constitutes a PRPS set. For example to correct for library
size if defined in the 'uv.variables' argument, several group of pseudo-samples will be created by averaging the top
and bottom-ranked samples by library size of the same biological subtype in each batch. Then those pseudo-samples will
be defined as pseudo-replicates which constitutes a PRPS set. Similarly to correct for purity if defined in the
'uv.variables' argument, several group of pseudo-samples will be created by averaging the top and bottom-ranked samples
by purity of the same biological subtype in each batch. Then those pseudo-samples will be defined as pseudo-replicates
which constitutes a PRPS set.
}
\references{
Molania R., ..., Speed, T. P., A new normalization for Nanostring nCounter gene expression data, Nucleic Acids Research,
2019.
Molania R., ..., Speed, T. P., Removing unwanted variation from large-scale RNA sequencing data with PRPS,
Nature Biotechnology, 2023
}
\author{
Ramyar Molania
}
