% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createUnSupervisedPRPSbyAnchors.R
\name{createUnSupervisedPRPSbyAnchors}
\alias{createUnSupervisedPRPSbyAnchors}
\title{create PRPS sets across batches using integration anchors.}
\usage{
createUnSupervisedPRPSbyAnchors(
  se.obj,
  assay.name,
  uv.variable,
  clustering.method = "kmeans",
  nb.clusters = 3,
  hvg = NULL,
  apply.log = TRUE,
  pseudo.count = 1,
  anchor.features = 2000,
  scale = TRUE,
  min.prps.samples = 3,
  max.prps.samples = "inf",
  max.prps.sets = 10,
  min.score = NULL,
  normalization.method = "LogNormalize",
  sct.clip.range = NULL,
  reduction = "cca",
  l2.norm = TRUE,
  dims = 1:10,
  k.anchor = 3,
  k.filter = 200,
  k.score = 30,
  max.features = 200,
  nn.method = "annoy",
  n.trees = 50,
  eps = 0,
  assess.se.obj = TRUE,
  remove.na = "assays",
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.name}{Symbol. A symbol indicating the assay name within the SummarizedExperiment object for the creation
of PRPS data. The assay must be the one that will be used as data input for the RUV-III-PRPS normalization.}

\item{uv.variable}{Symbol. A symbol specifying the name of columns in the sample annotation of the SummarizedExperiment
object. This variable can to be categorical or continuous variable. If a continuous variable is provide, this will be
divided into groups using the clusteintf method.}

\item{clustering.method}{Symbol.A symbol indicating the choice of clustering method for grouping the 'uv.variable'
if a continuous variable is provided. Options include 'kmeans', 'cut', and 'quantile'. The default is set to 'kmeans'.}

\item{nb.clusters}{Numeric. A numeric value indicating how many clusters should be found if the 'uv.variable' is a
continuous variable. The default is 3.}

\item{hvg}{Vector. A vector containing the names of highly variable genes. These genes will be utilized to identify
anchor samples across different batches. The default value is set to 'NULL'.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data or not. The default is 'TRUE'.}

\item{pseudo.count}{Numeric. A value as a pseudo count to be added to all measurements of the assay before applying
log transformation to avoid -Inf for raw counts that are equal to 0. The default is 1.}

\item{anchor.features}{Numeric. A numeric value indicating the provided number of features to be used in anchor finding.
The default is 2000. We refer to the FindIntegrationAnchors R function for more details.}

\item{scale}{Logical. Whether or not to scale the features provided. Only set to FALSE if you have previously scaled
the features you want to use for each object in the object.list.}

\item{min.prps.samples}{Numeric. The minimum number of samples to be averaged to create a pseudo-sample. The default
is 3. The minimum value is 2.}

\item{max.prps.samples}{Numeric value indicating the maximum number of samples to be averaged for creating a pseudo-sample.
The default is 'inf'. Please note that averaging a high number of samples may lead to inaccurate PRPS estimates.}

\item{max.prps.sets}{Numeric. The maximum number of PRPS sets across batches. The default in 10.}

\item{min.score}{Numeric. A cut off to filter anchors based the scores that the FindIntegrationAnchors calculates.
The default is NULL, indicating no filtration will be applied based on the scores. The value is between 0 or 1.}

\item{normalization.method}{Symbol. Indicate which normalization methods should be used before finding the anchors.
The options are "LogNormalize" or "SCT". The default is "LogNormalize".}

\item{sct.clip.range}{Numeric. Numeric of length two specifying the min and max values the Pearson residual will be
clipped to. The default is 'NULL'. We refer to the FindIntegrationAnchors R function for more details.}

\item{reduction}{Symbol. Indicates which dimensional reduction to perform when finding anchors. The options are "cca":
canonical correlation analysis, "rpca": reciprocal PCA and "rlsi": Reciprocal LSI. The default is "cca".}

\item{l2.norm}{Logical. Indicates whether to perform L2 normalization on the CCA samples embeddings after dimensional
reduction or not. The default is 'TRUE'.}

\item{dims}{Numeric. Indicates which dimensions to use from the CCA to specify the neighbor search space. Th default
is 10.}

\item{k.anchor}{Numeric. How many neighbors (k) to use when picking anchors. Th default is 3.}

\item{k.filter}{Numeric. How many neighbors (k) to use when filtering anchors. Th default is 200.}

\item{k.score}{Numeric. How many neighbors (k) to use when scoring anchors. Th default is 30.}

\item{max.features}{Numeric. The maximum number of features to use when specifying the neighborhood search space in
the anchor filtering.Th default is 30.}

\item{nn.method}{Symbol.Method for nearest neighbor finding. Options include: "rann", "annoy". The defauly is "annoy".}

\item{n.trees}{Numeric. More trees gives higher precision when using annoy approximate nearest neighbor search}

\item{eps}{Numeric. Error bound on the neighbor finding algorithm (from RANN/Annoy)}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment object or not. See the checkSeObj
function for more details.}

\item{remove.na}{Symbol. To remove NA or missing values from the assays or not. The options are 'assays' and 'none'.
The default is "assays", so all the NA or missing values from the assay(s) will be removed before computing RLE. See
the checkSeObj function for more details.}

\item{save.se.obj}{Logical. Indicates whether to save the RLE results in the metadata of the SummarizedExperiment
object or to output the result as list. By default it is set to TRUE.}

\item{verbose}{Logical. If 'TRUE', shows the messages of different steps of the function.}
}
\description{
This functions employs the FindIntegrationAnchors function from the Seurat R package to create PRPS sets for the
RUV-III normalization of RNA-seq data. This function can be used in situations in which biological variation are unknown.
}
\author{
Ramyar Molania
}
