% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createUnSupervisedPRPSbyMNN.R
\name{createUnSupervisedPRPSbyMNN}
\alias{createUnSupervisedPRPSbyMNN}
\title{create PRPS sets using k and mutual nearest neighbors in RNA-seq data.}
\usage{
createUnSupervisedPRPSbyMNN(
  se.obj,
  assay.name,
  uv.variable,
  data.input = "expr",
  nb.pcs = 2,
  center.pca = TRUE,
  scale.pca = FALSE,
  svd.bsparam = bsparam(),
  clustering.method = "kmeans",
  nb.clusters = 3,
  k = 2,
  hvg = NULL,
  normalization = "CPM",
  regress.out.bio.variables = NULL,
  apply.log = TRUE,
  pseudo.count = 1,
  mnn.bpparam = SerialParam(),
  assess.se.obj = TRUE,
  remove.na = "both",
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A summarized experiment object.}

\item{assay.name}{Symbol. A symbol for the selection of the name of the assay in the SummarizedExperiment object to be
used to find k nearest neighbors data.}

\item{uv.variable}{Symbol. Indicates the name of a column in the sample annotation of the SummarizedExperiment object.
The 'uv.variable' can be either categorical and continuous. If 'uv.variable' is a continuous variable, this will be
divided into 'nb.clusters' groups using the 'clustering.method'.}

\item{data.input}{Symbol. Indicates which data should be used an input for finding the k nearest neighbors data. Options
include: 'expr' and 'pcs'. If 'pcs' is selected, the first PCs of the data will be used as input. If 'expr' is selected,
the data will be selected as input.}

\item{nb.pcs}{Numeric. Indicates the number PCs should be used as data input for finding the k nearest neighbors. The
'nb.pcs' must be set when the "data.input = PCs". The default is 2.}

\item{center.pca}{Logical. Indicates whether to scale the data or not. If center is TRUE, then centering is done by
subtracting the column means of the assay from their corresponding columns. The default is TRUE.}

\item{scale.pca}{Logical. Indicates whether to scale the data or not before applying SVD.  If scale is TRUE, then scaling
is done by dividing the (centered) columns of the assays by their standard deviations if center is TRUE, and the root
mean square otherwise. The default is FALSE}

\item{svd.bsparam}{A BiocParallelParam object specifying how parallelization should be performed. The default is bsparam().
We refer to the 'runSVD' function from the BiocSingular R package.}

\item{clustering.method}{Symbol.A symbol indicating the choice of clustering method for grouping the 'uv.variable'
if a continuous variable is provided. Options include 'kmeans', 'cut', and 'quantile'. The default is set to 'kmeans'.}

\item{nb.clusters}{Numeric. A numeric value indicating how many clusters should be found if the 'uv.variable' is a
continuous variable. The default is 3.}

\item{k}{Numeric.The maximum number of nearest neighbors to compute. The default is set 3.}

\item{hvg}{Vector. A vector of the names of the highly variable genes. These genes will be used to find the anchors
samples across the batches. The default is NULL.}

\item{normalization}{Symbol. Indicates which normalization methods should be applied before finding the knn. The default
is 'cpm'. If is set to NULL, no normalization will be applied.}

\item{regress.out.bio.variables}{Symbols. Indicates the columns names that contain biological variables in the
SummarizedExperiment object. These variables will be regressed out from the data before finding genes that are highly
affected by unwanted variation variable. The default is NULL, indicates the regression will not be applied.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data or not. The default is TRUE.
Please, note, any RNA-seq data (assays) must be in log scale before computing RLE.}

\item{pseudo.count}{Numeric. A value as a pseudo count to be added to all measurements of the assay(s) before applying
log transformation to avoid -Inf for measurements that are equal to 0. The default is 1.}

\item{mnn.bpparam}{TTTT}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment object or not. See the checkSeObj
function for more details.}

\item{remove.na}{Symbol. To remove NA or missing values from the assays or not. The options are 'assays' and 'none'.
The default is "assays", so all the NA or missing values from the assay(s) will be removed before computing RLE. See
the checkSeObj function for more details.}

\item{save.se.obj}{Logical. Indicates whether to save the RLE results in the metadata of the SummarizedExperiment object
or to output the result as list. By default it is set to TRUE.}

\item{verbose}{Logical. If 'TRUE', shows the messages of different steps of the function.}
}
\description{
This function uses the k and mutual nearest neighbors approaches to create PRPS in the RNA-seq data. This function
can be used in situation that the biological variation are not known. The function applies the 'findKnn' function
to find similar samples per batch and then average them to create pseudo-samples. The, function uses the 'findMnn' to
match up pseudo samples across batches to create pseudo-replicates.
}
\author{
Ramyar Molania
}
