% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_analysis.R
\name{de_analysis}
\alias{de_analysis}
\title{is used to compute the differential analysis given a variable of the data}
\usage{
de_analysis(sce, normalization, regression_var, apply.log = FALSE, n.cores = 5)
}
\arguments{
\item{sce}{the dataset that will be used for this analysis}

\item{normalization}{All the available assays for the data (i.e. normalizations methods)}

\item{regression_var}{The regression variable that will be computed to the PCA of the data (library size)}

\item{apply.log}{Indicates whether to apply a log-transformation to the data}

\item{n.cores}{is the number of cpus used for mclapply parallelization}
}
\value{
list List containing the association plot and the computed regression
}
\description{
is used to compute the differential analysis given a variable of the data
}
