% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/de_analysis_wilcoxon_gene_exp_catvar_all_assays.R
\name{de_analysis_wilcoxon_gene_exp_catvar_all_assays}
\alias{de_analysis_wilcoxon_gene_exp_catvar_all_assays}
\title{is used to compute the differential gene expression analysis on all the assays
given a categorical variable of two groups of the samples using the Wilcoxon test}
\usage{
de_analysis_wilcoxon_gene_exp_catvar_all_assays(
  sce,
  regression_var,
  apply.log = FALSE,
  n.cores = 5
)
}
\arguments{
\item{sce}{the dataset that will be used for this analysis}

\item{regression_var}{The categorical variable of two groups that will be used
to computed the differential expression analysis (i.e. high vs low library size)}

\item{apply.log}{Indicates whether to apply a log-transformation to the data}

\item{n.cores}{is the number of cpus used for mclapply parallelization}
}
\value{
list List containing the association plot and the computed regression
}
\description{
is used to compute the differential gene expression analysis on all the assays
given a categorical variable of two groups of the samples using the Wilcoxon test
}
