% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/de_analysis_wilcoxon_gene_exp_catvar_single_assay.R
\name{de_analysis_wilcoxon_gene_exp_catvar_single_assay}
\alias{de_analysis_wilcoxon_gene_exp_catvar_single_assay}
\title{is used to compute the Wilcoxon Rank Sum and Signed Rank Test on the gene expression
of a single assay given a categorical variable of two groups}
\usage{
de_analysis_wilcoxon_gene_exp_catvar_single_assay(
  expr.data,
  apply.log = FALSE,
  variable,
  n.cores
)
}
\arguments{
\item{expr.data}{is the gene expression matrix genes by samples}

\item{apply.log}{Indicates whether to apply a log-transformation to the data}

\item{variable}{is a categorical variable such as sample types or batches}

\item{n.cores}{is the number of cpus used for mclapply parallelization}
}
\value{
dataframe containing the genes, the p-values before and after BH correction
}
\description{
is used to compute the Wilcoxon Rank Sum and Signed Rank Test on the gene expression
of a single assay given a categorical variable of two groups
}
