% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pca.R
\name{.pca}
\alias{.pca}
\title{is used to compute PCA of the data}
\usage{
.pca(sce, apply.log = FALSE)
}
\arguments{
\item{sce}{Dataset that will be used to compute the PCA}

\item{apply.log}{Indicates whether to apply a log-transformation to the data}
}
\value{
pca.all PCA components computed on the provided data
}
\description{
is used to compute PCA of the data
}
