% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBioGenes.R
\name{findBioGenes}
\alias{findBioGenes}
\title{Find highly variable genes.}
\usage{
findBioGenes(
  se.obj,
  assay.name,
  approach = "TwoWayAnova",
  nb.bio.genes = 0.1,
  bio.variables,
  uv.variables,
  bio.clustering.method = "kmeans",
  nb.bio.clusters = 3,
  uv.groups = NULL,
  uv.clustering.method = "kmeans",
  nb.uv.clusters = 3,
  apply.log = TRUE,
  pseudo.count = 1,
  normalization = "CPM",
  grid.nb = 20,
  regress.out.uv.variables = NULL,
  assess.bio.genes = TRUE,
  variables.to.assess.bio.genes = NULL,
  nb.pcs = 5,
  center = TRUE,
  scale = FALSE,
  anova.method = "aov",
  min.sample.for.aov = 3,
  corr.method = "spearman",
  a = 0.05,
  rho = 0,
  min.sample.for.correlation = 10,
  min.sample.for.mad = 3,
  assess.se.obj = TRUE,
  assess.variables = FALSE,
  cat.cor.coef = c(0.9, 0.9),
  cont.cor.coef = c(0.9, 0.9),
  remove.na = "none",
  save.se.obj = TRUE,
  output.name = NULL,
  save.imf = FALSE,
  imf.name = NULL,
  use.imf = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.name}{Character. A symbol indicating the name of the assay in the SummarizedExperiment object.}

\item{approach}{Character. A symbol specifying the biological gene selection method to be used. Options include:
'AnovaCorr.PerBatchPerBio', 'AnovaCorr.AcrossAllSamples', 'TwoWayAnova', and 'mad.unsupervised'. The default is 'TwoWayAnova'.
Refer to the function details for more information.}

\item{nb.bio.genes}{Numeric. A numeric value representing the proportion of the total genes to be selected as biological genes.
The default is 0.1.}

\item{bio.variables}{Character. A symbol or a vector of symbols indicating the column names that contain known biological
variable(s) in the SummarizedExperiment object. These biological variables can be categorical or continuous. Continuous
variables will be divided into 'nb.bio.clusters' groups based on the clustering method specified in the 'bio.clustering.method'
argument.}

\item{uv.variables}{Character. A symbol or a vector of symbols indicating the column names that contain unwanted variable(s)
in the SummarizedExperiment object. These variables can be categorical or continuous. Continuous variables will be divided into
'nb.uv.clusters' groups based on the clustering method specified in the 'uv.clustering.method' argument.}

\item{bio.clustering.method}{Character. A symbol specifying the clustering method to be used for grouping continuous sources
of biological variation. The default is 'kmeans'. Refer to the 'createHomogeneousBioGroups' function for more details.}

\item{nb.bio.clusters}{Numeric. A numeric value indicating the number of clusters for each continuous source of biological
variation. The default is 3, meaning each continuous biological variable will be divided into 3 groups.}

\item{uv.groups}{Character. A symbol or a vector of symbols indicating the column names that contain known unwanted
variable(s) in the SummarizedExperiment object. These variables can be categorical or continuous. Continuous variables will be
divided into 'nb.bio.clusters' groups based on the clustering method specified in the 'bio.clustering.method' argument.}

\item{uv.clustering.method}{Character. A symbol specifying the clustering method to be used for grouping continuous sources
of unwanted variation. The default is 'kmeans'. Refer to the 'createHomogeneousUvGroups' function for more details.}

\item{nb.uv.clusters}{Numeric. A numeric value indicating the number of clusters for each continuous source of unwanted
variation. The default is 2, meaning each continuous source of unwanted variation will be divided into 2 groups.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data before performing any statistical
analysis. The default is 'TRUE'.}

\item{pseudo.count}{Numeric. A value to be added as a pseudo count to all measurements before log transformation. The
default is 1.}

\item{normalization}{Character. Specifies which normalization method should be applied to the data before finding genes
affected by biological variation. The default is 'CPM'. If set to 'NULL', no normalization will be applied.
Refer to the 'applyOtherNormalizations' function for more details.}

\item{grid.nb}{Numeric. A numeric value (description needed).}

\item{regress.out.uv.variables}{Character. A symbol or a vector of symbols indicating the columns that contain unwanted
variation variables in the SummarizedExperiment object. These variables will be regressed out from the data before identifying
genes highly affected by biological variation. The default is NULL, meaning regression will not be applied.}

\item{assess.bio.genes}{Logical. Indicates whether to assess the performance of selected NCGs (biological genes). This
analysis involves principal component analysis (PCA) on the selected NCGs, followed by exploring the R² or vector correlation
between the first 'nb.pcs' principal components and the specified variables.}

\item{variables.to.assess.bio.genes}{Character. A symbol or a vector of symbols indicating the column names of the
SummarizedExperiment object that contain variables whose association with the selected NCGs needs to be evaluated.
The default is NULL, meaning all variables in 'bio.variables' and 'uv.variables' will be assessed.}

\item{nb.pcs}{Numeric. Indicates the number of principal components (PCs) to be used to assess the performance of the selected
NCGs. The default is 5.}

\item{center}{Logical. Indicates whether to center the data before applying principal component analysis (PCA). The default is 'TRUE'.
Refer to the 'computePCA' function for more details.}

\item{scale}{Logical. Indicates whether to scale the data before applying PCA. The default is 'FALSE'.
Refer to the 'computePCA' function for more details.}

\item{anova.method}{Character. Specifies the ANOVA method to be used to compute the association between gene-level
expression and other variables.}

\item{min.sample.for.aov}{Numeric. Specifies the minimum number of samples required in each group before applying ANOVA.
The default is 3.}

\item{corr.method}{Character. Specifies the correlation method to be used to compute the association between gene-level
expression and a continuous variable. The default is 'spearman'.}

\item{a}{Numeric. The significance level for confidence intervals in the correlation. The default is 0.05.}

\item{rho}{Numeric. The hypothesized correlation value to be used in hypothesis testing. The default is 0.}

\item{min.sample.for.correlation}{Numeric. Specifies the minimum number of samples required in each group before applying
the correlation analysis. The default is 10.}

\item{min.sample.for.mad}{Numeric. Specifies the minimum number of samples required to perform the Median Absolute
Deviation (MAD) calculation on each gene within homogeneous sample groups. The default is 3.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment object. If 'TRUE', the 'checkSeObj' function
will be applied. The default is 'TRUE'.}

\item{assess.variables}{Logical. Indicates whether to assess the correlation between biological and unwanted variation
variables separately when creating homogeneous sample groups. The default is 'FALSE'. Refer to the function 'assessVariablesAssociation'
for more details.}

\item{cat.cor.coef}{Numeric vector. Specifies the cut-off correlation coefficient between pairs of categorical variables.
The first value applies to correlations between 'uv.variables', and the second to correlations between 'bio.variables'.
Correlation is computed using the ContCoef function from the DescTools R package. If the correlation exceeds the cut-off,
only the variable with the highest number of factors will be retained. By default, both values are set to 0.9.}

\item{cont.cor.coef}{Numeric vector. Specifies the cut-off correlation coefficient between pairs of continuous variables.
The first value applies to correlations between 'uv.variables', and the second to correlations between 'bio.variables'.
Correlation is computed using the ContCoef function from the DescTools R package. If the correlation exceeds the cut-off,
only the variable with the highest variance will be retained. By default, both values are set to 0.9.}

\item{remove.na}{Character. Specifies whether to remove NA or missing values from 'assays', 'sample.annotation', 'both', or 'none'.
If 'assays' is selected, genes with missing values will be excluded. If 'sample.annotation' is selected, samples with missing
values for any 'bio.variables' or 'uv.variables' will be excluded. The default is 'none'.}

\item{save.se.obj}{Logical. Indicates whether to save the results in the metadata of the SummarizedExperiment object or output
as a logical vector. The default is 'TRUE'. The result will be saved in 'se.obj@metadata$NCG$supervised$output.name'.}

\item{output.name}{Character. The name of the output. If set to 'NULL', the function will automatically generate a name.
The default name is constructed as paste0(sum(ncg.selected),'|', paste0(bio.variables, collapse = '&'), '|', paste0(uv.variables,
collapse = '&'),'|TWAnova:', '|', assay.name).}

\item{save.imf}{Logical. Indicates whether to save the intermediate file in the SummarizedExperiment object. If 'TRUE',
the results of the two-way ANOVA will be saved, allowing users to adjust parameters without recalculating the ANOVA.
This speeds up parameter tuning for NCG selection. The default is 'FALSE'.}

\item{imf.name}{Character. The name of the intermediate file to be saved. If set to 'NULL', the function will generate a name,
such as paste0(assay.name, '|TwoWayAnova|'). The default is 'NULL'.}

\item{use.imf}{Logical. Indicates whether to use the intermediate file. The default is 'FALSE'.}

\item{verbose}{Logical. If 'TRUE', displays messages about the steps being executed.}
}
\description{
This function uses different approaches to find genes that are highly affected by biological variables or highly
variable genes. Refer to the details for more information.
}
\details{
Additional details...
\itemize{
\item TwoWayAnov: In this process, all the biological and unwanted variables will be grouped into two categorical variables
separately using the createHomogeneousBioGroups() and createHomogeneousUVGroups() functions. Then, a two-way ANOVA is
applied to the expression levels of individual genes, considering the summarized biological and unwanted variables as
two factors. A set of genes with the highest F-statistic the biological variables will be selected as biological genes.
\item AnovaCorr.AcrossAllSamples: computes gene-level correlation and ANOVA to find genes that are highly affected by continuous
and categorical sources biological variables across all samples, respectively.
Then, a set of genes with highest absolute correlation coefficients and F-statistics will be selected as biological.
\item AnovaCorr.PerBatchPerBio: computes gene-level correlation and ANOVA within groups of samples that are homogeneous
with respect to biological or unwanted variation. This is useful in situations where the biological and unwanted variation
are highly correlated.
\item mad.unsupervised: If sources of unwanted variation are unknown, the identifyUnknownUV must be applied to
estimate them. Subsequently, the functions conducts a Median Absolute Deviation (MAD) analysis on each gene within
sample groups homogeneous with respect to unwanted variables, to pinpoint genes highly
variable due to biological factors. The higher the MAD, the more likely the genes are to be biologically variable.
}
}
\author{
Ramyar Molania
}
