% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNcgPerBiologyPerBatch.R
\name{findNcgPerBiologyPerBatch}
\alias{findNcgPerBiologyPerBatch}
\title{Finds negative control genes using ANOVA and correlation per sample groups.}
\usage{
findNcgPerBiologyPerBatch(
  se.obj,
  assay.name,
  bio.variables,
  uv.variables,
  ncg.selection.method = "non.overlap",
  nb.ncg = 0.1,
  top.rank.bio.genes = 0.5,
  top.rank.uv.genes = 0.5,
  bio.percentile = 0.2,
  uv.percentile = 0.2,
  grid.group = "uv",
  grid.direction = "decrease",
  grid.nb = 20,
  min.sample.for.aov = 3,
  min.sample.for.correlation = 10,
  regress.out.bio.variables = NULL,
  regress.out.uv.variables = NULL,
  bio.groups = NULL,
  bio.clustering.method = "kmeans",
  nb.bio.clusters = 3,
  uv.groups = NULL,
  uv.clustering.method = "kmeans",
  nb.uv.clusters = 3,
  normalization = "CPM",
  apply.log = TRUE,
  pseudo.count = 1,
  corr.method = "spearman",
  a = 0.05,
  rho = 0,
  anova.method = "aov",
  assess.ncg = TRUE,
  variables.to.assess.ncg = NULL,
  nb.pcs = 5,
  center = TRUE,
  scale = FALSE,
  assess.se.obj = TRUE,
  remove.na = "none",
  save.se.obj = TRUE,
  output.name = NULL,
  ncg.group = NULL,
  plot.output = TRUE,
  save.imf = FALSE,
  imf.name = NULL,
  use.imf = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.name}{A character string. The name of the assay in the SummarizedExperiment object to be used for selecting
negative control genes.}

\item{bio.variables}{A character vector. The column names that contain biological variables in the SummarizedExperiment
object. Both continuous and categorical variables can be provided.}

\item{uv.variables}{A character vector. The column names that contain unwanted variation (UV) variables in the
SummarizedExperiment object.}

\item{ncg.selection.method}{A character string. Indicates how to select the genes as NCGs. For individual genes,
the two-way ANOVA calculates F-statistics for biological and unwanted variation factors separately. An ideal NCG set
should have high F-statistics for the unwanted variation variables and low F-statistics for the biological variables.
The function ranks the F-statistics obtained for the biological variables and the negative of the F-statistics obtained
for the unwanted variables. The function then offers 5 ways to summarize the ranks of the two F-statistics:
'Prod' (product of the ranks), 'Sum' (sum of the ranks), 'Average' (average of the ranks), 'AbsNoneOverlap' (non-overlapping genes
between 'top.rank.uv.genes' and 'top.rank.bio.genes'), and 'NoneOverlap' (non-overlapping genes between 'top.rank.uv.genes'
and at least 'top.rank.bio.genes'). The F-statistics for biological and UV are first ranked.}

\item{nb.ncg}{Numeric. The proportion of the total genes to be selected as a Negative Control Gene (NCG) set. The
default is .1.}

\item{top.rank.bio.genes}{Numeric. The percentage of top-ranked genes that are highly affected by the biological variation.
This parameter is required when \code{ncg.selection.method} is either 'NoneOverlap' or 'AbsNoneOverlap'.}

\item{top.rank.uv.genes}{Numeric. The percentage of top-ranked genes that are highly affected by the unwanted variation variables.
This parameter is required when \code{ncg.selection.method} is either 'NoneOverlap' or 'Auto'.}

\item{bio.percentile}{Numeric. The percentile cut-off for selecting genes that are highly affected by the biological
variation. The default is set to 0.8.}

\item{uv.percentile}{Numeric. The percentile cut-off for selecting genes that are highly affected by the unwanted
variation. The default is set to 0.8.}

\item{grid.group}{A character string. Indicates whether the grid search should be performed on biological ('top.rank.bio.genes'),
unwanted ('top.rank.uv.genes') or both factors. The options are 'bio', 'uv', or 'both'. The default is set to 'uv'.}

\item{grid.direction}{A character string. Indicates the direction of the grid search. The options are 'increase' or 'decrease'.
The default is set to 'decrease'.}

\item{grid.nb}{Numeric. The percentage for grid search when the 'ncg.selection.method' is 'NoneOverlap'. In the 'NoneOverlap'
approach, the grid search starts with the initial \code{top.rank.uv.genes} value and adds \code{grid.nb} in each loop to find
the desired 'nb.ncg'.}

\item{min.sample.for.aov}{Numeric. Indicates the minimum number of samples to be present in each group before applying
the ANOVA. The default is set to 3.}

\item{min.sample.for.correlation}{Numeric. Indicates the minimum number of samples to be considered in each group before
applying correlation analysis. The default is set to 10.}

\item{regress.out.bio.variables}{A character vector. Indicates the column names of biological variables in the
SummarizedExperiment object. These variables will be regressed out from the data before identifying genes that are
highly affected by unwanted variation. The default is NULL.}

\item{regress.out.uv.variables}{A character vector. Indicates the column names of unwanted variation variables in the
SummarizedExperiment object. These variables will be regressed out from the data before identifying genes that are
highly affected by biological variation. The default is NULL.}

\item{bio.groups}{A character vector or a symbol. Indicates the column names that contain biological variables in the
SummarizedExperiment object.
If not NULL, bio.groups' will be used to group samples into different homogeneous biological groups.}

\item{bio.clustering.method}{A character string. Indicates which clustering method should be used to group continuous
sources of biological variation if any are provided. The default is 'kmeans' clustering.}

\item{nb.bio.clusters}{Numeric. Indicates the number of clusters for each continuous source of biological variation.
The default is set to 3.}

\item{uv.groups}{A character vector or a symbol. Indicates the column names that contain unwanted variation variables in the
SummarizedExperiment object. If not NULL, \code{uv.groups} will be used to group samples into different homogeneous unwanted
variation groups.}

\item{uv.clustering.method}{A character string. Indicates which clustering method should be used to group continuous
sources of unwanted variation. The default is set to 'kmeans' clustering.}

\item{nb.uv.clusters}{Numeric. Indicates the number of clusters for each continuous source of unwanted variation (UV).
By default, it is set to 2.}

\item{normalization}{A character string. Indicates which normalization method should be applied to the data before
identifying genes that are affected by biological variation. The default is 'CPM'. Refer to the 'applyOtherNormalizations
function for more details.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data before performing correlation and
ANOVA. The default is TRUE.}

\item{pseudo.count}{Numeric. A value to be added as a pseudo-count to all measurements before log-transformation.}

\item{corr.method}{A character string. Indicates which correlation method should be used to compute associations between
gene-level expression and a continuous variable. The default is set to 'spearman'.}

\item{a}{Numeric. The significance level used for the confidence intervals in the correlation analysis. The default
is set to 0.05.}

\item{rho}{Numeric. The hypothesized correlation value to be used in the hypothesis testing. The default is set to 0.}

\item{anova.method}{A character string. Indicates which ANOVA method should be used to compute associations between
gene-level expression and a categorical variable. The default is 'aov'.}

\item{assess.ncg}{Logical. Indicates whether to assess the performance of selected NCGs. This analysis involves
principal component analysis (PCA) on the selected NCGs, followed by exploring the R-squared or vector correlation between
the first 'nb.pcs' principal components and biological and unwanted variation variables.}

\item{variables.to.assess.ncg}{A character vector. Indicates the column names of the SummarizedExperiment object that
contain variables whose association with the selected NCGs needs to be evaluated. The default is NULL, meaning all
variables in 'bio.variables' and 'uv.variables' will be assessed.}

\item{nb.pcs}{Numeric. Indicates the number of the first principal components of selected NCGs to be used to assess
the performance of NCGs.}

\item{center}{Logical. Indicates whether to center the data before applying principal component analysis. The default
is set to 'TRUE'.}

\item{scale}{Logical. Indicates whether to scale the data before applying principal component analysis. The default is
set to 'FALSE'.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment object. The default is set to  'TRUE'.}

\item{remove.na}{A character string. Indicates whether to remove NA or missing values from either the 'assays', 'sample.annotation',
'both', or 'none'. If 'assays' is selected, genes containing NA or missing values will be excluded. If 'sample.annotation'
is selected, samples with NA or missing values for any \code{bio.variables} and \code{uv.variables} will be excluded. The default
is set to  'both'.}

\item{save.se.obj}{Logical. Indicates whether to save the result in the metadata of the SummarizedExperiment object
('se.obj') or to output the result.
The default is TRUE.}

\item{output.name}{A character string. The name for the output. If set to NULL, the function will automatically generate a name.}

\item{ncg.group}{A character string. Indicates the name of the group of NCGs.}

\item{plot.output}{Logical. If TRUE, a plot of the NCG selection process will be generated.}

\item{save.imf}{Logical. Indicates whether to save the intermediate file in the SummarizedExperiment object. If set
to TRUE, the results of the two-way ANOVA
will be saved. If users wish to adjust parameters such as \code{nb.ncg}, \code{ncg.selection.method}, \code{top.rank.bio.genes},
and \code{top.rank.uv.genes}, the ANOVA
will not be recalculated. This accelerates parameter tuning for NCG selection. The default is FALSE.}

\item{imf.name}{A character string. The name to use when saving the intermediate file. If set to NULL, the function
will automatically generate a name.
The default is NULL.}

\item{use.imf}{Logical. Indicates whether to use the intermediate file for subsequent steps. The default is FALSE.}

\item{verbose}{Logical. If TRUE, process messages will be displayed.}
}
\value{
Either the SummarizedExperiment object containing a set of negative control genes, or a logical vector of
the selected negative control genes.
}
\description{
This function employs correlation and ANOVA analyses across within groups of samples that are homogeneous with respect
to biological or unwanted variation. Correlation analysis is utilized to identify genes highly affected by continuous
sources of variation, while ANOVA is used to identify genes affected by categorical sources of variation.
The function selects genes as negative control genes (NCG) based on high correlation coefficients and F-statistics for
unwanted sources of variation, and low correlation coefficients and F-statistics for biological sources of variation.
Various approaches are employed for the final gene selection; please refer to the details for more information.
}
\details{
The function uses 5 ways to summarize two gene-level F-statistics obtained for the biological and unwanted variation
. The function uses either the values or the ranks of F-statistics for NCGs selection. The function ranks the
negative of F-statistics values for unwanted variation. The lower the ranks, the greater the impact of unwanted
variation on genes. The function ranks the F-statistics for biological variation. The higher the ranks, the greater
the impact of biological variation on genes. The options are 'prod', sum', 'average', 'auto' or 'non.overlap' and
'quantile'.

If 'prod', 'sum' and 'average' is set:

The product, sum or average of ranks of F-statistics is calculated. Then, the function selects 'nb.ncg' 'numbers of
genes as negative control genes that have the lowest ranks.

If 'non.overlap' is selected:
\enumerate{
\item The function selects the top ‘top.rank.bio.genes’ genes that have the highest ranks of F-statistics
for biological variation.
\item The function selects the top ‘top.rank.uv.genes’ genes that have the lowest ranks of F-statistics for
unwanted variation.
\item The function excludes all genes obtained in 2 from the ones obtained 1. This will be a set of genes as
negative control genes.
}

If 'auto' is selected:
\enumerate{
\item The function selects the top ‘top.rank.bio.genes’ genes that have the highest ranks of F-statistics for
biological variation.
\item  The function selects the top ‘top.rank.uv.genes’ genes that have the lowest ranks of F-statistics  for
unwanted variation.
\item The function excludes all genes obtained in 2 from the ones obtained 1.
\item If the number of selected genes is larger or smaller than the specified ‘nb.ncg’, the function applies an
auto search to find approximate ‘nb.ncg’ of genes as negative control genes as follow. The auto search will either
decrease or increase the values of either ‘top.rank.bio.genes’ or ‘top.rank.uv.genes’ or both till to find
approximate ‘nb.ncg’ of genes as negative control genes.
}
If 'quantile' is selected:
\enumerate{
\item The function selects the ‘bio.percentile’ percentile of F-statistics for biological variation. Then, selects
all the genes that have F-statistics larger the calculated percentile.
\item The function selects the ‘uv.percentile’ percentile of F-statistics for unwanted variation. Then, selects
all the genes that have F-statistics larger the calculated percentile.
\item The function excludes all genes obtained in 2 from the ones obtained 1.
}

The data pre-processing for the functions findNcgAcrossSamples() and findNcgPerBiologyPerBatch().
The input data for correlation and ANOVA analyses can be one or a combination of the following options:
\enumerate{
\item item $log_2$ (raw data + pseudo-count).
\item Apply some library size normalizations (if the input data is gene count expression).
\item This will facilitate finding genes that are highly affected by biological variation. item Regress out
biological variation.
\item This will facilitate finding genes that are highly affected by unwanted variation.
item Regress out unwanted variation.
}
}
\author{
Ramyar Molania
}
