% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNcgUnSupervised.R
\name{findNcgsUnSupervised}
\alias{findNcgsUnSupervised}
\title{Finds negative control genes (NCG) using an unsupervised approach.}
\usage{
findNcgsUnSupervised(
  se.obj,
  assay.name,
  uv.variables,
  nb.ncg = 0.1,
  ncg.selection.method = "non.overlap",
  top.rank.bio.genes = 0.5,
  top.rank.uv.genes = 0.5,
  bio.percentile = 0.2,
  uv.percentile = 0.2,
  grid.group = "uv",
  grid.direction = "decrease",
  grid.nb = 20,
  clustering.method = "kmeans",
  nb.clusters = 3,
  normalization = "CPM",
  regress.out.variables = NULL,
  min.sample.for.mad = 3,
  min.sample.for.aov = 3,
  min.sample.for.correlation = 10,
  corr.method = "spearman",
  a = 0.05,
  rho = 0,
  anova.method = "aov",
  assess.ncg = TRUE,
  variables.to.assess.ncg = NULL,
  nb.pcs = 5,
  scale = FALSE,
  center = TRUE,
  apply.log = TRUE,
  pseudo.count = 1,
  assess.se.obj = TRUE,
  remove.na = "both",
  save.se.obj = TRUE,
  output.name = NULL,
  ncg.group = NULL,
  plot.output = TRUE,
  use.imf = FALSE,
  save.imf = FALSE,
  imf.name = NULL,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.name}{Character. The name of the assay in the SummarizedExperiment object to be used for RUV-III-PRPS
normalization.}

\item{uv.variables}{Character. A character vector indicating the names of the columns in the SummarizedExperiment object
that contain unwanted variables. If all unwanted variation is unknown, the 'identifyUnknownUV' function can estimate them.}

\item{nb.ncg}{Numeric. Specifies the number of genes to be chosen as negative control genes (NCG) when the '
ncg.selection.method' is set to 'auto'. This value corresponds to a fraction of the total genes in the SummarizedExperiment
object. The default is 0.1.}

\item{ncg.selection.method}{Character. Specifies the method used to select negative control genes (NCG). The available options
are 'prod', 'sum', 'average', 'non.overlap', 'quantile', and 'auto'. The default method is 'non.overlap'.}

\item{top.rank.bio.genes}{Numeric. Specifies the fraction of top-ranked genes that are highly affected by biological variation.
These genes have the highest MAD values. The default is 0.5.}

\item{top.rank.uv.genes}{Numeric. Specifies the fraction of top-ranked genes that are highly affected by unwanted variation
variables. These genes have the highest F-statistics and correlation coefficients. The default is 0.5.}

\item{bio.percentile}{Numeric. The percentile cut-off for selecting genes that are highly affected by biological variation.
The default is 0.8.}

\item{uv.percentile}{Numeric. The percentile cut-off for selecting genes that are highly affected by unwanted variation.
The default is 0.8.}

\item{grid.group}{Character. Specifies whether the grid search should focus on biological, unwanted, or both factors. The
options are 'bio', 'uv', or 'both'.}

\item{grid.direction}{Character. Specifies whether the grid search should proceed in decreasing or increasing order.
The options are 'increase' or 'decrease'.}

\item{grid.nb}{Numeric. Specifies the number of genes to consider during the grid search when the 'ncg.selection.method'
is set to 'auto'. The default is 20.}

\item{clustering.method}{Character. Specifies the clustering method to use for grouping continuous sources of unwanted variation.
The available options are 'kmeans', 'cut', and 'quantile'. The default is 'kmeans'.}

\item{nb.clusters}{Numeric. The number of clusters for grouping continuous sources of unwanted variation. The default is 3.}

\item{normalization}{Character. Specifies the normalization method used for library size adjustment before identifying
genes that are highly affected by biological variation. The default is 'CPM'. Refer to the 'applyOtherNormalization'
function for further details.}

\item{regress.out.variables}{Character. Specifies which variables to regress out from the data prior to analysis.}

\item{min.sample.for.mad}{Numeric. The minimum number of samples required to perform MAD analysis on each gene within homogeneous
sample groups. The default is 3.}

\item{min.sample.for.aov}{Numeric. The minimum number of samples required for ANOVA analysis between categorical sources
of variation with individual gene expression. The default is 3.}

\item{min.sample.for.correlation}{Numeric. The minimum number of samples required for correlation analyses between continuous
sources of unwanted variation and individual gene expression. The default is 10.}

\item{corr.method}{Character. The correlation method to use for the analysis. Options are 'pearson' or 'spearman'. The default
is 'spearman'.}

\item{a}{Numeric. The significance level used for the confidence intervals in the correlation analysis. The default is 0.05.}

\item{rho}{Numeric. The hypothesized correlation value to be used in the hypothesis testing. The default is 0.}

\item{anova.method}{Character. The ANOVA method to use. Options are 'aov' or 'welch'. The default is 'aov'. Refer to the function
'computeGenesVariableAnova' for more details.}

\item{assess.ncg}{Logical. Specifies whether to assess the performance of the selected NCGs. This involves principal component
analysis on the selected NCGs and exploration of R^2 or vector correlation between the first 'nb.pcs' principal components
and the biological and unwanted variables.}

\item{variables.to.assess.ncg}{Character. The column names of the SummarizedExperiment object that contain variables to be
assessed for their association with the selected NCGs. By default, all variables in 'uv.variables' will be assessed.}

\item{nb.pcs}{Numeric. Specifies the number of principal components from the selected NCGs to be used for performance assessment.
The default is 5.}

\item{scale}{Logical. Specifies whether to scale the data before applying SVD. The default is FALSE.}

\item{center}{Logical. Specifies whether to center the data before applying Singular Value Decomposition (SVD). The
default is set to 'TRUE'.}

\item{apply.log}{Logical. Specifies whether to apply a log-transformation to the data. The default is TRUE.}

\item{pseudo.count}{Numeric. The pseudo count to be added to all measurements before log transformation.}

\item{assess.se.obj}{Logical. Specifies whether to assess the SummarizedExperiment object. The default is TRUE.}

\item{remove.na}{Character. Specifies whether to remove missing values (NA) from the 'assays', the 'sample.annotation',
'both', or 'none'. The default is set to 'both'.}

\item{save.se.obj}{Logical. Specifies whether to save the result in the metadata of the SummarizedExperiment object or
output the result. The default is set to 'TRUE'.}

\item{output.name}{Character. Specifies the name for the output in the metadata of the SummarizedExperiment object. The
default is set to NULL, meaning the function will generate a name based on the specified argument.}

\item{ncg.group}{Character. Specifies the name of the group of NCGs. If NULL, the function will generate a name using
"ncg|unsupervised".}

\item{plot.output}{Logical. If TRUE, a plot of the NCG assessment will be displayed while the function is running.}

\item{use.imf}{Logical. Specifies whether to use the intermediate file. The default is FALSE.}

\item{save.imf}{Logical. Specifies whether to save the intermediate file. If TRUE, the function will save the results
of the ANOVA. If users want to change parameters such as 'nb.ncg', 'ncg.selection.method', 'top.rank.bio.genes', or
'top.rank.uv.genes', the analysis will not be recalculated.}

\item{imf.name}{Character. Specifies the name for saving the intermediate file. If NULL, the function will generate a name.}

\item{verbose}{Logical. If TRUE, process messages will be displayed.}
}
\value{
A SummarizedExperiment object containing the selected negative control genes or a logical vector indicating
the selected genes.
}
\description{
This function identifies a set of genes as negative control genes in cases where no biological variations are known.
The function uses gene-level correlation, ANOVA, and MAD analyses across and between sample groups to find a set
of genes as negative control genes for the RUV-III-PRPS normalization.
}
\details{
The function initially utilizes gene-level correlation and ANOVA to identify genes significantly influenced by continuous
and categorical sources of variation, respectively. Subsequently, it conducts Median Absolute Deviation (MAD) analysis
on each gene within homogeneous sample groups, considering unwanted variables, to pinpoint genes highly variable due
to biological factors. Lastly, various methods are employed to consolidate the statistical findings, ultimately determining
an appropriate set of genes as negative control genes for RUV-III-PRPS normalization.
}
\author{
Ramyar Molania
}
