% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findRepeatingPatterns.R
\name{findRepeatingPatterns}
\alias{findRepeatingPatterns}
\title{is used to find repeating patterns}
\usage{
findRepeatingPatterns(vector, n)
}
\arguments{
\item{vector}{TO BE DEFINED}

\item{n}{TO BE DEFINED}
}
\description{
is used to find repeating patterns
}
