% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findUnSupervisedNCGs.R
\name{findUnSupervisedNCGs}
\alias{findUnSupervisedNCGs}
\title{Find a set of negative control genes (NCG).}
\usage{
findUnSupervisedNCGs(
  se.obj,
  assay.name,
  uv.variables,
  nb.ncg = 10,
  ncg.selection.method = "AbsNoneOverlap",
  grid.nb = 1,
  top.rank.bio.genes = 50,
  top.rank.uv.genes = 50,
  min.sample.for.aov = 3,
  min.sample.for.correlation = 10,
  normalization = "CPM",
  corr.method = "spearman",
  a = 0.05,
  rho = 0,
  anova.method = "aov",
  clustering.method = "kmeans",
  nb.clusters = 3,
  assess.ncg = TRUE,
  variables.to.assess.ncg = NULL,
  nb.pcs = 5,
  scale.pca = FALSE,
  center.pca = TRUE,
  apply.log = TRUE,
  pseudo.count = 1,
  assess.se.obj = TRUE,
  assess.variables = TRUE,
  cat.cor.coef = c(0.95, 0.95),
  cont.cor.coef = c(0.95, 0.95),
  save.se.obj = TRUE,
  remove.na = "both",
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.name}{Symbol. Indicates a name of an assay in the SummarizedExperiment object. The selected assay should
be the one that will be used for RUV-III-PRPS normalization.}

\item{uv.variables}{Symbols. Indicates the columns names that contains UV variables in the SummarizedExperiment object.}

\item{nb.ncg}{Numeric. Indicates how many genes should be selected as NCG. The value is the percentage of the total
genes in the SummarizedExperiment object. The default is 10 percent.}

\item{ncg.selection.method}{Symbol.Indicates how to select a set genes as NCG.
For individual genes, the two-way ANOVA calculates F-statistics for biological and
unwanted variation factors separately. An ideal NCG set should have high F-statistics
for the unwanted variation variables and low F-statistics for the biological variables.
The function ranks the F-statistics obtained for the biological variable and negative of
the F-statistics obtained for the unwanted variables. Then this functions offers 5 ways to
summarize the ranks of the two F-statistics. Prod' is the product of the ranks. 'Sum', is
the sum of the ranks. 'Average' is the average of the ranks. 'AbsNoneOverlap' is the none overlapped
genes of the 'top.rank.uv.genes' and 'top.rank.bio.genes'. 'noneOverlap' is the none overlapped genes
of the 'top.rank.uv.genes' and at least 'top.rank.bio.genes'. The F-statistics for biological and UV
are first ranked.Then options are Prod (product), Sum, Average,}

\item{grid.nb}{Numeric. Indicates the percentage for grid search when the ncg.selection.method is
'noneOverlap'. In the 'noneOverlap' approach, the grid search starts with the initial
'top.rank.uv.genes' value an add the grid.nb in each loop to find the 'nb.ncg'.}

\item{top.rank.bio.genes}{Numeric.Indicates the percentage of top ranked genes that are highly affected
by the biological variation. This is required to be specified when the 'ncg.selection.method' is
either 'noneOverlap' or 'AbsNoneOverlap'.}

\item{top.rank.uv.genes}{Numeric.Indicates the percentage of top ranked genes that are highly affected
by the unwanted variation variables. This is required to be specified when the 'ncg.selection.method' is
either 'noneOverlap' or 'AbsNoneOverlap'.}

\item{min.sample.for.aov}{Numeric.Indicates the minimum number of samples to perform correlation analyses between continuous sources
of variation (biological and unwanted variation) with individual gene expression.}

\item{min.sample.for.correlation}{Numeric.Indicates the minimum number of samples to perform correlation analyses between continuous sources
of variation (biological and unwanted variation) with individual gene expression.}

\item{normalization}{Symbols.Indicates winch normalization method to use for library size adjustment before fining genes that are highly
affected by biological variation. The default is CPM. Refer to the 'applyOtherNormalization' function for further details.}

\item{corr.method}{Numeric.Indicates which correlation methods (pearson or spearman) should be used for the correlation analyses. The
default is 'spearman'.}

\item{a}{The significance level used for the confidence intervals in the correlation, by default it is set to 0.05.}

\item{rho}{The value of the hypothesised correlation to be used in the hypothesis testing, by default it is set to 0.}

\item{anova.method}{Symbols. Indicates which ANOVA methods to use. The default in 'aov'. Refer to the function "genesVariableAnova" for more details.}

\item{clustering.method}{TTTTT}

\item{nb.clusters}{TTTT}

\item{assess.ncg}{Logical. Indicates whether to assess the performance of selected NCG or not.
This analysis involves principal component analysis on the selected NCG and
then explore the R^2 or vector correlation between the 'nb.pcs' first principal
components and with biological and unwanted variables.}

\item{variables.to.assess.ncg}{Symbols. Indicates the column names of the SummarizedExperiment object that
contain variables whose association with the selected genes as NCG.
needs to be evaluated. The default is NULL. This means all the variables in the 'uv.variables' will be assessed.}

\item{nb.pcs}{Numeric. Indicates the number of the first principal components on selected NCG to be used to assess the performance of NCGs.}

\item{scale.pca}{Logical. Indicates whether to scale the data or not before applying SVD.  If scale is TRUE, then scaling
is done by dividing the (centered) columns of the assays by their standard deviations if center is TRUE, and the root
mean square otherwise. The default is FALSE.}

\item{center.pca}{Logical. Indicates whether to scale the data or not before applying SVD. If 'center.pca' is TRUE, then
centering is done by subtracting the column means of the assay from their corresponding columns. The default is TRUE.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data, by default it is set to TRUE.}

\item{pseudo.count}{Numeric. A value as a pseudo count to be added to all measurements before log transformation.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment class object or not.}

\item{assess.variables}{Logical. Indicates whether to assess the association between the biological and unwanted
variation variables separately. Refer to the variablesCorrelation function for more details. Th default is FALSE.}

\item{cat.cor.coef}{Vector of two numerical values. Indicates the cut-off of the correlation coefficient between each pair of
categorical variables. The first one is between each pair of 'uv.variables' and the second one is between each pair of 'bio.variables'.
The correlation is computed by the function ContCoef from the DescTools package. If the correlation of a pair of variable is higher than
the cut-off, then only the variable that has the highest number of factor will be kept and the other one will be excluded from the
remaining analysis. By default they are both set to 0.7.}

\item{cont.cor.coef}{Vector of two numerical values. Indicates the cut-off of the Spearman correlation coefficient between each pair of
continuous variables. The first one is between each pair of 'uv.variables' and the second one is between each pair of 'bio.variables'.
If the correlation of a pair of variable is higher than the cut-off, then only the variable that has the highest variance will
be kept and the other one will be excluded from the remaining analysis. By default they are both set to 0.7.}

\item{save.se.obj}{Logical. Indicates whether to save the result of the function in the metadata of the SummarizedExperiment object or
to output the result. The default is TRUE.}

\item{remove.na}{Symbol. Indicates whether to remove NA or missing values from either the 'assays', the 'sample.annotation',
'both' or 'none'. If 'assays' is selected, the genes that contains NA or missing values will be excluded. If 'sample.annotation' is selected, the
samples that contains NA or missing values for any 'uv.variables' will be excluded. By default, it is set to both'.}

\item{verbose}{Logical. If TRUE, displaying process messages is enabled.}
}
\value{
Either the SummarizedExperiment object containing the a set of negative control genes or a logical vector of
the selected negative control genes
}
\description{
This function uses the correlation and ANOVA analyses across all samples to find a set of genes as negative control
genes for RUV-III-PRPS normalization. The correlation and ANOVA are used to find genes that are highly affected by
continuous and categorical sources of variation respectively. The function selects genes as NCG that show possible
high correlation coefficients and F-statistics with the sources of unwanted variation and low correlation coefficients
and F-statistics with the sources of biological variation. The function uses different approaches to perform the final selection.
}
\author{
Ramyar Molania
}
