% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesDEA.R
\name{genesDEA}
\alias{genesDEA}
\title{is used to perform differential gene expression analysis of the assays in a SummarizedExperiment object.}
\usage{
genesDEA(
  se.obj,
  assay.names = "All",
  variable,
  apply.log = TRUE,
  pseudo.count = 1,
  assess.se.obj = TRUE,
  remove.na = "both",
  save.se.obj = TRUE,
  plot.output = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Optional string or list of strings for the selection of the name(s) of the assay(s) of the
SummarizedExperiment class object to compute the ANOVA. By default all the assays of the SummarizedExperiment object
will be selected.}

\item{variable}{String of the label of a categorical variable such as sample types or batches from colData(se.obj).}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data. By default
the log transformation will be selected.}

\item{pseudo.count}{Numeric. A value as a pseudo count to be added to all measurements before log transformation,
by default it is set to 1.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment class object.}

\item{remove.na}{TO BE DEFINED.}

\item{save.se.obj}{Logical. Indicates whether to save the result in the metadata of the SummarizedExperiment class object 'se.obj' or
to output the result. By default it is set to TRUE.}

\item{plot.output}{Logical. Indicates whether to plot the boxplot of the F-test statistics, by default it is set to TRUE.}

\item{verbose}{Logical. Indicates whether to show or reduce the level of output or messages displayed during the execution
of the functions, by default it is set to TRUE.}
}
\value{
SummarizedExperiment A SummarizedExperiment object containing the log2 F-statistics of ANOVA on the continuous variable
and if requested the associated boxplot.
}
\description{
is used to perform differential gene expression analysis of the assays in a SummarizedExperiment object.
}
