% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesVariableAnova.R
\name{genesVariableAnova}
\alias{genesVariableAnova}
\title{is used to compute the differential gene expression analysis between the gene expression (assay)
of a SummarizedExperiment class object and a categorical variable (i.e. batches) using ANOVA.}
\usage{
genesVariableAnova(
  se.obj,
  assay.names = "All",
  variable,
  apply.log = TRUE,
  method = "aov",
  save.se.obj = TRUE,
  boxplot.output = TRUE,
  plot.top.genes = FALSE,
  nb.top.genes = 3,
  assess.se.obj = TRUE,
  remove.na = "both",
  pseudo.count = 1,
  apply.round = TRUE,
  verbose = verbose
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object that will be used to compute the differential expression ANOVA analysis.}

\item{assay.names}{Optional string or list of strings for the selection of the name(s)
of the assay(s) of the SummarizedExperiment class object to compute the correlation. By default}

\item{variable}{String of the label of a categorical variable such as
sample types or batches from colData(se.obj).}

\item{apply.log}{Indicates whether to apply a log-transformation to the data. By default
the log transformation will be selected.}

\item{method}{A character string indicating which method
is to be used for the differential analysis: "aov" or "welch.correction". By default "aov" will
be selected.}

\item{save.se.obj}{Indicates whether to save the result in the metadata of the SummarizedExperiment class object 'se.obj' or
to output the result. By default it is set to TRUE.}

\item{boxplot.output}{Indicates whether to plot the boxplot of the F-test statistics, by default it is set to TRUE.}

\item{plot.top.genes}{Indicates whether to plot the gene expression of the number of genes
from the top listing of anova by default it is set to FALSE.}

\item{nb.top.genes}{Defines the number of genes from the top or bottom listing of anova to plot,
by default is set to 3.}

\item{assess.se.obj}{Indicates whether to assess the SummarizedExperiment class object.}

\item{remove.na}{TO BE DEFINED.}

\item{pseudo.count}{TO BE DEFINED.}

\item{apply.round}{TO BE DEFINED.}

\item{verbose}{Indicates whether to show or reduce the level of output or messages displayed during the execution
of the functions, by default it is set to TRUE.}
}
\value{
SummarizedExperiment A SummarizedExperiment object containing the log2 F-statistics of ANOVA on the continuous variable
and if requested the associated boxplot.
}
\description{
is used to compute the differential gene expression analysis between the gene expression (assay)
of a SummarizedExperiment class object and a categorical variable (i.e. batches) using ANOVA.
}
