% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetricFromSeObj.R
\name{getMetricFromSeObj}
\alias{getMetricFromSeObj}
\title{Get a metric or plot from the "metadata" of SummarizedExperiment object.}
\usage{
getMetricFromSeObj(
  se.obj,
  slot = "Metrics",
  assay.names,
  assessment.type = "gene.level",
  assessment,
  method = NULL,
  variables,
  file.name,
  sub.file.name = NULL,
  required.function,
  message.to.print,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{slot}{Symbol. A symbol indicating the name of the slots in the 'metadata' of the SummarizedExperiment object.}

\item{assay.names}{Symbol. A symbol or a vector of symbols specifying the name(s) of the assay(s) in the SummarizedExperiment
object. The default is set to 'all'.}

\item{assessment.type}{Symbol. A symbol indicating the type of assessment. Options are 'gene.level' or 'global.level'.}

\item{assessment}{Symbol. A symbol indicating the name of the metric to be checked.}

\item{method}{Symbol. A symbol indicating the method used to calculate the metric.}

\item{variables}{Symbol. A symbol or a vector of symbols indicating the variables used to calculate the metric.}

\item{file.name}{Symbol. A symbol indicating the file name to which the results of the metric are assigned.}

\item{sub.file.name}{Symbol. A symbol indicating the sub-file name to which the results of the metric are assigned.}

\item{required.function}{Symbol. A symbol indicating the name of function that is required to be applied before getting
the results of that metric.}

\item{message.to.print}{Symbol. A symbol to print.}

\item{verbose}{Logical. If 'TRUE', shows the messages of different steps of the function.}
}
\description{
Get a metric or plot from the "metadata" of SummarizedExperiment object.
}
\author{
Ramyar Molania
}
