% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupContinuousVariable.R
\name{groupContinuousVariable}
\alias{groupContinuousVariable}
\title{Group a continuous variable.}
\usage{
groupContinuousVariable(
  se.obj,
  variable,
  nb.clusters = 3,
  clustering.method = "kmeans",
  perfix = "_group",
  plot.output = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{variable}{Symbol. A symbol specifying the column in the SummarizedExperiment object. This should be a continuous
variable.}

\item{nb.clusters}{Numeric. A numeric value indicating the number of groups for continuous sources of unwanted variation.
The default is 3. This implies that each continuous source will be split into 3 groups using the specified
'clustering.method' method.}

\item{clustering.method}{Symbol. A symbol specifying the clustering method to be applied for grouping each continuous
source of unwanted variables. Options include 'kmeans', 'cut', and 'quantile'. The default is 'kmeans' clustering.}

\item{perfix}{Symbol. A symbol to add to each groups.}

\item{plot.output}{Logical. If is 'TRUE', the function plots the variable values colored by the groups.}

\item{verbose}{Logical. If 'TRUE', shows the messages of different steps of the function.}
}
\description{
Group a continuous variable.
}
\author{
Ramyar Molania
}
