% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_unwanted_variation.R
\name{identify_unwanted_variation}
\alias{identify_unwanted_variation}
\title{is used to identify the unwanted variation of a SummarizedExperiment class object.
Several assessment will be performed:
\enumerate{
\item PCA plot of each categorical variable.
\item Silhouette and ARI computed on categorical variable.
\item Combined Silhouette plot of the combined pair of categorical variable.
\item Linear regression between the first cumulative PC and continuous variable.
\item Spearman correlation between gene expression and continuous variable.
}}
\usage{
identify_unwanted_variation(
  se,
  assay_names = NULL,
  apply.log = FALSE,
  cat_var_label = NULL,
  cont_var_label = NULL,
  output_file = NULL,
  n.cores = 5
)
}
\arguments{
\item{se}{A SummarizedExperiment object that will be used to assess the performance of the normalisation of the data.}

\item{assay_names}{String or list of strings for selection of the name
of the assay of the SummarizedExperiment class object.}

\item{apply.log}{Indicates whether to apply a log-transformation to the data. By default
no transformation will be selected.}

\item{cat_var_label}{String or vector of strings of the label of categorical variable(s) such as
sample types or batches from colData(se).}

\item{cont_var_label}{String or vector of strings of the label of continuous variable(s)
such as library size from colData(se).}

\item{output_file}{Path and name of the output file to save the assessments plots in a pdf format.}

\item{n.cores}{is the number of cpus used for mclapply parallelization. Default is set to 5.}
}
\value{
plots List of assessments plots
}
\description{
is used to identify the unwanted variation of a SummarizedExperiment class object.
Several assessment will be performed:
\enumerate{
\item PCA plot of each categorical variable.
\item Silhouette and ARI computed on categorical variable.
\item Combined Silhouette plot of the combined pair of categorical variable.
\item Linear regression between the first cumulative PC and continuous variable.
\item Spearman correlation between gene expression and continuous variable.
}
}
