% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indentifyUnknownUV.R
\name{indentifyUnknownUV}
\alias{indentifyUnknownUV}
\title{is used to find potential unknown sources of unwanted variation..}
\usage{
indentifyUnknownUV(
  se.obj,
  assay.name,
  approach = "RLE",
  regress.out.bio.variables = NULL,
  regress.out.bio.gene.sets = NULL,
  uv.gene.sets = NULL,
  ncg = NULL,
  clustering.methods = "nbClust",
  nbClust.diss = NULL,
  nbClust.distance = "euclidean",
  nbClust.min.nc = 2,
  nbClust.max.nc = 5,
  nbClust.method = "kmeans",
  nbClust.index = "silhouette",
  nbClust.alphaBeale = 0.1,
  max.samples.per.batch = 0.1,
  nb.clusters = NULL,
  rle.comp = "median",
  apply.log = TRUE,
  pseudo.count = 1,
  nb.pcs = 2,
  center = TRUE,
  scale = FALSE,
  BSPARAM = bsparam(),
  remove.na = "none",
  assess.se.obj = TRUE,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.name}{A symbol indicates an assay name in SummarizedExperiment object.}

\item{approach}{A symbol indicates which approach should be used to identify unknown sources of unwanted variation. RLE:}

\item{regress.out.bio.variables}{One or more symbol indicate the names of columns in the sample annotation (colData) that will be regress out
from the data before finding sources of unwanted variation.}

\item{regress.out.bio.gene.sets}{List. A list of biological gene signatures. Individual gene set will be used to score samples and
then each scores will be regressed out from the data before finding negative control genes.}

\item{uv.gene.sets}{List. A list of unwanted variation gene signatures. Individual gene set will be used to score samples and
then each scores will be clustered to find possible unknown batches. The default is null.}

\item{ncg}{Logical or factor. Indicates a set of negative control genes to find sources of unwanted variation. If is not NULL,
the RLE or PCA and sample scoring will be based on only negative control genes. The default is null.}

\item{clustering.methods}{A symbol. Indicates winch clustering methods: 'kmeans', 'cut', 'quantile', 'nbClust' should be used to find
possible unknown batches. The default is nbClust.}

\item{nbClust.diss}{dissimilarity matrix to be used. By default, diss=NULL, but if it is replaced by a dissimilarity matrix, distance should be "NULL".}

\item{nbClust.distance}{the distance measure to be used to compute the dissimilarity matrix. This must be one of: "euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski" or "NULL".
By default, distance="euclidean". If the distance is "NULL", the dissimilarity matrix (diss) should be given by the user. If distance is not "NULL", the dissimilarity matrix should be "NULL".}

\item{nbClust.min.nc}{minimal number of clusters, between 1 and (number of objects - 1)}

\item{nbClust.max.nc}{minimal number of clusters, between 1 and (number of objects - 1)}

\item{nbClust.method}{maximal number of clusters, between 2 and (number of objects - 1), greater or equal to min.nc. By default, max.nc=15.}

\item{nbClust.index}{The index to be calculated. This should be one of : "kl", "ch", "hartigan", "ccc", "scott", "marriot", "trcovw", "tracew", "friedman", "rubin", "cindex", "db", "silhouette",
"duda", "pseudot2", "beale", "ratkowsky", "ball", "ptbiserial", "gap", "frey", "mcclain", "gamma", "gplus", "tau", "dunn", "hubert", "sdindex", "dindex", "sdbw", "all" (all indices except GAP, Gamma, Gplus and Tau),
"alllong" (all indices with Gap, Gamma, Gplus and Tau included).}

\item{nbClust.alphaBeale}{significance value for Beale's index.}

\item{max.samples.per.batch}{Numeric. Indicates the maximum number of samples per cluster when the clustering.methods is nbClust.
The default is .1 (10\%) of total samples in the SummarizedExperiment object.}

\item{nb.clusters}{Numeric. Indicates how many clusters should be found when clustering.methods kmeans, cut and quantile.}

\item{rle.comp}{A symbol. Indicates which properties: 'median' or 'IQR' or 'both' of the RLE data should be used for clustering when the approach is RLE.
The default is median.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data, by default it is set to TRUE.}

\item{pseudo.count}{Numeric. A value as a pseudo count to be added to all measurements before log transformation,
by default it is set to 1.}

\item{nb.pcs}{Numeric. A value to select the first principal components for clustering when the approach is set to 'PCA'.}

\item{center}{Logical. Indicates whether to center the data or not before performing PCA. The default is TRUE.}

\item{scale}{Logical. Indicates whether to scale the data or not before performing PCA. The default is FASLE.}

\item{BSPARAM}{TBBBB}

\item{remove.na}{A symbol. Indicates whether to remove NA or missing values from either the 'assays', 'sample.annotation', 'both' or 'none'.
If 'assays' is selected, the genes that contains NA or missing values will be excluded. If 'sample.annotation' is selected, the
samples that contains NA or missing values for any 'bio.variables' and 'uv.variables' will be excluded. By default, it is set to
'both'.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment class object.
By default it is set to TRUE.}

\item{save.se.obj}{Logical. Indicates whether to save the result in the metadata of the SummarizedExperiment class}

\item{verbose}{Logical. Indicates whether to show or reduce the level of output or messages displayed during the
execution of the functions, by default it is set to TRUE.}
}
\description{
is used to find potential unknown sources of unwanted variation..
}
