% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_assessment.R
\name{norm_assessment}
\alias{norm_assessment}
\title{is used to assess the performance of the normalisation of the data}
\usage{
norm_assessment(
  sce,
  apply.log = FALSE,
  biological_subtypes,
  library_size,
  time_effect,
  output_file = NULL
)
}
\arguments{
\item{sce}{Dataset that will be used to assess the performance of the normalisation of the data}

\item{apply.log}{Indicates whether to apply a log-transformation to the data}

\item{biological_subtypes}{Biological subtypes of each sample}

\item{library_size}{Library size of each sample}

\item{time_effect}{Time effect (plates/years ) of each sample}

\item{output_file}{Path and name of the output file to save the assessments plots in a pdf format}
}
\value{
plots List of assessments plots
}
\description{
is used to assess the performance of the normalisation of the data
}
