% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{normalise}
\alias{normalise}
\title{is used to normalise the gene expression (assay) of a SummarizedExperiment class object
using RUVIII-PRPS method for multiple k values in one go.}
\usage{
normalise(
  Y,
  M,
  ctl,
  k = NULL,
  eta = NULL,
  include.intercept = TRUE,
  average = FALSE,
  fullalpha = NULL,
  return.info = FALSE,
  inputcheck = TRUE
)
}
\arguments{
\item{Y}{A m by n matrix of the Raw gene expression matrix where m is the number of samples and n is the
number of features of a SummarizedExperiment variable to be normalised.}

\item{M}{Replicate matrix.}

\item{ctl}{Logical vector of length n of the negative control genes.}

\item{k}{A single value or a vector of values containing a single k or a range of k - the number of unwanted factors - to be tested.}

\item{eta}{A matrix with n columns, gene-wise (as opposed to sample-wise) covariates. By default is set to NULL.}

\item{include.intercept}{Add an intercept term to eta if it does not include one already. By default is set to TRUE.}

\item{average}{Average replicates after adjustment. By default is set to FALSE.}

\item{fullalpha}{Can be included to speed up execution. By default is set to NULL.}

\item{return.info}{If FALSE, only the adjusted data matrix is returned. If TRUE, additional information
is returned. By default is set to FALSE.}

\item{inputcheck}{Perform a basic sanity check on the inputs, and issue a warning if there is a problem.
By default is set to TRUE.}
}
\value{
list List containing the corrected gene expression matrix and the M, alpha and W for each k values tested.
}
\description{
is used to normalise the gene expression (assay) of a SummarizedExperiment class object
using RUVIII-PRPS method for multiple k values in one go.
}
