% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plot_squared.R
\name{pca_plot_squared}
\alias{pca_plot_squared}
\title{is used to plot PCA of the data colored by one variable}
\usage{
pca_plot_squared(
  pca,
  variable,
  variable.name,
  color,
  strokeSize = 0.2,
  pointSize = 0.8,
  strokeColor = "gray30",
  alpha = 0.5
)
}
\arguments{
\item{pca}{PCs of the dataset that will be used in the plot}

\item{variable}{The variable that will be used to display and color the PCA plot}

\item{variable.name}{The label of the variable that will be used on the PCA plot}

\item{color}{The color of the variable that will be used on the PCA plot}

\item{strokeSize}{geom_point aesthetics}

\item{pointSize}{geom_point aesthetics}

\item{strokeColor}{geom_point aesthetics}

\item{alpha}{geom_point aesthetics}
}
\value{
p PCA plot of the data colored by one variable
}
\description{
is used to plot PCA of the data colored by one variable
}
