% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCVariableRegression.R
\name{PCVariableRegression}
\alias{PCVariableRegression}
\title{is used to compute the linear regression.}
\usage{
PCVariableRegression(
  se.obj,
  assay.names = "All",
  variable,
  fast.pca = TRUE,
  nb.pcs = 10,
  save.se.obj = TRUE,
  plot.output = TRUE,
  assess.se.obj = TRUE,
  remove.na = "both",
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Optional string or list of strings for the selection of the name(s)
of the assay(s) of the SummarizedExperiment class object to compute the regression. By default}

\item{variable}{String of the label of a continuous variable such as
library size from colData(se.obj).}

\item{fast.pca}{Logical. Indicates whether to use the PCA calculated using a specific number of PCs instead of the
full range to speed up the process, by default is set to 'TRUE'.}

\item{nb.pcs}{Numeric. The number of few first cumulative PCs, by default is set to 10.}

\item{save.se.obj}{Logical. Indicates whether to save the result in the metadata of the SummarizedExperiment class
object 'se.obj' or to output the result. By default it is set to TRUE.}

\item{plot.output}{Logical. Indicates whether to plot the regression statistics, by default it is set to TRUE.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment class object.}

\item{remove.na}{Symbol. To remove NA or missing values from the assay(s) or variable or both. The options are
"assays", "sample.annotation, "both" or "none. "See the checkSeObj function for more details.}

\item{verbose}{Logical. Indicates whether to show or reduce the level of output or messages displayed during the
execution of the functions, by default it is set to TRUE.}
}
\value{
A SummarizedExperiment object containing the computed regression for
the continuous variable and if requested the associated plot.
}
\description{
This functions calculates the linear regression between the the first cumulative PCs of the gene expression (assays)
of a SummarizedExperiment object and a continuous variable (i.e. library size)
}
\details{
R2 values of fitted linear models are used to quantity the strength of the (linear) relationships between a single
quantitative source of unwanted variation, such as sample (log) library size or tumor purity, and global sample
summary statistics, such as the first k PCs (1 ≤ k ≤ 10).
}
\author{
Ramyar Molania
}
