% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combined_silh.R
\name{plot_combined_silh}
\alias{plot_combined_silh}
\title{is used to plot a combined plot of the silhouette computed on two caategorical variable
of a SummarizedExperiment class object (i.e. biology vs batch).}
\usage{
plot_combined_silh(silh1, silh2)
}
\arguments{
\item{silh1}{Vector of the Silhouette coefficient computed based on a categorical variable}

\item{silh2}{Vector of the Silhouette coefficient based on batch.}
}
\value{
Combined plot of the silhouette computed on batch versus the one
computed on biology
}
\description{
is used to plot a combined plot of the silhouette computed on two caategorical variable
of a SummarizedExperiment class object (i.e. biology vs batch).
}
