% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combined_silh_batch_bio.R
\name{plot_combined_silh_batch_bio}
\alias{plot_combined_silh_batch_bio}
\title{is used to plot a combined plot of the silhouette computed on batch versus the one
computed on biology of a SummarizedExperiment class object.}
\usage{
plot_combined_silh_batch_bio(silh_bio, silh_batch)
}
\arguments{
\item{silh_bio}{Vector of the Silhouette coefficient computed based on biology.}

\item{silh_batch}{Vector of the Silhouette coefficient based on batch.}
}
\value{
Combined plot of the silhouette computed on batch versus the one
computed on biology
}
\description{
is used to plot a combined plot of the silhouette computed on batch versus the one
computed on biology of a SummarizedExperiment class object.
}
