% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca}
\alias{plot_pca}
\title{is used to plot the pairwise plots of the first 3 PCA of the gene expression (assay)
of a SummarizedExperiment class object.}
\usage{
plot_pca(
  pca,
  assay_names = NULL,
  variable,
  variable.name,
  color,
  strokeSize = 0.2,
  pointSize = 0.8,
  strokeColor = "gray30",
  alpha = 0.5
)
}
\arguments{
\item{pca}{PCA components of a SummarizedExperiment variable that will be used in the plot.}

\item{assay_names}{Optional selection of names of the assays to compute the PCA.}

\item{variable}{The variable that will be used to display and color the PCA plot}

\item{variable.name}{The label of the variable that will be used on the PCA plot}

\item{color}{The color of the variable that will be used on the PCA plot}

\item{strokeSize}{geom_point aesthetics}

\item{pointSize}{geom_point aesthetics}

\item{strokeColor}{geom_point aesthetics}

\item{alpha}{geom_point aesthetics}
}
\value{
p PCA plot of the data colored by one variable
}
\description{
is used to plot the pairwise plots of the first 3 PCA of the gene expression (assay)
of a SummarizedExperiment class object.
}
