% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RLE.R
\name{plot_RLE}
\alias{plot_RLE}
\title{is used to plot the RLE of a SummarizedExperiment class object.}
\usage{
plot_RLE(se, assay_names = NULL, apply.log = FALSE)
}
\arguments{
\item{se}{A SummarizedExperiment object that will be used to compute the correlation}

\item{assay_names}{Optional string or list of strings for selection of the names
of the assays of the SummarizedExperiment class object to compute the correlation.}

\item{apply.log}{Indicates whether to apply a log-transformation to the data. By default
no transformation will be selected.}
}
\value{
list List of the computed RLE and the associated plot
}
\description{
is used to plot the RLE of a SummarizedExperiment class object.
}
