% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotARI.R
\name{plotARI}
\alias{plotARI}
\title{Generate barplot the adjusted rand index (ARI).}
\usage{
plotARI(
  se.obj,
  assay.names = "all",
  variables,
  ari.method = "hclust.complete.euclidian",
  plot.type = "single.plot",
  plot.output = TRUE,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Symbol. A symbol or list of symbols for the selection of the name(s) of the assay(s) in the
SummarizedExperiment object to generate barplot or scatter plots of the computed adjusted rand index. By default all
the assays of the SummarizedExperiment object will be selected.}

\item{variables}{Symbol. Indicates one or two column names in the SummarizedExperiment object that contains categorical
variables such as sample subtypes or batches. If two column names are provided, the function plots the two adjusted
rand index against each other for all the specified assays.}

\item{ari.method}{Symbol. A symbol that indicates what computed ARI method should be used for plotting. The "ari.method"
must be specified based on the "computeARI" function. The default is "hclust.complete.euclidian", which is the default
of the the "computeARI" function. We refer to the "computeARI" function for more detail.}

\item{plot.type}{Symbol. A symbol that specifies how to plot the adjusted rand index. The options are "single.plot" and
"combined.plot". If a variable is provided in the "variables" argument, then the "plot.type" must be set to "single.plot",
so, the function generates a barplot of the adjusted rand index. If two variables are provided and "plot.type" is set to
"combined.plot", then the function generates a scatter plot of the adjusted rand index of each variable against each other.
The default is set to 'single.plot'.}

\item{plot.output}{Logical. If 'TRUE', the individual barplots or scatter plots will be printed while functions is running.
The default is set to 'TRUE'.}

\item{save.se.obj}{Logical. Indicates whether to save the plots in the metadata of the SummarizedExperiment  object
or to output the result as list. The default is set to 'TRUE'.}

\item{verbose}{Logical. If 'TRUE', shows the messages of different steps of the function.}
}
\value{
A SummarizedExperiment object or a list that containing all the plots of the computed ARI on the categorical
variable.
}
\description{
This functions generates barplots of adjusted rand index for individual assays in the SummarizedExperiment object. If
two variables are provided, the function creates scatter plots of the adjusted rand index of each variable for
individual assays.
}
\references{
Molania R., ..., Speed, T. P., Removing unwanted variation from large-scale RNA sequencing data with PRPS,
Nature Biotechnology, 2023
}
\author{
Ramyar Molania
}
