% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAssessVariation.R
\name{plotAssessVariation}
\alias{plotAssessVariation}
\title{Plot the the variation of biological and unwanted variables.}
\usage{
plotAssessVariation(
  se.obj,
  assay.names = "all",
  variables,
  fast.pca = TRUE,
  anova.method = "aov",
  corr.method = "spearman",
  pcorr.method = "spearman",
  sil.dist.measure = "euclidian",
  ari.clustering.method = "hclust",
  ari.hclust.method = "complete",
  ari.hclust.dist.measure = "euclidian",
  output.file.name = NULL,
  pdf.width = 12,
  pdf.height = 12,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{#' @param se.obj A SummarizedExperiment object.}

\item{assay.names}{Symbol. A symbol or a vector of symbols for the selection of the name(s) of the assay(s) in the
SummarizedExperiment object. The default is set to 'all', so all the assays in the SummarizedExperiment object will
be selected.}

\item{variables}{Symbol. A symbol or vector of symbols indicating the name(s) of the column(s) in the SummarizedExperiment
object for which the specified metrics below have been calculated.}

\item{fast.pca}{Logical. Indicates whether to the fast SVD approach has been used in the 'computePCA' function or not.
The default is set to 'TRUE'.}

\item{anova.method}{Symbol. A symbol indicating which anova method has been used to compute association between
gene-level expression and a categorical variable. The options are 'aov' and 'welch'. The default is set to 'aov'.}

\item{corr.method}{Symbol. A symbol indicating which correlation method should be used to compute correlation between
gene-level expression and a continuous variable. The options are 'pearson', 'kendall', or "spearman". The default is
set to 'spearman'.}

\item{pcorr.method}{Symbol. A symbol indicating which correlation method should be used to compute gene-gene partial
correlation. The options are 'pearson', 'kendall', or "spearman". The default is set to 'spearman'.}

\item{sil.dist.measure}{Symbol. A symbol indicating which ditsance measure to be applied on the PCs to calculate silhouette.
The options are 'euclidean', maximum', manhattan', 'canberra', 'binary' or 'minkowski'. The default is set to 'euclidean'.
Refer to the function 'computeSilhouette' for more details.}

\item{ari.clustering.method}{Symbol. A symbol that indicates which clustering methods should be applied on the PCs to
calculate the ARI. The options are 'mclust' or 'hclust' methods. The default is set to 'hclust'. Refer to the 'computeARI'
for more details.}

\item{ari.hclust.method}{Symbol. A symbol specifying the agglomeration method to be used when the 'clustering.method'
is set to 'hclust' method for computing ARI. The options are: 'ward.D', 'ward.D2', 'single', 'complete', 'average' (= UPGMA),
'mcquitty' (= WPGMA), median' (= WPGMC) or centroid' (= UPGMC). The default is set to 'complete'. Refer to the 'computeARI'
for more details.}

\item{ari.hclust.dist.measure}{Symbol. A symbol indicating the distance measure to be used in the 'dist' function when
the 'clustering.method' is set to 'hclust' method for computing ARI. The options are 'euclidean', 'maximum', 'manhattan',
'canberra', 'binary' or 'minkowski'. The default is set to 'euclidean'. Refer to the 'computeARI' for more details.}

\item{output.file.name}{Path and name of the output file to save the assessments plots in a pdf format.}

\item{pdf.width}{TTT}

\item{pdf.height}{TTT}

\item{verbose}{Logical. If 'TRUE', shows the messages of different steps of the function.}
}
\description{
Plot the the variation of biological and unwanted variables.
}
