% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCombinedSilhouette.R
\name{plotCombinedSilhouette}
\alias{plotCombinedSilhouette}
\title{is used to plot a combined plot of the silhouette computed on two categorical variables
of a SummarizedExperiment class object (i.e. biology vs batch).}
\usage{
plotCombinedSilhouette(
  se.obj,
  assay.names = "All",
  variable1,
  variable2,
  method = "euclidian",
  save.se.obj = TRUE,
  assess.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object that will be used to compute the differential expression ANOVA analysis.}

\item{assay.names}{Optional string or list of strings for the selection of the name(s)
of the assay(s) of the SummarizedExperiment class object to compute the correlation. By default}

\item{variable1}{String of the label of a categorical variable such as
sample types or batches from colData(se.obj).}

\item{variable2}{String of the label of a categorical variable such as
sample types or batches from colData(se.obj).}

\item{method}{A character string indicating which method
is to be used for the differential analysis: 'euclidean', 'maximum', 'manhattan', 'canberra', 'binary' or 'minkowski'.
By default 'euclidean' will be selected.}

\item{save.se.obj}{Logical. Indicates whether to save the result in the metadata of the SummarizedExperiment class object 'se.obj' or
to output the result. By default it is set to TRUE.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment class object.}

\item{verbose}{Logical. Indicates whether to show or reduce the level of output or messages displayed during the execution
of the functions, by default it is set to TRUE.}
}
\value{
plot Combined plot of the silhouette coefficient on the two categorical variables.
}
\description{
is used to plot a combined plot of the silhouette computed on two categorical variables
of a SummarizedExperiment class object (i.e. biology vs batch).
}
