% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneSetScore.R
\name{plotGeneSetScore}
\alias{plotGeneSetScore}
\title{Plot gene set score of a gene set.}
\usage{
plotGeneSetScore(
  se.obj,
  assay.names = "all",
  reference.score = NULL,
  gene.set.name,
  plot.output = TRUE,
  assess.se.obj = TRUE,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{TTT}

\item{assay.names}{TTT}

\item{reference.score}{TTT}

\item{gene.set.name}{TTT}

\item{plot.output}{TTTT}

\item{assess.se.obj}{TTT}

\item{save.se.obj}{TTT}

\item{verbose}{TTT}

\item{name}{TTT}
}
\description{
This function computes all possible gene-gene pairwise ordinary and partial correlation of the assays in the
SummarizedExperiment object.
}
\details{
Partial correlation is used to estimate correlation between two variables while controlling for third
variables
}
\author{
Ramyar Molania
}
