% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGenesPartialCorrelation.R
\name{plotGenesPartialCorrelation}
\alias{plotGenesPartialCorrelation}
\title{Plot gene-gene partial correlation.}
\usage{
plotGenesPartialCorrelation(
  se.obj,
  assay.names = "all",
  variable = NULL,
  method = "spearman",
  plot.type = "barplot",
  filter.genes = TRUE,
  corr.dif.cutoff = 0.4,
  plot.ncol = 2,
  plot.nrow = 3,
  plot.output = TRUE,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Symbol. A symbol or a vector of symbols for the selection of the name(s) of the assay(s) in the
SummarizedExperiment object to obtain the computed correlations. The default is set to "all, which
indicates all the assays of the SummarizedExperiment object will be selected.}

\item{variable}{Symbol. A symbol that indicates the column name of the SummarizedExperiment object that has been used
to compute partial correlation.}

\item{method}{Symbol. A symbol that indicates which correlation methods has been computed for the "variable". The options
are 'pearson' or "spearman". The default is set to 'spearman'.}

\item{plot.type}{Symbol. A symbol specifying how to plot the results of the partial pairwise gene correlation. The
options are:}

\item{filter.genes}{Logical. Indicates to filter genes before generating the plots or not. The default is set to 'TRUE'.}

\item{corr.dif.cutoff}{Numeric. A numeric value to be used as a cutoff for filtering genes. The default is set to
0.4.}

\item{plot.ncol}{Numeric. Indicates number of columns in the plot grid. When the number of selected assay is more than
1, the function puts all the RLE boxplots in one grid.}

\item{plot.nrow}{Numeric. Indicates number of rows in the plot grid. When the number of selected assay is more than
3, the function puts all the RLE boxplots in one grid.}

\item{plot.output}{Logical. Indicates to print the histogram of partial correlation coefficients or not. The default is
set to 'FLASE'.}

\item{save.se.obj}{Logical. Indicates whether to save the RLE results in the metadata of the SummarizedExperiment object or
to output the result as list. By default it is set to TRUE.}

\item{verbose}{Logical. If 'TRUE', shows the messages of different steps of the function.}
}
\value{
The SummarizedExperiment object that contains the correlation plots or a list of the correlation plots
for individual assays.
}
\description{
This function computes all possible gene-gene pairwise ordinary and partial correlation of the assays in the
SummarizedExperiment object.
}
\details{
Partial correlation is used to estimate correlation between two variables while controlling for third
variables.
}
\author{
Ramyar Molania
}
