% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMetric.R
\name{plotMetric}
\alias{plotMetric}
\title{is used to plot the metric computed on a variable across the assays selected of a
SummarizedExperiment class object.}
\usage{
plotMetric(
  se.obj,
  assay.names = "All",
  variable,
  metric = c("gene.pearson.corr", "gene.spearman.corr", "gene.aov.anova",
    "gene.welch.correction.anova", "ari", "pcs.vect.corr", "pcs.lm", "sil.euclidean",
    "sil.maximum", "sil.manhattan", "sil.canberra", "sil.binary", "sil.minkowski"),
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object that will be used.}

\item{assay.names}{String or list of strings for the selection of the name
of the assays of the SummarizedExperiment class object.}

\item{variable}{String of the label of a variable such as
sample types or batches from colData(se.obj).}

\item{metric}{Metric being assessed.}

\item{verbose}{Logical. Indicates whether to show or reduce the level of output or messages displayed during the execution
of the functions, by default it is set to TRUE.}
}
\value{
SummarizedExperiment A SummarizedExperiment object containing the associated plot.
}
\description{
is used to plot the metric computed on a variable across the assays selected of a
SummarizedExperiment class object.
}
