% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCAAssay.R
\name{plotPCAassay}
\alias{plotPCAassay}
\title{is used to plot the pairwise plots of the first PCA of the gene expression
of a single assay of the SummarizedExperiment class object.}
\usage{
plotPCAassay(
  se.obj,
  pca_x,
  variable,
  nb.pcs = 3,
  fast.pca = TRUE,
  color = NULL,
  strokeSize = 0.2,
  pointSize = 1.5,
  strokeColor = "gray30",
  alpha = 0.5
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object that will be used to compute the differential expression ANOVA analysis.}

\item{pca_x}{The PCA of a SummarizedExperiment object that will be used to plot.}

\item{variable}{String of the label of a categorical variable such as
sample types or batches from colData(se.obj).}

\item{nb.pcs}{TO BE DEFINED.}

\item{fast.pca}{TO BE DEFINED.}

\item{color}{The color of the variable that will be used on the PCA plot}

\item{strokeSize}{geom_point aesthetics}

\item{pointSize}{geom_point aesthetics}

\item{strokeColor}{geom_point aesthetics}

\item{alpha}{geom_point aesthetics}
}
\value{
plot PCA plot of the data colored by one variable
}
\description{
is used to plot the pairwise plots of the first PCA of the gene expression
of a single assay of the SummarizedExperiment class object.
}
