% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCVariableCorrelation.R
\name{plotPCVariableCorrelation}
\alias{plotPCVariableCorrelation}
\title{plot the vector correlation.}
\usage{
plotPCVariableCorrelation(
  se.obj,
  assay.names = "all",
  variable,
  fast.pca = TRUE,
  nb.pcs = 10,
  plot.output = TRUE,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Symbol. A symbol or list of symbols for the selection of the name(s) of the assay(s) in the
SummarizedExperiment object to generate a vector correlation. The default is "all, which indicates all
the assays of the SummarizedExperiment object will be selected.}

\item{variable}{Symbol. Indicates a name of the column in the sample annotation of the SummarizedExperiment object.
The variable must be a categorical variable.}

\item{fast.pca}{Logical. Indicates whether to use the fast PCA or PCA results computed by the computePCA function. The
default is 'TRUE'.}

\item{nb.pcs}{Numeric. The number of first PCs to use to plot the vector correlation. The default is 10.}

\item{plot.output}{Logical. If TRUE, the individual vector correlation plot(s) will be printed while functions is running.}

\item{save.se.obj}{Logical. Indicates whether to save the vector correlation plots to the meta data of the
SummarizedExperiment object or to output the result as list. By default it is set to TRUE.}

\item{verbose}{Logical. If TRUE, displaying process messages is enabled.}
}
\value{
A SummarizedExperiment object or a list that contains all the vector correlation plots for the individual
assay(s).
}
\description{
This function generate a dot-line plot between the first cumulative PCs and the correlation coefficient (R2) obtained
from the vector correlation analysis. An ideal normalization should results a low correlation with unwanted variation
variables and high correlation with known biology.
}
\references{
Molania R., ..., Speed, T. P., Removing unwanted variation from large-scale RNA sequencing data with PRPS,
Nature Biotechnology, 2023
}
\author{
Ramyar Molania
}
