% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareSeObj.R
\name{prepareSeObj}
\alias{prepareSeObj}
\title{Prepare a SummarizedExperiment object.}
\usage{
prepareSeObj(
  data = NULL,
  sample.annotation = NULL,
  raw.count.assay.name = NULL,
  remove.lowly.expressed.genes = FALSE,
  count.cutoff = 10,
  biological.group = NULL,
  minimum.proportion = 0.5,
  calculate.library.size = FALSE,
  estimate.tumor.purity = NULL,
  assay.name.to.estimate.purity = NULL,
  create.sample.annotation = FALSE,
  gene.annotation = NULL,
  create.gene.annotation = FALSE,
  add.gene.details = FALSE,
  gene.group = NULL,
  gene.details = NULL,
  add.housekeeping.genes = FALSE,
  add.immun.stroma.genes = FALSE,
  metaData = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data}{List. A list containing data sets (assay(s)) or expression data. For individual datasets, genes should be arranged
in rows and samples in columns. If multiple datasets are provided, ensure that the row names of the assays are in
the same order.}

\item{sample.annotation}{Data frame. Contains sample-specific information for the assays. The order of rows should
match the column names of the assay(s).}

\item{raw.count.assay.name}{Character. The name of the raw RNA-seq count data within the list of assays or expression data.
Raw counts data is required if 'remove.lowly.expressed.genes' or 'calculate.library.size' is set to 'TRUE'.}

\item{remove.lowly.expressed.genes}{Logical. If 'TRUE', lowly expressed genes will be identified and removed from
the assay specified in the 'raw.count.assay.name' argument. The default is 'FALSE'.}

\item{count.cutoff}{Numeric. Minimum count required for at least some sample groups. If the 'biological.group' argument
is set to 'NULL', all samples will be considered as a single group. Otherwise, the smallest subgroups of the
'biological.group' will be considered for filtering lowly expressed genes. This follows the 'filterByExpr' function
from the edgeR package.}

\item{biological.group}{Character. The column name in the sample annotation specifying the biological groups.
The smallest biological groups will be considered for filtering lowly expressed genes. If 'NULL', all samples will
be considered as a single group.}

\item{minimum.proportion}{Numeric. The minimum proportion of samples within a group in which a gene must be expressed.
The default is 0.5.}

\item{calculate.library.size}{Logical. If 'TRUE', the library size (total counts) will be calculated using the data
provided in 'raw.count.assay.name'. Library size will be computed after removing lowly expressed genes.}

\item{estimate.tumor.purity}{Character. A method to estimate tumor purity. Options include: 'estimate', 'singscore',
'both', and 'NULL'. If 'estimate', the ESTIMATE method will be used. If 'singscore', the 'singscore' method will be
applied. If 'both', both methods will be applied. The default is 'NULL'.}

\item{assay.name.to.estimate.purity}{Character. The name of an assay within the list of assays or expression data
to be used for estimating tumor purity.}

\item{create.sample.annotation}{Logical. If 'TRUE', a sample annotation will be generated, initially containing
the column names of the assay(s). Additional variables, such as library size, will be included in the annotation.}

\item{gene.annotation}{Data frame. Contains gene-specific details (e.g., chromosome names, GC content). The list of
housekeeping genes and immune/stromal genes will be included in the gene annotation.}

\item{create.gene.annotation}{Logical. If 'TRUE', a gene annotation will be generated, initially containing row names
of the assay(s). Additional gene details and lists will be added to the annotation.}

\item{add.gene.details}{Logical. If 'TRUE', preset or provided gene details in the 'gene.details' argument will be
added to the gene annotation.}

\item{gene.group}{Character. The name of a gene class from the row names of the assay(s). This must be one of the following:
'entrezgene_id', 'hgnc_symbol', or 'ensembl_gene_id'.}

\item{gene.details}{Character or character vector. Indicates the gene details to be included in the gene annotation.}

\item{add.housekeeping.genes}{Logical. If 'TRUE', publicly available "housekeeping" gene sets will be added to the gene
annotation. These genes may be used as negative controls for RUV normalization.}

\item{add.immun.stroma.genes}{Logical. If 'TRUE', immune and stromal gene signatures from Kosuke Yoshihara et al. will
be added to the gene annotation. These signatures can be used to estimate tumor purity in cancer RNA-seq data.}

\item{metaData}{Any. Metadata in any format and dimensions.}

\item{verbose}{Logical. If 'TRUE', the function will print messages at various steps of the process.}
}
\value{
A SummarizedExperiment object containing assay(s) and also samples annotation, gene annotation and metadata
if there are specified.
}
\description{
This function creates a SummarizedExperiment object from tabular expression data set(s) and sample annotation (if
available). The function can identify and remove lowly expressed genes if a raw count data is provided, add a range
of annotations for individual genes e.g. biotype, chromosome names, ... . , estimate tumor purity, and provide several
sets of housekeeping genes and a immune and stromal gene signatures. The housekeeping gene sets could be suitable
negative control genes for identification of sources of unknown unwanted variation and the implementation of the RUV
methods. The immune and stromal gene signature could be used to estimate tumor purity variation in cancer RNA-seq data.
}
\details{
The SummarizedExperiment object is a data structure used in the R for representing and manipulating high-dimensional
experimental data. Here are some key features and components of the SummarizedExperiment object:
Assays:
The SummarizedExperiment allows for the incorporation of multiple assays (data). Each assay is a separate matrix of data
associated with the same features and samples. Rows typically represent features (e.g., genes, transcripts), and columns
represent samples or experimental conditions.
Row and Column Metadata:
The SummarizedExperiment object includes metadata associated with both rows and columns. Row metadata can contain
information about the features, such as gene annotations or genomic coordinates. Column metadata may include sample
information, experimental conditions, or other relevant details.
Metadata
The Metadata of the SummarizedExperiment allows for flexibility in terms of data types and structures. This makes it
suitable for saving metrics and plots in the RUVIIIPRPS R package. We refer to the SummarizedExperiment R package for
more details.
}
\author{
Ramyar Molania
}
