% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printColoredMessage.R
\name{printColoredMessage}
\alias{printColoredMessage}
\title{is used to print colored messages.}
\usage{
printColoredMessage(message, color, verbose)
}
\arguments{
\item{message}{Message to be printed.}

\item{color}{The color of the message.}

\item{verbose}{whether to show the messages or not.}
}
\description{
is used to print colored messages.
}
