% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prpsForContinuousUV.R
\name{prpsForContinuousUV}
\alias{prpsForContinuousUV}
\title{is used to create PRPS for a continuous variable as source of unwanted variation.}
\usage{
prpsForContinuousUV(
  se.obj,
  assay.name,
  apply.log = TRUE,
  pseudo.count = 1,
  uv.variable,
  bio.variable,
  batch.variable = NULL,
  min.sample.for.prps = 3,
  min.sample.per.batch = 10,
  assess.se.obj = TRUE,
  remove.na = "both",
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object that will be used to create PRPS.}

\item{assay.name}{String for the selection of the name of the assay of the SummarizedExperiment class object.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data, by default it is set to TRUE.}

\item{pseudo.count}{Numeric. A value as a pseudo count to be added to all measurements before log transformation,
by default it is set to 1.}

\item{uv.variable}{String of the label of a continuous or continuous variable.
such as samples types or batch from colData(se) that will be used to define PRPS.}

\item{bio.variable}{String of the label of a categorical variable that specifies major biological groups
such as samples types from colData(se).}

\item{batch.variable}{String of the label of a categorical variable that specifies major batch groups
such as plates from colData(se).}

\item{min.sample.for.prps}{Numeric. Indicates the minimum number of samples to create one pseudo-sample,
by default it is set to 3.}

\item{min.sample.per.batch}{Numeric. Indicates the minimum number of homogeneous biological samples within each batch
to create a PRPS set for each continuous variable. The minimum should be '2*min.sample.for.prps'. By default it is set to 6.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment class object.
By default it is set to TRUE.}

\item{remove.na}{String. Indicates whether to remove NA or missing values from either the 'assays', the 'sample.annotation',
'both' or 'none'. If 'assays' is selected, the genes that contains NA or missing values will be excluded. If 'sample.annotation' is selected, the
samples that contains NA or missing values for any 'bio.variables' and 'uv.variables' will be excluded. By default, it is set to
'both'.}

\item{save.se.obj}{Logical. Indicates whether to save the result in the metadata of the SummarizedExperiment class
object 'se.obj' or to output the result, by default it is set to TRUE.}

\item{verbose}{Logical. Indicates whether to show or reduce the level of output or
messages displayed during the execution of the functions, by default it is set to TRUE.}
}
\value{
SummarizedExperiment A SummarizedExperiment object containing the PRPS data or just PRPS data.
}
\description{
We will create distinct group of pseudo-replicates for each source of unwanted variation defined in the 'uv.variables' argument.
For example to correct for library size if defined in the 'uv.variables' argument, several group of pseudo-samples
will be created by averaging the top and bottom-ranked samples by library size of the same biological subtype in each batch.
Then those pseudo-samples will be defined as pseudo-replicates which constitutes a PRPS set.
}
