% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_pc_contvar.R
\name{regression_pc_contvar}
\alias{regression_pc_contvar}
\title{is used to compute the linear regression of a continuous variable and the first cumulative PCs
of the gene expression (assay) of a SummarizedExperiment class object.}
\usage{
regression_pc_contvar(
  pca,
  cont_var,
  cont_var_label,
  assay_names = NULL,
  nb_pca_comp = 10
)
}
\arguments{
\item{pca}{PCA components of a SummarizedExperiment variable that will be used in the plot.}

\item{cont_var}{Vector of a categorical variable such as sample types
(i.e. biological subtypes) or batches.}

\item{cont_var_label}{String or vector of strings of the label of categorical variable(s) such as
sample types or batches from colData(se).}

\item{assay_names}{Optional selection of names of the assays to compute the PCA.}

\item{nb_pca_comp}{The number of components of the PCA used to compute the regression.}
}
\value{
list List containing the association plot and the computed regression
}
\description{
is used to compute the linear regression of a continuous variable and the first cumulative PCs
of the gene expression (assay) of a SummarizedExperiment class object.
}
