% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_pc_contvar_single_assay.R
\name{regression_pc_contvar_single_assay}
\alias{regression_pc_contvar_single_assay}
\title{is used to compute the linear regression of a continuous variable and the first cumulative PCs
of the data on a single assay}
\usage{
regression_pc_contvar_single_assay(pca, regression_var, nb_pca_comp = 10)
}
\arguments{
\item{pca}{PCs of the dataset that will be used in the plot}

\item{regression_var}{The continous variable that will be computed to the PCA of the data
(i.e. library size)}

\item{nb_pca_comp}{The number of components of the PCA used to compute the regression}
}
\value{
vector containing the computed regression
}
\description{
is used to compute the linear regression of a continuous variable and the first cumulative PCs
of the data on a single assay
}
