% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruv_multiple_k.R
\name{ruv_multiple_k}
\alias{ruv_multiple_k}
\title{is used to normalise the gene expression (assay) of a SummarizedExperiment class object
using RUVIII-PRPS method for multiple k values in one go.}
\usage{
ruv_multiple_k(Y, Yrep, M, ctl, k = c(1:20), eta = NULL)
}
\arguments{
\item{Y}{A m by n matrix of the Raw gene expression where m is the number of samples and n is the
number of features of a SummarizedExperiment variable to be normalised.}

\item{Yrep}{A matrix of the gene expression of the obtained PRPS}

\item{M}{Replicate matrix.}

\item{ctl}{Logical vector of length n of the negative control genes.}

\item{k}{A vector containing the range of k - the number of unwanted factors - to be tested.}

\item{eta}{A matrix with n columns, gene-wise (as opposed to sample-wise) covariates. By default is set to NULL.}
}
\value{
list List containing the corrected gene expression matrix and the M, alpha and W for each k values tested.
}
\description{
is used to normalise the gene expression (assay) of a SummarizedExperiment class object
using RUVIII-PRPS method for multiple k values in one go.
}
