% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUVIII.PRPS.R
\name{RUVIII.PRPS}
\alias{RUVIII.PRPS}
\title{RUVIII with technical replicates or PRPS normalization.}
\usage{
RUVIII.PRPS(
  se.obj,
  assay.name,
  prps.type = "supervised",
  prps.group,
  prps.set.names = NULL,
  ncg.type = "supervised",
  ncg.group,
  ncg.set.names = NULL,
  technical.replicates = NULL,
  k = 1,
  apply.log = TRUE,
  data.to.log = "assay",
  pseudo.count = 1,
  eta = NULL,
  include.intercept = TRUE,
  return.info = FALSE,
  assess.se.obj = TRUE,
  remove.na = "none",
  output.name = NULL,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.name}{Symbol. A symbol indicating the name of the assay in the SummarizedExperiment object. The selected
assay should be the one that has been used for PRPS and NCGs selection.}

\item{prps.type}{Symbol. A symbol denoting which groups of the generated PRPS should be utilized for normalization. The
options include 'supervised,' 'unsupervised,' or 'both.' When 'supervised' is chosen, the function employs the PRPS
set(s) generated through supervised functions. Opting for 'unsupervised' utilizes the PRPS set(s) created through
unsupervised functions. If 'both' is selected, the function utilizes the PRPS set(s) generated by both supervised and
unsupervised functions. The default setting is 'supervised.'}

\item{prps.group}{Symbol. A symbol indicating the name of PRPS group under the 'prps.type' slot.}

\item{prps.set.names}{A symbol specifying th exact name of a PPRS set under the 'prps.group' name in the supervised or
un.supervised slots in the SummarizedExperiment object. The default is set to 'all'. This will select all available
PRPS sets for the specified 'prps.group'.}

\item{ncg.type}{Symbol. A symbol indicating which groups of the computed NCGs should be utilized for normalization. The
options include 'supervised,' 'unsupervised,' or 'both.' When 'supervised' is chosen, the function employs the PRPS
set(s) generated through supervised functions. Opting for 'unsupervised' utilizes the PRPS set(s) created through
unsupervised functions. If 'both' is selected, the function utilizes the PRPS set(s) generated by both supervised and
unsupervised functions. The default setting is 'supervised.'}

\item{ncg.group}{Symbol. A symbol indicating the name of NCGs group under the 'prps.type' slot.}

\item{ncg.set.names}{A symbol specifying th exact name of a NCG set under the 'ncg.group' name in the 'supervised' or
'un.supervised' slots in the SummarizedExperiment object. The default is set to 'all'. This will select all available
NCG sets for the specified 'ncg.group'.}

\item{technical.replicates}{Vector. A vector of sample names that contain at least a technical duplicate. The length of
vector must as similar to the number of sample in the SummarizedExperiment object. The sample names of each set of
technical replicate must be the same. See the details for more information.}

\item{k}{Numeric. A numeric value or a vector of numeric values indicating the number of unwanted factors to be calculated
and used to remove unwanted variation from the data. If a vector is provided, the function applies the RUV-III method
for each individual value, up to a maximum of possible k values for each normalization. See the details for more information.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data or PRPS or both. The default it
is set to 'TRUE'.}

\item{data.to.log}{Symbol. A symbol specifying the data that must be logged transformed before applying the RUV-III
method.  The options are 'assay', 'prps' , 'both' or 'non'. Remember, both data and PRPS must be in log scale before
applying the RUV-III data.}

\item{pseudo.count}{Numeric. A numeric value as a pseudo count to be added to all measurements before log transformation.
The default it is set to 1. Note, the current RUV-III method does not support NA or missing values.}

\item{eta}{Symbol. Gene-wise (as opposed to sample-wise) covariates. These covariates are adjusted for by RUV-1 before any
further analysis proceeds. Can be either (1) a matrix with n columns, (2) a matrix with n rows, (3) a dataframe with
n rows, (4) a vector or factor of length n, or (5) simply 1, for an intercept term.}

\item{include.intercept}{When eta is specified (not NULL) but does not already include an intercept term, this will
automatically include one.}

\item{return.info}{If 'TRUE', additional information including the W matrix of the RUV-III method is returned. The
default is set to 'TRUE'.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment object or not. The default is set
to 'TRUE'. Refer to the checkSeObj function for more details.}

\item{remove.na}{Symbol. A symbol indicating to remove NA or missing values from the assays or not. The options are
'assays' and 'none'. The default is "assays", so all the NA or missing values from the assay(s) will be removed before
applying the RUV-III method. See the 'checkSeObj' function for more details.}

\item{output.name}{Symbol. A Symbol to use for the name(s) of the RUV-III normalized datasets. The default is set to
'NULL', them individual RUV-III normalized data will be named as follow
paste0('RUVIIIPRPS_K.', k.vals)}

\item{save.se.obj}{Logical. Indicates whether to save the RUV-III-PRPS normalized data as new assay(s) in the
SummarizedExperiment object or to output them as list. The default it is set to 'TRUE'.}

\item{verbose}{Logical. If 'TRUE', shows the messages of different steps of the function.}
}
\value{
Either the SummarizedExperiment object containing the RUV-III normalized gene expression data as new assay(s)
or a list of the RUV-III normalized gene expression data set(s).
}
\description{
This function applies the RUV-III with technical replicates or PRPS on transcriptions data.
}
\details{
RUV-III normalization. Before we can describe the linear model underlying RUV-III, we need to introduce the m × m1
mapping matrix M connecting assays to distinct samples, which captures the pattern of replication in our assays. Here,
m is the number of assays, and m1 is the number of distinct samples being assayed. M(i,h) = 1if assay i is on sample
h and M(i,h) = 0 otherwise. Each row of M sums to 1, and the columns sum to the distinct sample replication numbers, the
elements of MTM. We also define an m1 × p design matrix X to capture the biological factor(s) of interest indexed by
sample rather than assay. There are no constraints on p; indeed, X could be the m1 × m1 identity matrix. Our goal
here is to remove unwanted variation, not to estimate regression parameters.
The linear model we use is: Y = 1μ + MXβ + Wα + ε
where the data Y = (yij)and unobserved errors ε = (εij)are m × n; the matrices X and M have just been defined; μ is
the 1 × n row of gene means; β is p × n; the matrix W whose columns capture the unwanted variation is m × k; and α
is k × n. 1 = 1m is the m × 1 column vector of 1s. Here, m = number of assays, n = number of genes and p is the
dimension of the wanted variation X and k that of the unwanted variation W. Assume that W⊥1. Also, we suppose that we
have a subset of nc of negative control genes whose m × nc submatrix Yc satisfies Yc = 1μc + Wαc + εc, where we have
assumed that βc = 0—that is, that there is no true association between these genes and the biology of interest.
The projection PM = M(M M) M replaces each entry yij of Y by the simple average of the entries yi′j over all i′ for
which M (i, h) = M (i′ , h) = 1—that is, over all i′ such that i′ and i label replicate assays of the same unique
sample (or pseudo-sample) labeled h. Write RM = I − PM for the corresponding residual projector. This is our source
of information on the unwanted variation that we will remove. If the replication is technical at some level, then RMY
mainly contains information about unwanted variation in the system after the technical replicates were created.
Depending on the study details, technical replicates could be created immediately beforethe assay was run, in parallel
with or immediately after sample was collected or somewhere in between. The earlier the creation of technical replicates,
the more unwanted variation will be captured in their differences. The use of pseudo-replicates of suitable pseudo-samples
enables us to start to deal with pre-technical unwanted variation. Write the spectral decomposition of RMYY RM = UDU ,
where U is an m × m orthogonal matrix and D is an m × n diagonal matrix with entries ordered from largest to smallest
eigenvalue. Let P1 be the orthogonal projection onto 1m. For a chosen k, 1 ≤ k ≤ m−m1,
I- Define αˆ(k) = U(k)TY, where U(k) is the first k columns of U
II- Estimate W by regressing the centered negative controls (I − P1)Yc on αˆ(k)T c
III-
}
\references{
\itemize{
\item Molania R., ..., Speed, T. P., A new normalization for Nanostring nCounter gene expression data, Nucleic Acids Research,
}
\enumerate{
\item 
}
\itemize{
\item Molania R., ..., Speed, T. P., Removing unwanted variation from large-scale RNA sequencing data with PRPS,
Nature Biotechnology, 2023
}
}
\author{
Ramyar Molania
}
