% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruvIII.R
\name{ruvIII}
\alias{ruvIII}
\title{is used to normalise the gene expression (assay) of a SummarizedExperiment class object
using the RUVIII-PRPS method.}
\usage{
ruvIII(
  se.obj,
  assay.name,
  replicate.data,
  M,
  ctl,
  k = NULL,
  eta = NULL,
  include.intercept = TRUE,
  average = FALSE,
  fullalpha = NULL,
  return.info = FALSE,
  inputcheck = TRUE,
  assess.se.obj = TRUE,
  remove.na = "measurements",
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object that will be used to computer fastRUV-III}

\item{assay.name}{String for the selection of the name of the assay data
of the SummarizedExperiment class object}

\item{replicate.data}{TO BE DEFINED}

\item{M}{Replicate matrix.}

\item{ctl}{Logical vector of length n of the negative control genes.}

\item{k}{The number of unwanted factors to use.}

\item{eta}{A matrix with n columns, gene-wise (as opposed to sample-wise) covariates. By default is set to NULL.}

\item{include.intercept}{Logical. Add an intercept term to eta if it does not include one already. By default is set to TRUE.}

\item{average}{Logical. Average replicates after adjustment. By default is set to FALSE.}

\item{fullalpha}{Can be included to speed up execution. By default is set to NULL.}

\item{return.info}{Logical. If FALSE, only the adjusted data matrix is returned. If TRUE, additional information
is returned. By default is set to FALSE.}

\item{inputcheck}{Logical. Perform a basic sanity check on the inputs, and issue a warning if there is a problem.
By default is set to TRUE.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment class object.}

\item{remove.na}{TO BE DEFINED.}

\item{save.se.obj}{Logical. Indicates whether to save the result in the metadata of the SummarizedExperiment class object 'se.obj' or
to output the result. By default it is set to TRUE.}

\item{verbose}{Logical. Indicates whether to show or reduce the level of output or messages displayed during the execution
of the functions, by default it is set to TRUE.}
}
\value{
SummarizedExperiment A SummarizedExperiment object containing the normalised gene expression
into a new assay (RUVIII_K).
}
\description{
is used to normalise the gene expression (assay) of a SummarizedExperiment class object
using the RUVIII-PRPS method.
}
