% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette_coef_catvar.R
\name{silhouette_coef_catvar}
\alias{silhouette_coef_catvar}
\title{is used to compute the Silhouette coefficient from the PC of all assays
given a categorical variable}
\usage{
silhouette_coef_catvar(pca, assay_names = NULL, cat_var, nPCs = 3)
}
\arguments{
\item{pca}{PCs of the dataset that will be used}

\item{assay_names}{Optional selection of names of the assays to compute the PCA}

\item{cat_var}{is a categorical variable such as sample types or batches}

\item{nPCs}{is the number of PCs used to measure the distance}
}
\value{
list List containing the association plot and the computed silhouette
}
\description{
is used to compute the Silhouette coefficient from the PC of all assays
given a categorical variable
}
