% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette_coef_catvar_all_assays.R
\name{silhouette_coef_catvar_all_assays}
\alias{silhouette_coef_catvar_all_assays}
\title{is used to compute the Silhouette coefficient from the PC of all assays
given a categorical variable}
\usage{
silhouette_coef_catvar_all_assays(pca, normalization, cat_var, nPCs)
}
\arguments{
\item{pca}{PCs of the dataset that will be used}

\item{normalization}{All the available assays for the data (i.e. normalizations methods)}

\item{cat_var}{is a categorical variable such as sample types or batches}

\item{nPCs}{is the number of PCs used to measure the distance}
}
\value{
list List containing the association plot and the computed silhouette
}
\description{
is used to compute the Silhouette coefficient from the PC of all assays
given a categorical variable
}
