% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette_coef_catvar_single_assay.R
\name{silhouette_coef_catvar_single_assay}
\alias{silhouette_coef_catvar_single_assay}
\title{is used to compute the Silhouette coefficient from the PC of a single assay
given a categorical variable}
\usage{
silhouette_coef_catvar_single_assay(pca, cat_var, nPCs = 3)
}
\arguments{
\item{pca}{PCs of the dataset that will be used}

\item{cat_var}{is a categorical variable such as sample types or batches}

\item{nPCs}{is the number of PCs used to measure the distance}
}
\value{
a single value, the silhouette coefficient computed
}
\description{
is used to compute the Silhouette coefficient from the PC of a single assay
given a categorical variable
}
