% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette_coeff.R
\name{silhouette_coeff}
\alias{silhouette_coeff}
\title{is used to compute the Silhouette coefficient given a categorical variable}
\usage{
silhouette_coeff(pca, variable, nPCs)
}
\arguments{
\item{pca}{PCs of the dataset that will be used}

\item{variable}{is a categorical variable such as sample types or batches}

\item{nPCs}{is the number of PCs used to measure the distance}
}
\value{
a single value, the silhouette coefficient computed
}
\description{
is used to compute the Silhouette coefficient given a categorical variable
}
