% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervisedFindNCG.R
\name{supervisedFindNCG}
\alias{supervisedFindNCG}
\title{is used to find a set of negative control genes  (NCG) of an assay in a SummarizedExperiment object.}
\usage{
supervisedFindNCG(
  se.obj,
  assay.name,
  approach = "PbPbio",
  nb.ncg = 10,
  top.rank.bio.genes = 50,
  top.rank.uv.genes = 50,
  bio.variables,
  bio.groups = NULL,
  nb.bio.clusters = 2,
  uv.variables,
  uv.groups = NULL,
  nb.uv.clusters = 2,
  bio.clustering.method = "kmeans",
  uv.clustering.method = "kmeans",
  normalization = "CPM",
  regress.out.uv.variables = NULL,
  regress.out.bio.variables = NULL,
  apply.log = TRUE,
  pseudo.count = 1,
  min.sample.for.aov = 3,
  min.sample.for.correlation = 10,
  corr.method = "spearman",
  a = 0.05,
  rho = 0,
  grid.nb = 10,
  anova.method = "aov",
  ncg.selection.method = "Prod",
  assess.ncg = TRUE,
  variables.to.assess.ncg = NULL,
  nb.pcs = 5,
  center = TRUE,
  scale = FALSE,
  assess.se.obj = TRUE,
  assess.variables = TRUE,
  cat.cor.coef = c(0.95, 0.95),
  cont.cor.coef = c(0.95, 0.95),
  remove.na = "both",
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.name}{A name of an assay data in the SummarizedExperiment object.}

\item{approach}{TTTT}

\item{nb.ncg}{TTTTT}

\item{top.rank.bio.genes}{TTTTT}

\item{top.rank.uv.genes}{TTTTT}

\item{bio.variables}{TTTT}

\item{bio.groups}{TTTT}

\item{nb.bio.clusters}{TTTTT}

\item{uv.variables}{YYYY}

\item{uv.groups}{YYYY}

\item{nb.uv.clusters}{YYYYY}

\item{bio.clustering.method}{YYYY}

\item{uv.clustering.method}{UUUUU}

\item{normalization}{HHHH}

\item{regress.out.uv.variables}{UUUUU}

\item{regress.out.bio.variables}{UUUUUU}

\item{apply.log}{UUUUU}

\item{pseudo.count}{YYYYY}

\item{min.sample.for.aov}{UUUUU}

\item{min.sample.for.correlation}{UUUUU}

\item{corr.method}{Logical. Indicates whether to apply a log-transformation to the data.}

\item{a}{Logical. Indicates whether to apply a log-transformation to the data.}

\item{rho}{Logical. Indicates whether to apply a log-transformation to the data.}

\item{grid.nb}{NNNNNN}

\item{anova.method}{PPPPP}

\item{ncg.selection.method}{PPPP}

\item{assess.ncg}{PPPP}

\item{variables.to.assess.ncg}{PPPP}

\item{nb.pcs}{PPPPPP}

\item{center}{PP}

\item{scale}{PP}

\item{assess.se.obj}{Logical. Indicates whether to apply a log-transformation to the data.}

\item{assess.variables}{PPPP}

\item{cat.cor.coef}{PPPP}

\item{cont.cor.coef}{PPPP}

\item{remove.na}{PPPPP}

\item{save.se.obj}{PPPP}

\item{verbose}{Logical. Indicates whether to apply a log-transformation to the data.}
}
\value{
a list of negative control genes.
}
\description{
is used to find a set of negative control genes  (NCG) of an assay in a SummarizedExperiment object.
}
