% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervisedFindNcgAnoCorrAs.R
\name{supervisedFindNcgAnoCorrAs}
\alias{supervisedFindNcgAnoCorrAs}
\title{is used to find a set of negative control genes (NCG) using ANOVA and correlation analyses across all samples.}
\usage{
supervisedFindNcgAnoCorrAs(
  se.obj,
  assay.name,
  nb.ncg = 10,
  bio.variables,
  uv.variables,
  ncg.selection.method = "AbsNoneOverlap",
  grid.nb = 0.05,
  top.rank.bio.genes = 80,
  top.rank.uv.genes = 80,
  min.sample.for.aov = 3,
  min.sample.for.correlation = 10,
  regress.out.uv.variables = NULL,
  regress.out.bio.variables = NULL,
  normalization = "CPM",
  corr.method = "spearman",
  a = 0.05,
  rho = 0,
  apply.log = TRUE,
  pseudo.count = 1,
  anova.method = "aov",
  remove.na = "both",
  assess.ncg = TRUE,
  variables.to.assess.ncg = NULL,
  nb.pcs = 5,
  scale = FALSE,
  center = TRUE,
  assess.se.obj = TRUE,
  assess.variables = TRUE,
  cat.cor.coef = c(0.95, 0.95),
  cont.cor.coef = c(0.95, 0.95),
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.name}{Symbol.Indicates a name of assay in the SummarizedExperiment object.
This assay will be used to find a set set of negative control genes.}

\item{nb.ncg}{Numeric.Indicates the percentage of the total genes to be selected a NCG set,
by default it is set to 10 percent.}

\item{bio.variables}{Symbole. Indicates the columns names that contain bioloical variables in the SummarizedExperiment object.}

\item{uv.variables}{Symbole. Indicates the columns names that contain unwanted variation variables in the SummarizedExperiment object.}

\item{ncg.selection.method}{Symbol.Indicates how to select a set genes as NCG.
For individual genes, the two-way ANOVA calculates F-statistics for biological and
unwanted variation factors separately. An ideal NCG set should have high F-statistics
for the unwanted variation variables and low F-statistics for the biological variables.
The function ranks the F-statistics obtained for the biological variable and negative of
the F-statistics obtained for the unwanted variables. Then this functions offers 5 ways to
summarize the ranks of the two F-statistics. Prod' is the product of the ranks. 'Sum', is
the sum of the ranks. 'Average' is the average of the ranks. 'AbsNoneOverlap' is the none overlapped
genes of the 'top.rank.uv.genes' and 'top.rank.bio.genes'. 'noneOverlap' is the none overlapped genes
of the 'top.rank.uv.genes' and at least 'top.rank.bio.genes'. The F-statistics for biological and UV
are first ranked.Then options are Prod (product), Sum, Average,}

\item{grid.nb}{Numeric. Indicates the percentage for grid search when the ncg.selection.method is
'noneOverlap'. In the 'noneOverlap' approach, the grid search starts with the initial
'top.rank.uv.genes' value an add the grid.nb in each loop to find the 'nb.ncg'.}

\item{top.rank.bio.genes}{Numeric.Indicates the percentage of top ranked genes that are highly affected
by the biological variation. This is required to be specified when the 'ncg.selection.method' is
either 'noneOverlap' or 'AbsNoneOverlap'.}

\item{top.rank.uv.genes}{Numeric.Indicates the percentage of top ranked genes that are highly affected
by the unwanted variation variables. This is required to be specified when the 'ncg.selection.method' is
either 'noneOverlap' or 'AbsNoneOverlap'.}

\item{min.sample.for.aov}{Numeric.Indicates the minimum number of samples to perform correlation analyses between continuous sources
of variation (biological and unwanted variation) with individual gene expression.}

\item{min.sample.for.correlation}{Numeric.Indicates the minimum number of samples to perform correlation analyses between continuous sources
of variation (biological and unwanted variation) with individual gene expression.}

\item{regress.out.uv.variables}{Symbol.Indicates the column names in the SummarizedExperiment object that will be regressed out from the data
before performing correlation and ANOVA. The default is NULL.}

\item{regress.out.bio.variables}{Symbol.Indicates the column names in the SummarizedExperiment object that will be regressed out from the data
before performing correlation and ANOVA. The default is NULL.}

\item{normalization}{Symbols.Indicates winch normalization method to use for library size adjustment before fining genes that are highly
affected by biological variation. The default is CPM. Refer to the 'applyOtherNormalization' function for further details.}

\item{corr.method}{Numeric.Indicates which correlation methods (pearson or spearman) should be used for the correlation analyses. The
default is 'spearman'.}

\item{a}{TBBB}

\item{rho}{TBBB}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data, by default it is set to TRUE.}

\item{pseudo.count}{Numeric. A value as a pseudo count to be added to all measurements before log transformation.}

\item{anova.method}{Symbols. Indicates which ANOVA methods to use. The default in 'aov'. Refer to the function "genesVariableAnova" for more details.}

\item{remove.na}{Symbol. Indicates whether to remove NA or missing values from either the 'assays', the 'sample.annotation',
'both' or 'none'. If 'assays' is selected, the genes that contains NA or missing values will be excluded. If 'sample.annotation' is selected, the
samples that contains NA or missing values for any 'bio.variables' and 'uv.variables' will be excluded. By default, it is set to both'.}

\item{assess.ncg}{Logical. Indicates whether to assess the performance of selected NCG or not.
This analysis involves principal component analysis on the selected NCG and
then explore the R^2 or vector correlation between the 'nb.pcs' first principal
components and with biological and unwanted variables.}

\item{variables.to.assess.ncg}{Symbols. Indicates the column names of the SummarizedExperiment object that
contain variables whose association with the selected genes as NCG.
needs to be evaluated. The default is NULL. This means all the variables in the 'bio.variables' and 'uv.variables' will be assessed.}

\item{nb.pcs}{Numeric. Indicates the number of the first principal components on selected NCG to be used to assess the performance of NCGs.}

\item{scale}{Logical. Indicates whether to scale the data before applying Principal component analysis. The default is FALSE.}

\item{center}{Logical. Indicates whether to center the data before applying principal component analysis or not. The default is TRUE.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment class object or not.}

\item{assess.variables}{Logical. Indicates whether to assess the association between the bioloical and unwanted variation variables seprartly.}

\item{cat.cor.coef}{Vector of two numerical values. Indicates the cut-off of the correlation coefficient between each pair of
categorical variables. The first one is between each pair of 'uv.variables' and the second one is between each pair of 'bio.variables'.
The correlation is computed by the function ContCoef from the DescTools package. If the correlation of a pair of variable is higher than
the cut-off, then only the variable that has the highest number of factor will be kept and the other one will be excluded from the
remaining analysis. By default they are both set to 0.7.}

\item{cont.cor.coef}{Vector of two numerical values. Indicates the cut-off of the Spearman correlation coefficient between each pair of
continuous variables. The first one is between each pair of 'uv.variables' and the second one is between each pair of 'bio.variables'.
If the correlation of a pair of variable is higher than the cut-off, then only the variable that has the highest variance will
be kept and the other one will be excluded from the remaining analysis. By default they are both set to 0.7.}

\item{save.se.obj}{Logical. Indicates whether to save the result in the metadata of the SummarizedExperiment class object 'se.obj' or
to output the result. By default it is set to TRUE.}

\item{verbose}{Logical. Indicates whether to show or reduce the level of output or messages displayed during the execution
of the functions, by default it is set to TRUE.}
}
\value{
Either the SummarizedExperiment object containing the a set of negative control genes
or a logical vector of the selected negative control genes
}
\description{
is used to find a set of negative control genes (NCG) using ANOVA and correlation analyses across all samples.
}
