% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variablesCorrelation.R
\name{variablesCorrelation}
\alias{variablesCorrelation}
\title{is used to assess the association between variables in a SummarizedExperiment object.}
\usage{
variablesCorrelation(
  se.obj,
  bio.variables,
  uv.variables,
  cat.cor.coef = c(0.7, 0.7),
  cont.cor.coef = c(0.7, 0.7),
  assess.se.obj = TRUE,
  remove.na = "sample.annotation",
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{bio.variables}{String of the label of (a) categorical variable(s) that specifies major biological groups
such as samples types from colData(se).}

\item{uv.variables}{String or vector of strings of the label of continuous or categorical variable(s)
such as samples types, batch or library size from colData(se) that will be used to define PRPS.}

\item{cat.cor.coef}{Vector of two numerical values. Indicates the cut-off of the correlation coefficient between each
pair of categorical variables. The first one is between each pair of 'uv.variables' and the second one is between each
pair of 'bio.variables'. The correlation is computed by the function ContCoef from the DescTools package. If the
correlation of a pair of variable is higher than the cut-off, then only the variable that has the highest number of
factor will be kept and the other one will be excluded from the remaining analysis. By default they are both set to 0.7.}

\item{cont.cor.coef}{Vector of two numerical values. Indicates the cut-off of the Spearman correlation coefficient
between each pair of continuous variables. The first one is between each pair of 'uv.variables' and the second one is
between each pair of 'bio.variables'. If the correlation of a pair of variable is higher than the cut-off, then only
the variable that has the highest variance will be kept and the other one will be excluded from the remaining analysis.
By default they are both set to 0.7.}

\item{assess.se.obj}{Logical. Whether to assess the SummarizedExperiment object or not.}

\item{remove.na}{String. Indicates whether to remove NA or missing values from either the 'sample.annotation', 'both'
or 'none'. If 'assays' is selected, the genes that contains NA or missing values will be excluded. If 'sample.annotation'
is selected, the samples that contains NA or missing values for any 'bio.variables' and 'uv.variables' will be excluded.
By default, it is set to
'sample.annotation'.}

\item{verbose}{Logical. Indicates whether to show or reduce the level of output or messages displayed
during the execution of the functions, by default it is set to TRUE.}
}
\value{
a SummarizedExperiment object and the selected 'uv.variables' and 'bio.variables'.
}
\description{
The function assesses the association between all selected biological and unwanted variation variables separately.
If two categorical variables are highly correlated, the function keeps one that has the highest number of factors.
For two continuous variables, the one with higher variance will be kept.
}
\details{
For each pair of categorical variables from 'uv.variables' and each pair of categorical variables from 'bio.variables',
the correlation is computed using the function ContCoef from the DescTools R package. For each pair of continuous
variables from 'uv.variables' and each pair of continuous variables from 'bio.variables', the correlation is computed
using the Spearman correlation. The user defines a minimum cut-off of the correlation coefficient between each pair of
categorical variables in the 'cat.cor.coef' for the unwanted variables, followed by the minimum cut-off of the correlation
coefficient for the biological variables. If the correlation of a pair of those variables is higher than the minimum
cut-off, only the variable that has the highest number of factor will be kept and the other one will be excluded from
the remaining analysis. The user defines a minimum cut-off of the correlation coefficient between each pair of continuous
variables in the 'cont.cor.coef' for the unwanted variables, followed by the minimum cut-off of the correlation coefficient
for the biological variables. If the correlation of a pair of variable is higher than the minimum cut-off, only the variable
that has the highest variance will be kept and the other one will be excluded from the remaining analysis.
}
\author{
Ramyar Molania
}
