% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_correlation_pc_catvar.R
\name{vector_correlation_pc_catvar}
\alias{vector_correlation_pc_catvar}
\title{is used to compute the vector correlation between the first  PCs of the gene expression (assay)
of a SummarizedExperiment class object and a continous variable (i.e. library size).}
\usage{
vector_correlation_pc_catvar(
  pca,
  cat_var,
  cat_var_label,
  assay_names = NULL,
  plot_output = TRUE,
  nb_pca_comp = 10
)
}
\arguments{
\item{pca}{PCA components of a SummarizedExperiment variable that will be used in the plot.}

\item{cat_var}{Vector of a categorical variable such as sample types
(i.e. biological subtypes) or batches.}

\item{cat_var_label}{String or vector of strings of the label of categorical variable(s) such as
sample types or batches from colData(se).}

\item{assay_names}{Optional string or list of strings for selection of the names
of the assays of the SummarizedExperiment class object to compute the correlation.}

\item{plot_output}{Indicates whether to plot the results}

\item{nb_pca_comp}{The number of components of the PCA used to compute the regression.}
}
\description{
is used to compute the vector correlation between the first  PCs of the gene expression (assay)
of a SummarizedExperiment class object and a continous variable (i.e. library size).
}
