% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon_test.R
\name{wilcoxon_test}
\alias{wilcoxon_test}
\title{is used to compute the Wilcoxon Rank Sum and Signed Rank Test to data on a categorical variable}
\usage{
wilcoxon_test(expr.data, apply.log = FALSE, variable, n.cores)
}
\arguments{
\item{expr.data}{is the gene expression matrix genes by samples}

\item{apply.log}{Indicates whether to apply a log-transformation to the data}

\item{variable}{is a categorical variable such as sample types or batches}

\item{n.cores}{is the number of cpus used for mclapply parallelization}
}
\value{
dataframe containing the genes, the p-values before and after BH correction
}
\description{
is used to compute the Wilcoxon Rank Sum and Signed Rank Test to data on a categorical variable
}
