% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyLog.R
\name{applyLog}
\alias{applyLog}
\title{Apply log2 on assay(s) + pseudo.count.}
\usage{
applyLog(
  se.obj,
  assay.names = "all",
  apply.log = TRUE,
  pseudo.count = 1,
  assessment,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Symbol. A symbol or a vector of symbols for the selection of the name(s) of the assay(s) in the
SummarizedExperiment object. The default is set to "all, which indicates all the assays of the SummarizedExperiment
object will be selected.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data or not. The default is 'TRUE'.}

\item{pseudo.count}{Numeric. A value as a pseudo count to be added to all measurements of the assay(s) before applying
log transformation to avoid -Inf for measurements that are equal to 0. The default is 1.}

\item{assessment}{Symbol. Indicates the downstream analyses impacted by not applying log transformation to the data.}

\item{verbose}{Logical. If 'TRUE', shows the messages of different steps of the function.}
}
\value{
logged data
}
\description{
Apply log2 on assay(s) + pseudo.count.
}
\author{
Ramyar Molania
}
