% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyOtherNormalizations.R
\name{applyOtherNormalizations}
\alias{applyOtherNormalizations}
\title{Perform several normalization methods for RNA-seq data.}
\usage{
applyOtherNormalizations(
  se.obj,
  assay.name,
  method = "CPM",
  apply.log = TRUE,
  pseudo.count = 1,
  remove.na = "assays",
  assess.se.obj = TRUE,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A summarized experiment object.}

\item{assay.name}{Symbol.  Specifies the name of the assay within the SummarizedExperiment object. The specified assay
should be in raw count format.}

\item{method}{Symbol. Indicates the normalization method to apply on the 'assay.name'. Options are:
'CPM': Count Per Million from the edger R package.
'TMM': Trimmed Mean of M-values from the edger R package,
'upper':a scaling normalization that forces the median of each lane to be the same from the EDAseq R package.
'median': a scaling normalization that forces the upper quartile of each lane to be the same from the EDAseq R package.
'full': a non linear full quantile normalization from the EDAseq R package
'VST', variance stabilizing transformation from the DESeq2 R package.
By default it is set to 'CPM'.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data or not after the normalization
specified in the 'method' is performed. The default is 'TRUE'.}

\item{pseudo.count}{Numeric. A value as a pseudo count to be added to all measurements of the assay(s) after applying
the normalization to avoid -Inf for measurements that are equal to 0. The default is 1.}

\item{remove.na}{Logical. Indicates whether to remove NA or missing values from the assay or not.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment class object.
By default it is set to TRUE.}

\item{save.se.obj}{Logical. Indicates whether to save the result as new assay to the SummarizedExperiment
object or to output the result. The default it is set to TRUE.}

\item{verbose}{Logical. Indicates whether to show or reduce the level of output or messages displayed
during the execution of the functions, by default it is set to TRUE.}
}
\value{
a SummarizedExperiment object containing an assay with the selected normalisation method
}
\description{
This functions provides different normalization methods: CPM, TMM, upper, full, median, VST for RNA-seq
data.
}
\author{
Ramyar Molania
}
