% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyOtherNormalizations.R
\name{applyOtherNormalizations}
\alias{applyOtherNormalizations}
\title{is used to perform several normalization methods for RNA-seq data.}
\usage{
applyOtherNormalizations(
  se.obj,
  assay.name,
  method = "CPM",
  apply.log = TRUE,
  pseudo.count = 1,
  remove.na = "assays",
  assess.se.obj = TRUE,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A summarized experiment object.}

\item{assay.name}{Symbol. Indicates a name of the assay of the SummarizedExperiment object. This assay should be in raw count format.}

\item{method}{Symbol.Indicates the normalization method to use from 'CPM', 'TMM', 'upper', 'full', 'median', 'VST',
and 'Quantile'. By default it is set to 'CPM'.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data after normalization.
By default it is set to TRUE.}

\item{pseudo.count}{Numeric. Indicates a positive value as a pseudo count to be added to all measurements before log transformation,
by default it is set to 1.}

\item{remove.na}{Logical. Indicates whether to remove NA or missing values from the assay or not.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment class object.
By default it is set to TRUE.}

\item{save.se.obj}{Logical. Indicates whether to save the result as new assay to the SummarizedExperiment
object or to output the result. The default it is set to TRUE.}

\item{verbose}{Logical. Indicates whether to show or reduce the level of output or messages displayed
during the execution of the functions, by default it is set to TRUE.}
}
\value{
a SummarizedExperiment object containing an assay with the selected normalisation method
}
\description{
This functions provides different notmalization methods:  CPM, TMM, upper, full, median, VST, Quantile for RNA-seq
data.
}
\author{
Ramyar Molania
}
