% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyOtherNormalizations.R
\name{applyOtherNormalizations}
\alias{applyOtherNormalizations}
\title{is used to perform several normalization methods: CPM, TMM, upper, full, median, VST, Quantile.}
\usage{
applyOtherNormalizations(
  se.obj,
  assay.name,
  method = "CPM",
  pseudo.count = 1,
  apply.log = TRUE,
  assess.se.obj = TRUE,
  save.se.obj = TRUE,
  remove.na = "none",
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A summarized experiment object.}

\item{assay.name}{String for the selection of the name of the assay of the SummarizedExperiment class object used to define PRPS.}

\item{method}{String defining the normalization method to use from 'CPM', 'TMM', 'upper', 'full', 'median', 'VST',
and 'Quantile'. By default it is set to 'CPM'.}

\item{pseudo.count}{Numeric. A value as a pseudo count to be added to all measurements before log transformation,
by default it is set to 1.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment class object.
By default it is set to TRUE.}

\item{save.se.obj}{Logical. Indicates whether to save the result in the metadata of the SummarizedExperiment class
object 'se.obj' or to output the result, by default it is set to TRUE.}

\item{remove.na}{TO BE DEFINED.}

\item{verbose}{Logical. Indicates whether to show or reduce the level of output or messages displayed
during the execution of the functions, by default it is set to TRUE.}
}
\value{
a SummarizedExperiment object containing an assay with the selected normalisation method
}
\description{
is used to perform several normalization methods: CPM, TMM, upper, full, median, VST, Quantile.
}
