% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessNormalization.R
\name{assessNormalization}
\alias{assessNormalization}
\title{Assess the performance of normalization methods.}
\usage{
assessNormalization(
  se.obj,
  assay.names = "all",
  bio.variables,
  uv.variables,
  metrics.to.exclude = NULL,
  assessments.to.exclude = NULL,
  fast.pca = TRUE,
  sil.dist.measure = "euclidian",
  sil.nb.pcs = 3,
  ari.clustering.method = "hclust",
  ari.hclust.method = "complete",
  ari.hclust.dist.measure = "euclidian",
  ari.nb.pcs = 3,
  vca.nb.pcs = 3,
  lra.nb.pcs = 3,
  fstat.cutoff = 1,
  corr.coef.cutoff = 0.2,
  corr.method = "spearman",
  anove.method = "aov",
  output.file = NULL,
  plot.output = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Symbol. A symbol or a vector of symbols for the selection of the name(s) of the assay(s) in the
SummarizedExperiment object. The default is set to 'all', so all the assays in the SummarizedExperiment object will
be selected.}

\item{bio.variables}{Symbols. Indicating the column names that contain known biological variable(s) in the
SummarizedExperiment object. These biological variables can be categorical or continuous.}

\item{uv.variables}{Symbols. Indicating the column names that contain unwanted variable(s) in the SummarizedExperiment
object. These unwanted variables can be categorical or continuous.}

\item{metrics.to.exclude}{Symbols. A symbol or vector of symbols indication which metrics to be excluded.
The metrics can be obtained using the  'getAssessmentMetrics' function. The default is set to NULL. Refer
to the details for more information.}

\item{assessments.to.exclude}{Symbols. A symbol or vector of symbols indicating which assessment metrics to exclude.
These metrics can be acquired using the 'getAssessmentMetrics' function. Default is set to NULL. See the details for
further information.}

\item{fast.pca}{Logical. Indicates whether to calculate a specific number of PCs instead of the full range
to speed up the process, by default is set to 'TRUE'.}

\item{sil.dist.measure}{A character string indicating which method
is to be used for the differential analysis: 'euclidean', 'maximum', 'manhattan', 'canberra', 'binary' or 'minkowski'.
By default 'euclidean' will be selected.}

\item{sil.nb.pcs}{Numeric. Indicates the number of PCs that should be used to compute the silhouette coefficients.}

\item{ari.clustering.method}{Symbol. Indicates which clustering methods should be applied on the PCs calculate the ARI.
The function provides the 'mclust' or 'hclust' methods. The default is 'hclust'.}

\item{ari.hclust.method}{Symbol. Indicate the agglomeration method to be used for the 'hclust' method. This should be
one of 'ward.D', 'ward.D2', 'single', 'complete', 'average' (= UPGMA), 'mcquitty' (= WPGMA), 'median' (= WPGMC) or
'centroid' (= UPGMC). See the hclust function for more details.}

\item{ari.hclust.dist.measure}{Symbol. Indicates the distance measure to be used in the dist function. This must be
one of euclidean', 'maximum', 'manhattan', 'canberra', 'binary' or 'minkowski'. See the dist function for more details.}

\item{ari.nb.pcs}{Numeric. Indicates the number of PCs that should be used to compute the ARI.}

\item{vca.nb.pcs}{Numeric. Indicates the number of PCs that should be used to compute the ARI.}

\item{lra.nb.pcs}{Numeric. Indicates the number of PCs that should be used to compute the ARI.}

\item{fstat.cutoff}{Numeric. Specifies the threshold for selecting genes that have a F-statistics lower than this value
with categorical variables. The default value is 1.}

\item{corr.coef.cutoff}{Numeric. Specifies the threshold for selecting genes that have an absolute correlation coefficient
lower than this value with continuous variables. The default value is 0.2.}

\item{corr.method}{Symbol. Indicates which correlation methods should be used. Options are The default is 'spearman'.}

\item{anove.method}{Symbol. Indicates which ANOVA methods should be used. Options are ... . The default is 'aov'.}

\item{output.file}{Path and name of the output file to save the assessments plots in a pdf format.}

\item{plot.output}{BB}

\item{verbose}{Logical. If TRUE, displaying process messages is enabled.}
}
\value{
A SummarizedExperiment object containing all the assessments plots and metrics. If specified it will generate
a pdf containing the assessments plots and metrics used for the assessment.
}
\description{
This function summarizes a range of global and gene level metrics obtained using the 'assessVariation' function for
individual biological and unwanted variables. The functions returns numerical assessments in a table to assess the
performance of difference normalization. Refer to details for more information.
}
\details{
It is essential to assess the performance of a normalization as a remover of unwanted variation and a preserver of
biological variation in the data. An ideal normalization should preserve all known sources of biological variation
and reduce or remove the impact of all known sources of unwanted variation. To rank methods according to their
performance, we aggregate all evaluation metrics into a overall score as follows.
Several assessment will be performed:
For each categorical variable:
\itemize{
\item PCA plot of the categorical variable.
\item Silhouette and ARI computed on the categorical variable.
\item Differential analysis based ANOVA between the gene expression and the categorical variable.
\item Vector correlation between the first cumulative PCs of the gene expression and the categorical variable.
For each continous variable:
\item Linear regression between the first cumulative PC and continuous variable.
\item Correlation between gene expression and continuous variable.
}

It will output the following plots:
\itemize{
\item PCA plot of each categorical variable.
\item Boxplot of the F-test distribution from ANOVA between the gene expression and each categorical variable.
\item Vector correlation between the first cumulative PCs of the gene expression and each categorical variable.
\item Combined Silhouette plot of the combined pair of all categorical variables.
\item Linear regression between the first cumulative PC and continuous variable.
\item Boxplot of the correlation between gene expression and continuous variable.
\item It will also output the RLE plot distribution.
}
}
\references{
**Molania R., ..., Speed, T. P., A new normalization for Nanostring nCounter gene expression data, Nucleic Acids Research,
2019.
**Molania R., ..., Speed, T. P., Removing unwanted variation from large-scale RNA sequencing data with PRPS,
Nature Biotechnology, 2023
}
\author{
Ramyar Molania
}
