% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessVariation.R
\name{assessVariation}
\alias{assessVariation}
\title{is used to assess the performance of the normalisation of a SummarizedExperiment class object.}
\usage{
assessVariation(
  se.obj,
  assay.names = "all",
  variables,
  metrics = "all",
  metrics.to.exclude = NULL,
  rle.iqr.width = 2,
  rle.plot.ncol = 1,
  rle.median.points.size = 1,
  rle.geom.hline.color = "cyan",
  fast.pca = TRUE,
  compute.nb.pcs = 10,
  center.pca = TRUE,
  scale.pca = FALSE,
  pca.bsparam = bsparam(),
  nb.pcs.toplot.pca = 3,
  apply.log = TRUE,
  pseudo.count = 1,
  general.points.size = 1,
  sil.dist.measure = "euclidian",
  sil.nb.pcs = 3,
  ari.clustering.method = "hclust",
  ari.hclust.method = "complete",
  ari.hclust.dist.measure = "euclidian",
  ari.nb.pcs = 3,
  corr.method = "spearman",
  anove.method = "aov",
  assess.se.obj = TRUE,
  remove.na = "both",
  output.file = NULL,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object that will be used to compute the PCA.}

\item{assay.names}{Symbol. A symbol or a vector of symbols for the selection of the name(s) of the assay(s) of the
SummarizedExperiment class object. By default all the assays of the SummarizedExperiment class object will be selected.}

\item{variables}{String or vector of strings of the label of continuous or categorical variable(s)
such as samples types, batch or library size from colData(se).}

\item{metrics}{TTTT}

\item{metrics.to.exclude}{TTTT}

\item{rle.iqr.width}{TTT}

\item{rle.plot.ncol}{TTT}

\item{rle.median.points.size}{TTT}

\item{rle.geom.hline.color}{TTT}

\item{fast.pca}{Logical. Indicates whether to calculate a specific number of PCs instead of the full range
to speed up the process, by default is set to 'TRUE'.}

\item{compute.nb.pcs}{Numeric. The number of first PCs to be calculated for the fast pca process, by default is set to 10.}

\item{center.pca}{TTT}

\item{scale.pca}{TTTT}

\item{pca.bsparam}{TTTT}

\item{nb.pcs.toplot.pca}{TTTT}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data. By default
no transformation will be selected.}

\item{pseudo.count}{Numeric. A value as a pseudo count to be added to all measurements before log transformation,
by default it is set to 1.}

\item{general.points.size}{TTTT}

\item{sil.dist.measure}{TTTT}

\item{sil.nb.pcs}{TTTT}

\item{ari.clustering.method}{TTTT}

\item{ari.hclust.method}{TTTTT}

\item{ari.hclust.dist.measure}{TTTTT}

\item{ari.nb.pcs}{TTTTT}

\item{corr.method}{TTTTTT}

\item{anove.method}{TTTTTTT}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment class object.}

\item{remove.na}{TTTT}

\item{output.file}{Path and name of the output file to save the assessments plots in a pdf format.}

\item{verbose}{Logical. Indicates whether to show or reduce the level of output or messages displayed during the execution
of the functions, by default it is set to TRUE.}
}
\value{
SummarizedExperiment A SummarizedExperiment object containing all the assessments plots and metrics.
If specified it will generate a pdf containing the assessments plots and metrics used for the assessment.
}
\description{
Several assessment will be performed:
For each categorical variable:
\itemize{
\item PCA plot of the categorical variable.
\item Silhouette and ARI computed on the categorical variable.
\item Differential analysis based ANOVA between the gene expression and the categorical variable.
\item Vector correlation between the first cumulative PCs of the gene expression and the categorical variable.
For each continous variable:
\item Linear regression between the first cumulative PC and continuous variable.
\item Correlation between gene expression and continuous variable.
}
}
\details{
It will output the following plots:
\itemize{
\item PCA plot of each categorical variable.
\item Boxplot of the F-test distribution from ANOVA between the gene expression and each categorical variable.
\item Vector correlation between the first cumulative PCs of the gene expression and each categorical variable.
\item Combined Silhouette plot of the combined pair of all categorical variables.
\item Linear regression between the first cumulative PC and continuous variable.
\item Boxplot of the correlation between gene expression and continuous variable.
\item It will also output the RLE plot distribution.
}
}
