% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPRPSconnectedness.R
\name{checkPRPSconnectedness}
\alias{checkPRPSconnectedness}
\title{is used to check PRPS sets connection.}
\usage{
checkPRPSconnectedness(data.input, min.samples, batch.name, verbose = TRUE)
}
\arguments{
\item{data.input}{TTTTT}

\item{min.samples}{TTTTTT}

\item{batch.name}{TTTTTT}

\item{verbose}{TTTTTT}
}
\description{
is used to check PRPS sets connection.
}
