% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSeObj.R
\name{checkSeObj}
\alias{checkSeObj}
\title{Assess the assay names, variables, and missing values in a SummarizedExperiment object.}
\usage{
checkSeObj(
  se.obj,
  assay.names = "all",
  variables = "all",
  remove.na = "both",
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Symbol. A symbol or vector of symbols used to specify the name(s) of the assay(s) in the
SummarizedExperiment object. The default is "all," indicating that all assays in the SummarizedExperiment object will
be assessed.}

\item{variables}{Symbol. A symbol or a vector of symbols specifying the name(s) of the column(s) in the sample
annotation of the SummarizedExperiment object. By default, it is set to 'all', then all the columns will be examined.
We recommend specifying those column(s) that are of interest to your analysis.}

\item{remove.na}{Symbol. Specifies whether to eliminate missing values from either 'assays', 'sample.annotation',}

\item{verbose}{Logical. If 'TRUE', shows the messages of different steps of the function.}
}
\value{
A SummarizedExperiment object.
}
\description{
This function assesses the structure of a SummarizedExperiment object and removes any missing values from both
assay(s) and sample annotation. When multiple assays are provided, if they are missing in only one of the assays,
the corresponding rows in other assays will be remove as well. Please note that, the current RUV-III-PRPS method does
not support missing values in the assay(s).
}
\author{
Ramyar Molania
}
