% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckSeObj.R
\name{checkSeObj}
\alias{checkSeObj}
\title{is used to check the SummarizedExperiment class object and if selected to exclude some of the samples or genes that
contains NA or missing values from the analysis.}
\usage{
checkSeObj(
  se.obj,
  assay.names = "All",
  variables = "All",
  remove.na = "both",
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{String for the selection of the name, or names of the assay of the SummarizedExperiment class object.
By default it is set to 'All'.}

\item{variables}{String. A name or names of the variables in the sample annotation of the SummarizedExperiment class object.
By default it is set to 'All'.}

\item{remove.na}{String. Indicates whether to remove NA or missing values from either the 'assays', 'sample.annotation', 'both' or 'none'.
If 'assays' is selected, the genes that contains NA or missing values will be excluded. If 'sample.annotation' is selected, the
samples that contains NA or missing values for any 'bio.variables' and 'uv.variables' will be excluded. By default, it is set to
'both'.}

\item{verbose}{Logical. Indicates whether to show or reduce the level of output or messages displayed
during the execution of the functions, by default it is set to TRUE.}
}
\value{
a SummarizedExperiment object
}
\description{
is used to check the SummarizedExperiment class object and if selected to exclude some of the samples or genes that
contains NA or missing values from the analysis.
}
