% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSeObj.R
\name{checkSeObj}
\alias{checkSeObj}
\title{is used to check the assay names, variables and missing/NA values in a SummarizedExperiment object.}
\usage{
checkSeObj(
  se.obj,
  assay.names = "all",
  variables = "all",
  remove.na = "both",
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Symbol. Indicates the name(s) of the assay(s) in the SummarizedExperiment object. By default it is
set to 'all', then all the assay(s) will be checked.}

\item{variables}{Symbol. Indicates the name(s) of the column(s) in the sample annotation of the SummarizedExperiment
object. By default it is set to 'all', then all the columns will be checked.}

\item{remove.na}{Symbol. Indicates whether to remove missing/NA values from either the 'assays', 'sample.annotation',
'both' or 'none'. If 'assays' is selected, the genes that contains missing/NA values will be excluded. If 'sample.annotation'
is selected, the samples that contains NA or missing values for each 'variables' will be excluded. By default, it is
set to 'both'.}

\item{verbose}{Logical. If TRUE shows the messages.}
}
\value{
A SummarizedExperiment object.
}
\description{
This functions assesses the structure of a SummarizedExperiment object and removes any missing/NA values from both
assay(s) and sample annotation. Please note that, the current RUV-III-PRPS method cannot handle missing/NA values in
the assay(s).
}
\author{
Ramyar Molania
}
