% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckSeObj.R
\name{checkSeObj}
\alias{checkSeObj}
\title{is used to check summarized experiment object.}
\usage{
checkSeObj(
  se.obj,
  assay.names = "All",
  variables = "All",
  remove.na = "both",
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A summarized experiment object.}

\item{assay.names}{A name or names of the assays in summarized experiment object.}

\item{variables}{A name or names of the variables in the sample annotation of the summarized experiment object.}

\item{remove.na}{'both', 'measurements', 'sample.annotation', 'none'}

\item{verbose}{whether to show the messages or not.}
}
\description{
is used to check summarized experiment object.
}
