% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ari.R
\name{compute_ari}
\alias{compute_ari}
\title{is used to compute the adjusted rank index (ARI) from the first PC of all assays
given a categorical variable}
\usage{
compute_ari(pca, cat_var, assay_names = NULL, plot = FALSE, nPCs = 3)
}
\arguments{
\item{pca}{PCs of the dataset that will be used}

\item{cat_var}{is a categorical variable such as sample types or batches}

\item{assay_names}{Optional selection of names of the assays to compute the ARI}

\item{plot}{Optional output of a plot, default set to FALSE.}

\item{nPCs}{is the number of PCs used to measure the distance, default is set to 3.}
}
\value{
list List containing the association plot and the computed ari
}
\description{
is used to compute the adjusted rank index (ARI) from the first PC of all assays
given a categorical variable
}
