% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pca.R
\name{compute_pca}
\alias{compute_pca}
\title{is used to compute PCA of the data on all assays}
\usage{
compute_pca(sce, assay_names = NULL, apply.log = FALSE)
}
\arguments{
\item{sce}{Dataset that will be used to compute the PCA}

\item{assay_names}{Optional selection of names of the assays to compute the PCA}

\item{apply.log}{Indicates whether to apply a log-transformation to the data}
}
\value{
pca.all PCA components computed on the provided data
}
\description{
is used to compute PCA of the data on all assays
}
