% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeARI.R
\name{computeARI}
\alias{computeARI}
\title{is used to compute the adjusted rand index (ARI).}
\usage{
computeARI(
  se.obj,
  assay.names = "All",
  variable,
  fast.pca = TRUE,
  nb.pcs = 3,
  clustering.method = "hclust",
  hclust.method = "complete",
  dist.measure = "euclidian",
  plot.output = FALSE,
  save.se.obj = TRUE,
  assess.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Symbol. A symbol or list of symbols for the selection of the name(s) of the assay(s) in the
SummarizedExperiment object to compute PCA. By default all the assays of the SummarizedExperiment object will be selected.}

\item{variable}{Symbol. Indicates the column name in the SummarizedExperiment object that contains a categorical
variable such as sample types or batches.}

\item{fast.pca}{Logical. Indicates whether to use the PCA calculated using a specific number of PCs instead of the
full range to speed up the process, by default is set to 'TRUE'.}

\item{nb.pcs}{Numeric. The number of first PCs to be calculated for the fast pca process, by default is set to 3.}

\item{clustering.method}{Symbol. Indicates which clustering methods to be used. The function provides the mclust or
hclust methods. The default is hclust.}

\item{hclust.method}{Symbol. Indicate the agglomeration method to be used for the hclust methid. This should be
one of "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or
"centroid" (= UPGMC). See the hclust function for more details.}

\item{dist.measure}{Symbol. Indicates the distance measure to be used in the dist function. This must be one of
"euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski". See the dist function for more details.}

\item{plot.output}{Logical. Indicates whether to plot the ARI, by default it is set to FALSE.}

\item{save.se.obj}{Logical. Indicates whether to save the result in the metadata of the SummarizedExperiment class
object 'se.obj' or to output the result. By default it is set to TRUE.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment class object, by default it is
set to TRUE.}

\item{verbose}{Logical. Indicates whether to show or reduce the level of output or messages displayed during the
execution of the functions, by default it is set to TRUE.}
}
\value{
A SummarizedExperiment object or a list that containing the computed ARI on the categorical variable.
}
\description{
This functions computes the adjusted rand index using the first PCs of the assays in a SummarizedExperiment object
given a categorical variable.
}
\author{
Ramyar Molania
}
