% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeARI.R
\name{computeARI}
\alias{computeARI}
\title{is used to compute the adjusted rand index (ARI) using the first PC of the assays in a SummarizedExperiment
object given a categorical variable.}
\usage{
computeARI(
  se.obj,
  assay.names = "All",
  variable,
  fast.pca = TRUE,
  nb.pcs = 3,
  clustering.method = "hclust",
  hclust.method = "complete",
  dist.measure = "euclidian",
  save.se.obj = TRUE,
  plot.output = FALSE,
  assess.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object that will be used to compute the PCA.}

\item{assay.names}{Optional string or list of strings for the selection of the name(s)
of the assay(s) of the SummarizedExperiment class object to compute the ARI. By default}

\item{variable}{String of the label of a categorical variable such as
sample types or batches from colData(se.obj).}

\item{fast.pca}{Logical. Indicates whether to use the PCA calculated using a specific number of PCs instead of the full range
to speed up the process, by default is set to 'TRUE'.}

\item{nb.pcs}{Numeric. The number of first PCs to be calculated for the fast pca process, by default is set to 3.}

\item{clustering.method}{Symbol. Indicates the clustering methods.}

\item{hclust.method}{Symbol.}

\item{dist.measure}{Indicates the clustering methods.}

\item{save.se.obj}{Logical. Indicates whether to save the result in the metadata of the SummarizedExperiment class object 'se.obj' or
to output the result. By default it is set to TRUE.}

\item{plot.output}{Logical. Indicates whether to plot the ARI, by default it is set to FALSE.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment class object, by default it is set to TRUE.}

\item{verbose}{Logical. Indicates whether to show or reduce the level of output or messages displayed during the execution
of the functions, by default it is set to TRUE.}
}
\value{
SummarizedExperiment A SummarizedExperiment object containing the computed ARI
on the categorical variable.
}
\description{
It can be used to assess how a group of biological samples are distributed across batches (i.e. example subtypes vs batch), and how batches ### update
}
