% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeARI.R
\name{computeARI}
\alias{computeARI}
\title{compute the adjusted rand index (ARI).}
\usage{
computeARI(
  se.obj,
  assay.names = "all",
  variable,
  clustering.method = "hclust",
  hclust.method = "complete",
  hclust.dist.measure = "euclidian",
  fast.pca = TRUE,
  nb.pcs = 3,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Symbol. A symbol or a vector of symbols for the selection of the name(s) of the assay(s) in the
SummarizedExperiment object to compute ARI. The default is set to 'all', so all the assays of the SummarizedExperiment
object will be selected.}

\item{variable}{Symbol. Indicates the column name that contain categorical variable in the SummarizedExperiment object.
The variable can be biological or unwanted variable.}

\item{clustering.method}{Symbol. A symbol that indicates which clustering methods should be applied on the PCs to
calculate the ARI. The options are 'mclust' or 'hclust' methods. The default is set to 'hclust'.}

\item{hclust.method}{Symbol. Indicate the agglomeration method to be used when the  'clustering.method' is 'hclust'
method. The options are: 'ward.D', 'ward.D2', 'single', 'complete', 'average' (= UPGMA), 'mcquitty' (= WPGMA),
'median' (= WPGMC) or centroid' (= UPGMC). Refer to the 'hclust' function from the stats R package for more details.}

\item{hclust.dist.measure}{Symbol. Indicates the distance measure to be used in the 'dist' function. The options are
'euclidean', 'maximum', 'manhattan', 'canberra', 'binary' or 'minkowski'. Refer to the 'dist' function from the stats
R package for more details.}

\item{fast.pca}{Logical. Indicates whether to use the PC calculated using fast PCA or not. The default is set to 'TRUE'.
The fast PCA and ordinary PCA do not affect the silhouette coefficient calculation. Refer to detail for more information.}

\item{nb.pcs}{Numeric. The number of first PCs to be used to calculated the distances between samples. The default is
set to 3.}

\item{save.se.obj}{Logical. Indicates whether to save the result in the metadata of the SummarizedExperiment class
object 'se.obj' or to output the result. By default it is set to TRUE.}

\item{verbose}{Logical. If 'TRUE', shows the messages of different steps of the function.}
}
\value{
A SummarizedExperiment object or a list that containing the computed ARI on the categorical variable.
}
\description{
This functions computes the adjusted rand index for given a categorical variable using the first PCs of the assay(s)
in a SummarizedExperiment object.
}
\details{
The ARI is the corrected-for-chance version of the Rand index. The ARI measures the percentage of matches between
two label lists. We use the ARI to assess the performance of normalization methods in terms of sample subtype
separation and batch mixing. We first calculate PCs and use the first PCs to perform ARI.
}
\references{
Molania R., ..., Speed, T. P., Removing unwanted variation from large-scale RNA sequencing data with PRPS,
Nature Biotechnology, 2023
}
\author{
Ramyar Molania
}
