% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeGenesPartialCorrelation.R
\name{computeGenesPartialCorrelation}
\alias{computeGenesPartialCorrelation}
\title{Compute gene-gene partial correlation.}
\usage{
computeGenesPartialCorrelation(
  se.obj,
  assay.names,
  variable,
  method = "spearman",
  genes = NULL,
  select.genes = TRUE,
  reference.data = NULL,
  corr.coff.cutoff = 0.3,
  apply.log = TRUE,
  pseudo.count = 1,
  apply.round = TRUE,
  assess.se.obj = TRUE,
  remove.na = "none",
  override.check = FALSE,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Symbol. A symbol or a vector of symbols for the selection of the name(s) of the assay(s) in the
SummarizedExperiment object to calculate RLE data, medians and interquartile ranges. The default is set to "all, which
indicates all the assays of the SummarizedExperiment object will be selected.}

\item{variable}{Symbol. A symbol that indicates the column name of the SummarizedExperiment object that contains a
continuous variable such as library size, tumor purity, ... .}

\item{method}{Symbol. A symbol that indicates which correlation methods should be used. The options are 'pearson',
'kendall', or "spearman". The default is set to 'spearman'.}

\item{genes}{Vector. A vector comprising genes upon which the correlation analysis will be performed. The vector can be
logical, numeric or gene names. The default is 'NULL'.}

\item{select.genes}{Logical. If 'TRUE', the function will compute correlation between individual genes and the specified
variable, then select a subset of gene based the "corr.coff.cutoff" for down-stream analysis. The default is set to
'TRUE'.}

\item{reference.data}{Symbol. A symbol specifying the name of the assay to be used to select genes using the correlation
analysis. The default is NULL. This means all the specified assays will be used.}

\item{corr.coff.cutoff}{Numeric. A numeric value to be used as a cutoff for selecting the genes. The default is set to
0.7.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data or not. The default is TRUE.}

\item{pseudo.count}{Numeric. A value as a pseudo count to be added to all measurements of the assay(s) before applying
log transformation to avoid -Inf for measurements that are equal to 0. The default is 1.}

\item{apply.round}{Logical. Indicates whether to round the correlations coefficients or not. The default is set to 'TRUE'.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment object or not. See the checkSeObj
function for more details.}

\item{remove.na}{Symbol. To remove NA or missing values from the assays or not. The options are 'assays' and 'none'.
The default is "assays", so all the NA or missing values from the assay(s) will be removed before computing RLE. See
the checkSeObj function for more details.}

\item{override.check}{Logical. When set to 'TRUE', the function verifies the current SummarizedExperiment object to
determine if the PPcorr has already been computed for the current parameters. If it has, the metric will not be recalculated.
The default is set to FALSE.}

\item{save.se.obj}{Logical. Indicates whether to save the RLE results in the metadata of the SummarizedExperiment object or
to output the result as list. By default it is set to TRUE.}

\item{verbose}{Logical. If 'TRUE', shows the messages of different steps of the function.}
}
\value{
The SummarizedExperiment object that contains the correlation coefficients or a list of the correlation coefficient
for individual assays.
}
\description{
This function computes all possible gene-gene pairwise ordinary and partial correlation of the assays in the
SummarizedExperiment object.
}
\details{
Partial correlation is used to estimate correlation between two variables while controlling for third
variables.
}
\author{
Ramyar Molania
}
