% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeGenesVariableAnova.R
\name{computeGenesVariableAnova}
\alias{computeGenesVariableAnova}
\title{Compute ANOVA between individual gene expression and a categorical variable.}
\usage{
computeGenesVariableAnova(
  se.obj,
  assay.names = "all",
  variable,
  method = "aov",
  apply.log = TRUE,
  pseudo.count = 1,
  plot.top.genes = FALSE,
  nb.top.genes = 3,
  apply.round = TRUE,
  assess.se.obj = TRUE,
  remove.na = "both",
  save.se.obj = TRUE,
  override.check = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Symbol. A symbol or vector of symbols for the selection of the name(s) of the assay(s) of the
SummarizedExperiment object to compute the correlation. By default all the assays of the SummarizedExperiment class
object will be selected.}

\item{variable}{Symbol. A symbol indicating a column name of the SummarizedExperiment object that contains a
categorical variable such as experimental batches, ... .}

\item{method}{Symbol. A symbol indicating which method is to be used for the ANOVA analysis. The options are "aov" and
"welch". The default is set to "aov".}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data before applying ANOVA. The default
is set to 'TRUE'.}

\item{pseudo.count}{Numeric. A numeric value as a pseudo count to be added to all measurements before applying log transformation.
The default is set to 1.}

\item{plot.top.genes}{Logical. Indicates whether to plot the gene expression of "nb.top.genes" number genes with highest
and lowest F-statistics across the specified variable. The default is set to 'FALSE'.}

\item{nb.top.genes}{Numeric. A numeric value Defining the number of genes from the top or bottom listing of ANOVA
F-statistics to plot. The default is set to 3.}

\item{apply.round}{Logical. Indicates whether to round the values of F-statistics. The default is set to 'TRUE'.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment class object. The default is set to
'TRUE'.}

\item{remove.na}{Symbol. A symbol that specifies whether to eliminate missing values from either 'assays', 'sample.annotation',}

\item{save.se.obj}{Logical. Indicates whether to save the results, ANOVA F-statistics and p-values, in the metadata
of the SummarizedExperiment object or to output the result as a list or vector. The default is set to 'TRUE'.}

\item{override.check}{Logical. When set to 'TRUE', the function verifies the current SummarizedExperiment object to
determine if the ANOVA has already been computed for the current parameters. If it has, the metric will not be recalculated.
The default is set to FALSE.}

\item{verbose}{Logical. If 'TRUE', shows the messages of different steps of the function.}
}
\value{
A SummarizedExperiment object containing the log2 F-statistics of ANOVA on the continuous
variable and if requested the associated boxplot.
}
\description{
This function calculates the ANOVA between individual gene expression of the assays in a SummarizedExperiment object
and a categorical variable as factor.
}
\details{
ANOVA enables us to assess the effects of a given qualitative variable (which we call a factor) on gene expression
measurements across any set of groups (labeled by the levels of the factor) under study. We use ANOVA F-statistics
to summarize the effects of a qualitative source of unwanted variation (for example, batches) on the expression levels
of individual genes, where genes having large F-statistics are deemed to be affected by the unwanted variation.
We also use ANOVA tests (the aov() function in R) to assign P values to the association between tumor purity and
molecular subtypes.
}
\author{
Ramyar Molania
}
