% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeGenesVariableAnova.R
\name{computeGenesVariableAnova}
\alias{computeGenesVariableAnova}
\title{compute ANOVA between individual gene expression and a categorical variable.}
\usage{
computeGenesVariableAnova(
  se.obj,
  assay.names = "all",
  variable,
  method = "aov",
  apply.log = TRUE,
  pseudo.count = 1,
  plot.top.genes = FALSE,
  nb.top.genes = 3,
  apply.round = TRUE,
  assess.se.obj = TRUE,
  remove.na = "both",
  save.se.obj = TRUE,
  plot.output = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Optional string or list of strings for the selection of the name(s)
of the assay(s) of the SummarizedExperiment class object to compute the ANOVA. By default}

\item{variable}{String of the label of a categorical variable such as
sample types or batches from colData(se.obj).}

\item{method}{A character string indicating which method
is to be used for the differential analysis: "aov" or "welch.correction". By default "aov" will
be selected.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data. By default
the log transformation will be selected.}

\item{pseudo.count}{Numeric. A value as a pseudo count to be added to all measurements before log transformation,
by default it is set to 1.}

\item{plot.top.genes}{Logical. Indicates whether to plot the gene expression of the number of genes
from the top listing of anova by default it is set to FALSE.}

\item{nb.top.genes}{Defines the number of genes from the top or bottom listing of anova to plot,
by default is set to 3.}

\item{apply.round}{Logical. Indicates whether to round the ARI results, by default it is set to TRUE.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment class object.}

\item{remove.na}{TO BE DEFINED.}

\item{save.se.obj}{Logical. Indicates whether to save the result in the metadata of the SummarizedExperiment object
'se.obj' or to output the result. By default it is set to TRUE.}

\item{plot.output}{Logical. Indicates whether to plot the boxplot of the F-test statistics, by default it is set to TRUE.}

\item{verbose}{Logical. If TRUE, displaying process messages is enabled.}
}
\value{
SummarizedExperiment A SummarizedExperiment object containing the log2 F-statistics of ANOVA on the continuous
variable and if requested the associated boxplot.
}
\description{
This function calculates the ANOVA between individual gene expression of the assays in a SummarizedExperiment object
and a categorical variable as factor.
}
\details{
ANOVA enables us to assess the effects of a given qualitative variable (which we call a factor) on gene expression
measurements across any set of groups (labeled by the levels of the factor) under study. We use ANOVA F-statistics
to summarize the effects of a qualitative source of unwanted variation (for example, batches) on the expression levels
of individual genes, where genes having large F-statistics are deemed to be affected by the unwanted variation.
We also use ANOVA tests (the aov() function in R) to assign P values to the association between tumor purity and
molecular subtypes.
}
\author{
Ramyar Molania
}
