% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePCA.R
\name{computePCA}
\alias{computePCA}
\title{is used to perform Principal Component Analysis (PCA) using singular value decomposition of the assays in a SummarizedExperiment object.}
\usage{
computePCA(
  se.obj,
  assay.names = "All",
  fast.pca = TRUE,
  nb.pcs = 10,
  scale = FALSE,
  center = TRUE,
  apply.log = TRUE,
  pseudo.count = 1,
  BSPARAM = NULL,
  assess.se.obj = TRUE,
  remove.na = "assays",
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Symbol. Optional symbol or list of symbols for the selection of the name(s)
of the assay(s) of the SummarizedExperiment object to compute PCA. By default
all the assays of the SummarizedExperiment object will be selected.}

\item{fast.pca}{Logical. Indicates whether to calculate a specific number of left singular vectors instead of the full range
to speed up the process, by default is set to 'TRUE'.}

\item{nb.pcs}{Numeric. The number of first left singular vectors to be calculated for the fast PCA process, by default is set to 10.}

\item{scale}{Logical. Indicates whether to scale the data or not.  If scale is TRUE then scaling is done by dividing the (centered)
columns of assays by their standard deviations if center is TRUE, and the root mean square otherwise.The default is FALSE.}

\item{center}{Logical.Indicates whether to scale the data or not. If center is TRUE then centering is done by subtracting the column means (omitting NAs)
of x from their corresponding columns. The default is TRUE.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data. By default
log2 transformation will be applied.}

\item{pseudo.count}{Numeric. A value as a pseudo count to be added to all measurements before log transformation,
by default it is set to 1.}

\item{BSPARAM}{A BiocParallelParam object specifying how parallelization should be performed.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment class object.}

\item{remove.na}{To remove NA or missing values from the assays or not. The options are 'assays' and 'none'.}

\item{save.se.obj}{Logical. Indicates whether to save the result in the metadata of the SummarizedExperiment object 'se.obj' or
to output the result as list. By default it is set to TRUE.}

\item{verbose}{Logical. Indicates whether to show or reduce the level of output or messages displayed during the execution
of the functions, by default it is set to TRUE.}
}
\value{
SummarizedExperiment A SummarizedExperiment object or a list that containing the SVD results.
}
\description{
is used to perform Principal Component Analysis (PCA) using singular value decomposition of the assays in a SummarizedExperiment object.
}
