% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePCA.R
\name{computePCA}
\alias{computePCA}
\title{is used to perform principal component analysis (PCA) using singular value decomposition (SVD).}
\usage{
computePCA(
  se.obj,
  assay.names = "All",
  fast.pca = TRUE,
  nb.pcs = 10,
  scale = FALSE,
  center = TRUE,
  apply.log = TRUE,
  pseudo.count = 1,
  BSPARAM = NULL,
  assess.se.obj = TRUE,
  remove.na = "assays",
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Symbol. A symbol or list of symbols of the assay(s) in the SummarizedExperiment object to compute
PCA. By default all the assays of the SummarizedExperiment object will be selected.}

\item{fast.pca}{Logical. Indicates whether to calculate a specific number of left singular vectors instead of the
full range to speed up the process, by default is set to 'TRUE'.}

\item{nb.pcs}{Numeric. The number of first left singular vectors to be calculated for the fast PCA process, by default
is set to 10.}

\item{scale}{Logical. Indicates whether to scale the data or not before applying SVD.  If scale is TRUE, then scaling
is done by dividing the (centered) columns of the assays by their standard deviations if center is TRUE, and the root
mean square otherwise. The default is FALSE.}

\item{center}{Logical. Indicates whether to scale the data or not. If center is TRUE, then centering is done by
subtracting the column means of the assay from their corresponding columns. The default is TRUE.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data. By default
log2 transformation will be applied.}

\item{pseudo.count}{Numeric. A value as a pseudo count to be added to all measurements before log transformation,
by default it is set to 1.}

\item{BSPARAM}{A BiocParallelParam object specifying how parallelization should be performed.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment class object.}

\item{remove.na}{To remove NA or missing values from the assays or not. The options are 'assays' and 'none'.}

\item{save.se.obj}{Logical. Indicates whether to save the result in the metadata of the SummarizedExperiment object
'se.obj' or to output the result as list. By default it is set to TRUE.}

\item{verbose}{Logical. Indicates whether to show or reduce the level of output or messages displayed during the
execution of the functions, by default it is set to TRUE.}
}
\value{
A SummarizedExperiment object or a list that containing the singular value decomposition results and the
percentage variation of each PCs.
}
\description{
This function uses ingular value decomposition to perform principal component on the assay(s) in a SummarizedExperiment
object. The function provides fast singular value decomposition using the BiocSingular R package.
}
\details{
The PCs (in this context also called singular vectors) of the sample × transcript array of log counts are the linear
combinations of the transcript measurements having the largest, second largest, third largest, etc., variation,
standardized to be of unit length and orthogonal to the preceding components. Each will give a single value for
each sample.
}
