% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePCA.R
\name{computePCA}
\alias{computePCA}
\title{is used to compute PCA of the gene expression (assay) of a SummarizedExperiment class object.}
\usage{
computePCA(
  se.obj,
  assay.names = "All",
  apply.log = TRUE,
  fast.pca = TRUE,
  nb.pcs = 10,
  return.pc.percentage = TRUE,
  scale = FALSE,
  center = TRUE,
  BSPARAM = NULL,
  save.se.obj = TRUE,
  assess.se.obj = TRUE,
  verbose = TRUE,
  pseudo.count = 1
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object that will be used to compute the PCA.}

\item{assay.names}{Optional string or list of strings for the selection of the name(s)
of the assay(s) of the SummarizedExperiment class object to compute the correlation. By default}

\item{apply.log}{Indicates whether to apply a log-transformation to the data. By default
no transformation will be selected.}

\item{fast.pca}{logical. Indicates whether to calculate a specific number of PCs instead of the full range
to speed up the process, by default is set to 'TRUE'.}

\item{nb.pcs}{Numeric. The number of first PCs to be calculated for the fast pca process, by default is set to 10.}

\item{return.pc.percentage}{TO BE DEFINED.}

\item{scale}{Either a logical value or a numeric-alike vector of length equal
to the number of columns of the gene expression (assay) of a SummarizedExperiment class object.
It is a generic function to scale the columns of a numeric matrix, by default it is set to 'FALSE'.}

\item{center}{Either a logical value or a numeric-alike vector of length equal
to the number of columns of the gene expression (assay) of a SummarizedExperiment class object.
It is a generic function to center the columns of a numeric matrix, by default is set to 'TRUE'.}

\item{BSPARAM}{A BiocParallelParam object specifying how parallelization should be performed.}

\item{save.se.obj}{Indicates whether to save the result in the metadata of the SummarizedExperiment class object 'se.obj' or
to output the result. By default it is set to TRUE.}

\item{assess.se.obj}{Indicates whether to assess the SummarizedExperiment class object.}

\item{verbose}{Indicates whether to show or reduce the level of output or messages displayed during the execution
of the functions, by default it is set to TRUE.}

\item{pseudo.count}{Numeric. A value as a pseudo count to be added to all measurements before log transformation,
by default it is set to 1.}
}
\value{
SummarizedExperiment A SummarizedExperiment object containing the PCA.
}
\description{
is used to compute PCA of the gene expression (assay) of a SummarizedExperiment class object.
}
