% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePCVariableCorrelation.R
\name{computePCVariableCorrelation}
\alias{computePCVariableCorrelation}
\title{is used to compute the vector correlation.}
\usage{
computePCVariableCorrelation(
  se.obj,
  assay.names = "all",
  variable,
  fast.pca = TRUE,
  nb.pcs = 10,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Symbol. A symbol or list of symbols for the selection of the name(s) of the assay(s) in the
SummarizedExperiment object to calculate RLE data, medians and interquartiles. The default is "all, which indicates all
the assays of the SummarizedExperiment object will be selected.}

\item{variable}{Symbol. Indicates a name of the column in the sample annotation of the SummarizedExperiment object.
The variable must be a categorical variable.}

\item{fast.pca}{Logical. Indicates whether to use the fast PCA or PCA results computed by the computePCA function. The
default is 'TRUE'.}

\item{nb.pcs}{Numeric. The number of first PCs to use to calculate the vector correlation. The default is 10. This
number cannot be bigger that number of PCs calculated by the computePCA function.}

\item{save.se.obj}{Logical. Indicates whether to save the vector correlation plots in the metadata of the SummarizedExperiment
object or to output the results as list. By default it is set to 'TRUE'.}

\item{verbose}{Logical. If TRUE, displaying process messages is enabled.}
}
\value{
A SummarizedExperiment object or a list that contains the vector correlation plots of individual assay(s) for
the categorical variable.
}
\description{
This function calculates the the vector correlation between the first cumulative PCs of the gene expression (assay)
of a SummarizedExperiment object and a categorical variable (i.e. batch). Then, the functions generates a line-dot plot
between the first cumulative PCs and the correlation coefficient to see the relationship between different PCs with the
variable. An ideal normalization should results a low correlation with unwanted variation variables and high correlation
with known biology.
}
\details{
We use the Rozeboom squared vector correlation to quantify the strength of (linear) relationships between two sets
of variables, such as the first k PCs (that is 1 ≤ k ≤ 10) and dummy variables representing time, batches, plates and
biological variables. Not only does this quantity summarize the full set of canonical correlations, but it also reduces
to the familiar R2 from multiple regression when one of the variable sets contains just one element.
}
\references{
Molania R., ..., Speed, T. P., Removing unwanted variation from large-scale RNA sequencing data with PRPS,
Nature Biotechnology, 2023
}
\author{
Ramyar Molania
}
