% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePCVariableRegression.R
\name{computePCVariableRegression}
\alias{computePCVariableRegression}
\title{compute the linear regression.}
\usage{
computePCVariableRegression(
  se.obj,
  assay.names = "all",
  variable,
  fast.pca = TRUE,
  nb.pcs = 10,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Symbol. A symbol or list of symbols for the selection of the name(s) of the assay(s) in the
SummarizedExperiment object to calculate regression analysis. The default is "all, which indicates all
the assays of the SummarizedExperiment object will be selected.}

\item{variable}{Symbol. Indicates a name of the column in the sample annotation of the SummarizedExperiment object.
The variable must be a continuous variable.}

\item{fast.pca}{Logical. Indicates whether to use the computed fast PCA or PCA results computed by the computePCA function. The
default is 'TRUE'.}

\item{nb.pcs}{Numeric. The number of first PCs to use to plot the vector correlation. The default is 10.}

\item{save.se.obj}{Logical. Indicates whether to save the vector correlation plots to the meta data of the
SummarizedExperiment object or to output the result as list. By default it is set to TRUE.}

\item{verbose}{Logical. If TRUE, displaying process messages is enabled.}
}
\value{
A SummarizedExperiment object containing the computed regression for
the continuous variable and if requested the associated plot.
}
\description{
This functions calculates the linear regression between the the first cumulative PCs of the gene expression (assays)
of a SummarizedExperiment object and a continuous variable (i.e. library size).
}
\details{
R2 values of fitted linear models are used to quantity the strength of the (linear) relationships between a single
quantitative source of unwanted variation, such as sample (log) library size or tumor purity, and global sample
summary statistics, such as the first k PCs (1 ≤ k ≤ 10).
}
\author{
Ramyar Molania
}
