% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeRLE.R
\name{computeRLE}
\alias{computeRLE}
\title{calculate relative log expression (RLE) of RNA-seq data.}
\usage{
computeRLE(
  se.obj,
  assay.names = "all",
  apply.log = TRUE,
  pseudo.count = 1,
  outputs.to.return = "all",
  assess.se.obj = TRUE,
  remove.na = "assays",
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Symbol. A symbol or list of symbols for the selection of the name(s) of the assay(s) in the
SummarizedExperiment object to calculate RLE data, medians and interquartile ranges. The default is "all, which indicates all
the assays of the SummarizedExperiment object will be selected.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data or not. The default is TRUE.
Please, note, any RNA-seq data (assays) must be in log scale before computing RLE.}

\item{pseudo.count}{Numeric. A value as a pseudo count to be added to all measurements of the assay(s) before applying
log transformation to avoid -Inf for measurements that are equal to 0. The default is 1.}

\item{outputs.to.return}{Symbol. Indicates what kind of RLE computations should be performed and data should be
returned. The options are "all", "rle", "rle.med", "rle.iqr" and "rle.med.iqr". If "all" is selected, the function returns the RLE data,
medians and interquartile ranges. If "rle" is selected, the function returns only the RLE data for each assay. If "rle.med"
is selected, the function returns the only the medians of the RLE data. If "rle.iqr" is selected, the function returns
the only the interquartile ranges. of the RLE data.  If "rle.med.iqr" is selected, the function returns
both the RLE medians and interquartile ranges of the RLE data.The default is 'all'.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment object or not. See the checkSeObj
function for more details.}

\item{remove.na}{Symbol. To remove NA or missing values from the assays or not. The options are 'assays' and 'none'.
The default is "assays", so all the NA or missing values from the assay(s) will be removed before computing RLE. See
the checkSeObj function for more details.}

\item{save.se.obj}{Logical. Indicates whether to save the RLE results in the metadata of the SummarizedExperiment object or
to output the result as list. By default it is set to TRUE.}

\item{verbose}{Logical. If TRUE, displaying process messages is enabled.}
}
\value{
A SummarizedExperiment object or a list that contains all the RLE data of individual assay(s) in the
SummarizedExperiment object.
}
\description{
This function calculates relative log expression (RLE) of the assay(s) in a SummarizedExperiment object. In addition,
the function returns the RLE medians and interquartile ranges (IQRs) of each sample for individual assays.
}
\details{
RLE plots are used to reveal trends, temporal clustering and other non-random patterns resulting from unwanted variation
in gene expression data. To generate RLE plots, we first form the log ratio log(yig/yg) of the raw count yig for
gene g in the sample labeled i relative to the median value yg of the counts for gene g taken across all samples. We
then generate a box plot from all the log ratios for sample i and plotted all such box plots along a line, where i
varies in a meaningful order, usually sample processing date. An ideal RLE plot should have its medians centered around
zero, and its box widths and their interquartile ranges (IQRs) should be similar in magnitude. Because of their
sensitivity to unwanted variation, we also examine the relationships between RLE medians and interquartile ranges with
potential sources of unwanted variation and individual gene expression levels in the datasets. In the absence of any
influence of unwanted variation in the data, we should see no such associations.
}
\references{
Gandolfo L. C. & Speed, T. P., RLE plots: visualizing unwanted variation in high dimensional data. PLoS ONE, 2018.
Molania R., ..., Speed, T. P., A new normalization for Nanostring nCounter gene expression data, Nucleic Acids Research,
2019.
Molania R., ..., Speed, T. P., Removing unwanted variation from large-scale RNA sequencing data with PRPS,
Nature Biotechnology, 2023
}
\author{
Ramyar Molania
}
