% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSilhouette.R
\name{computeSilhouette}
\alias{computeSilhouette}
\title{is used to compute the Silhouette coefficient from the first PC of a SummarizedExperiment class
object given a categorical variable.}
\usage{
computeSilhouette(
  se.obj,
  assay.names = "All",
  variable,
  method = "euclidian",
  fast.pca = TRUE,
  nb.pcs = 3,
  save.se.obj = TRUE,
  plot.output = FALSE,
  assess.se.obj = TRUE,
  apply.round = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object that will be used to compute the PCA.}

\item{assay.names}{Optional string or list of strings for the selection of the name(s)
of the assay(s) of the SummarizedExperiment class object to compute the Silhouette coefficients. By default}

\item{variable}{String of the label of a categorical variable such as
sample types or batches from colData(se.obj).}

\item{method}{A character string indicating which method
is to be used for the differential analysis: 'euclidean', 'maximum', 'manhattan', 'canberra', 'binary' or 'minkowski'.
By default 'euclidean' will be selected.}

\item{fast.pca}{Logical. Indicates whether to use the PCA calculated using a specific number of PCs instead of the full range
to speed up the process, by default is set to 'TRUE'.}

\item{nb.pcs}{Numeric. The number of first PCs to be calculated for the fast pca process, by default is set to 3.}

\item{save.se.obj}{Logical. Indicates whether to save the result in the metadata of the SummarizedExperiment class object 'se.obj' or
to output the result. By default it is set to TRUE.}

\item{plot.output}{Logical. Indicates whether to plot the Silhouette coefficients, by default it is set to FALSE.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment class object.}

\item{apply.round}{Logical. Indicates whether to round the ARI results,by default it is set to TRUE.}

\item{verbose}{Logical. Indicates whether to show or reduce the level of output or messages displayed during the execution
of the functions, by default it is set to TRUE.}
}
\value{
SummarizedExperiment A SummarizedExperiment object containing the computed silhouette
on the categorical variable.
}
\description{
is used to compute the Silhouette coefficient from the first PC of a SummarizedExperiment class
object given a categorical variable.
}
