% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSilhouette.R
\name{computeSilhouette}
\alias{computeSilhouette}
\title{compute the average Silhouette coefficient.}
\usage{
computeSilhouette(
  se.obj,
  assay.names = "all",
  variable,
  dist.measure = "euclidian",
  fast.pca = TRUE,
  nb.pcs = 3,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object that will be used to compute the PCA.}

\item{assay.names}{Optional string or list of strings for the selection of the name(s)
of the assay(s) of the SummarizedExperiment class object to compute the Silhouette coefficients. By default}

\item{variable}{String of the label of a categorical variable such as
sample types or batches from colData(se.obj).}

\item{dist.measure}{A character string indicating which method
is to be used for the differential analysis: 'euclidean', 'maximum', 'manhattan', 'canberra', 'binary' or 'minkowski'.
By default 'euclidean' will be selected.}

\item{fast.pca}{Logical. Indicates whether to use the PCA calculated using a specific number of PCs instead of the
full range to speed up the process, by default is set to 'TRUE'.}

\item{nb.pcs}{Numeric. The number of first PCs to be calculated for the fast pca process, by default is set to 3.}

\item{save.se.obj}{Logical. Indicates whether to save the result in the metadata of the SummarizedExperiment class
object 'se.obj' or to output the result. By default it is set to TRUE.}

\item{verbose}{Logical. Indicates whether to show or reduce the level of output or messages displayed during the
execution of the functions, by default it is set to TRUE.}
}
\value{
SummarizedExperiment A SummarizedExperiment object containing the computed silhouette
on the categorical variable.
}
\description{
This function calculates the mean Silhouette coefficient for categorical variables such as sample subtypes, batches,
etc., and a distance matrix based on the principal components of assay(s).
}
\details{
We use silhouette coefficient analysis to assess the separation of biological populations and batch effects. The
silhouette function uses Euclidean distance to calculate both the similarity between one patient and the other patients
in each cluster and the separation between patients in different clusters. A better normalization method will lead to
higher and lower silhouette coefficients for biological and batch labels, respectively.
}
\references{
Molania R., ..., Speed, T. P., Removing unwanted variation from large-scale RNA sequencing data with PRPS,
Nature Biotechnology, 2023
}
\author{
Ramyar Molania
}
