% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSilhouette.R
\name{computeSilhouette}
\alias{computeSilhouette}
\title{Compute the average silhouette width.}
\usage{
computeSilhouette(
  se.obj,
  assay.names = "all",
  variable,
  dist.measure = "euclidian",
  fast.pca = TRUE,
  nb.pcs = 3,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Symbol. A symbol or a vector of symbols for the selection of the name(s) of the assay(s) in the
SummarizedExperiment object to calculate Silhouette coefficient. The default is set to 'all', which indicates all the
assays of the SummarizedExperiment object will be selected.}

\item{variable}{Symbol. A symbol that indicates the column name that contain categorical variable in the
SummarizedExperiment object. The variable can be biological or unwanted variable.}

\item{dist.measure}{Symbol. A symbol that indicates which ditsance measure to be used to calculate silhouette coefficient.
The options are 'euclidean', 'maximum', manhattan', 'canberra', 'binary' or 'minkowski'. The default is set to 'euclidean'.
Refer to the function 'dist' from the 'stats' R package for more details.}

\item{fast.pca}{Logical. Indicates whether to use the PC calculated using fast PCA or not. The default is set to 'TRUE'.
The fast PCA and ordinary PCA do not affect the silhouette coefficient calculation.}

\item{nb.pcs}{Numeric. A numeric value indicates the number of first PCs to be used to calculated the distances between
samples. The default is set to 3.}

\item{save.se.obj}{Logical. Indicates whether to save the result in the metadata of the SummarizedExperiment object
or to output the result. By default it is set to 'TRUE'.}

\item{verbose}{Logical. If 'TRUE', shows the messages of different steps of the function.}
}
\value{
Either the SummarizedExperiment object containing the computed silhouette coefficient in the metadata or a list
of silhouette coefficient for all specified assay(s)
}
\description{
This function calculates the average silhouette width for each categorical variables such as sample biological subtypes,
batches, etc.. The distance matrix is based on the principal components.
}
\details{
The Silhouette coefficient analysis is used to assess the separation of categorical samples groups. We use first few
principal components to calculate both the similarity between one sample and the other samples in each cluster and the
separation between samples in different clusters. A better normalization method will lead to higher and lower silhouette
coefficients for biological and batch variables, respectively.
}
\references{
Molania R., ..., Speed, T. P., Removing unwanted variation from large-scale RNA sequencing data with PRPS,
Nature Biotechnology, 2023
}
\author{
Ramyar Molania
}
