% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compuetSurvival.R
\name{computeSurvival}
\alias{computeSurvival}
\title{Generates Kaplan-Meier survival plots.}
\usage{
computeSurvival(
  se.obj,
  assay.names = "all",
  genes = "all",
  variable = NULL,
  survival.time,
  survival.events,
  expr.stratify = 2,
  return.p.values = TRUE,
  return.survival.plots = FALSE,
  plot.output = TRUE,
  assess.se.obj = TRUE,
  remove.na = "none",
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Symbol. A symbol or a vector of symbols for the selection of the name(s) of the assay(s) in the
SummarizedExperiment object to calculate RLE data, medians and interquartile ranges. The default is set to "all, which
indicates all the assays of the SummarizedExperiment object will be selected.}

\item{genes}{Symbol. A symbol or a vector of symbols for the selection of genes for gene level survival analysis. The
default is set to 'all'.}

\item{variable}{Symbol. A symbol that indicates the column name of the SummarizedExperiment object that contains
a categorical variable e.g. tumour subtypes, ... for variable-level survival analysis. The default is set to 'NULL'.}

\item{survival.time}{Symbol.A symbol that indicates the column name of the SummarizedExperiment object that contains
survival time.}

\item{survival.events}{Symbol.A symbol that indicates the column name of the SummarizedExperiment object that contains
survival event (0 and 1).}

\item{expr.stratify}{Numeric. A numeric value that indicates how the gene expression of each genes should be stratified
for gene-level survival analysis. The stratification is based on the quantiles of gene expression. The default is set
to 2. This means expression level of each specified genes will be divvied in to group based on 50\% quantiles.}

\item{return.p.values}{Logical. Indicates whether to calculate p.values in the survival analysis or not. The default
is set to 'TRUE'.}

\item{return.survival.plots}{Logical. Indicates whether to generate survival plot or not. The default
is set to 'FALSE'.}

\item{plot.output}{Logical. Indicates whether to plot variable-level survival analysis or not. The default is set to
'TRUE'.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment object or not. See the checkSeObj
function for more details.}

\item{remove.na}{Symbol. To remove NA or missing values from the assays or not. The options are 'assays' and 'none'.
The default is "assays", so all the NA or missing values from the assay(s) will be removed before computing RLE. See
the checkSeObj function for more details.}

\item{save.se.obj}{Logical. Indicates whether to save the RLE results in the metadata of the SummarizedExperiment object or
to output the result as list. By default it is set to TRUE.}

\item{verbose}{Logical. If 'TRUE', shows the messages of different steps of the function.}
}
\description{
This function computes gene-level and variable-level survival analysis.
}
\author{
Ramyar Molania
}
