% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_gene_exp_contvar.R
\name{correlation_gene_exp_contvar}
\alias{correlation_gene_exp_contvar}
\title{is used to compute the correlation between the gene expression (assay)
of a SummarizedExperiment class object and a continuous variable (i.e. library size).}
\usage{
correlation_gene_exp_contvar(
  se,
  cont_var,
  cont_var_label,
  assay_names = NULL,
  method = "spearman",
  apply.log = FALSE,
  boxplot_output = TRUE,
  ranked_genes_plot_output = FALSE,
  a = 0.05,
  rho = 0,
  nb_ranked_genes = 3
)
}
\arguments{
\item{se}{A SummarizedExperiment object that will be used to compute the correlation}

\item{cont_var}{Vector of a continous variable such as sample types
(i.e. biological subtypes) or batches.}

\item{cont_var_label}{String or vector of strings of the label of continous variable(s) such as
sample types or batches from colData(se).}

\item{assay_names}{Optional string or list of strings for selection of the names
of the assays of the SummarizedExperiment class object to compute the correlation.}

\item{method}{A character string indicating which correlation coefficient
is to be used for the test: "pearson", "kendall", or "spearman". By default 'spearman will
be selected.}

\item{apply.log}{Indicates whether to apply a log-transformation to the data. By default
no transformation will be selected.}

\item{boxplot_output}{Indicates whether to plot the boxplot of the correlation.}

\item{ranked_genes_plot_output}{Indicates whether to plot the gene expression of the number of genes
from the high or low correlation.}

\item{a}{The significance level used for the confidence intervals in the correlation,
by default it is set to 0.05.}

\item{rho}{The value of the hypothesised correlation to be used in the hypothesis testing,
by default it is set to 0.}

\item{nb_ranked_genes}{Defines the number of genes from the top or bottom listing of anova to plot,
by default is set to 3.}
}
\value{
list List containing the associated plot and the computed correlation on the continuous variable.
}
\description{
is used to compute the correlation between the gene expression (assay)
of a SummarizedExperiment class object and a continuous variable (i.e. library size).
}
