% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_gene_exp_contvar_all_assays.R
\name{correlation_gene_exp_contvar_all_assays}
\alias{correlation_gene_exp_contvar_all_assays}
\title{is used to compute the Spearman correlation between the data and library size
on all assay}
\usage{
correlation_gene_exp_contvar_all_assays(
  sce,
  corr_var,
  apply.log = FALSE,
  n.cores = 5
)
}
\arguments{
\item{sce}{the dataset that will be used for this analysis}

\item{corr_var}{The continuous variable that will be used to compute to correlation}

\item{apply.log}{Indicates whether to apply a log-transformation to the data}

\item{n.cores}{is the number of cpus used for mclapply parallelization}
}
\value{
list List containing the association plot and the computed correlation
}
\description{
is used to compute the Spearman correlation between the data and library size
on all assay
}
