% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_gene_exp_contvar_single_assay.R
\name{correlation_gene_exp_contvar_single_assay}
\alias{correlation_gene_exp_contvar_single_assay}
\title{is used to compute the correlation (Spearman or Pearson) between the data and
a continuous variable (i.e. library size)}
\usage{
correlation_gene_exp_contvar_single_assay(
  expr.data,
  apply.log,
  variable,
  method,
  n.cores
)
}
\arguments{
\item{expr.data}{is the gene expression matrix genes by samples}

\item{apply.log}{Indicates whether to apply a log-transformation to the data}

\item{variable}{is a continous variable such as library size or purity}

\item{method}{to select which method to use either Spearman or Pearson correlation}

\item{n.cores}{is the number of cpus used for mclapply parallelization}
}
\value{
dataframe containing the genes, the rho, the pvalues before and after BH correction
}
\description{
is used to compute the correlation (Spearman or Pearson) between the data and
a continuous variable (i.e. library size)
}
