% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createHomogeneousUVGroups.R
\name{createHomogeneousUVGroups}
\alias{createHomogeneousUVGroups}
\title{This is used to create all possible groups with respect to unwanted variation variables in the SummarizedExperiment objct.}
\usage{
createHomogeneousUVGroups(
  se.obj,
  uv.variables,
  nb.clusters,
  clustering.method = "kmeans",
  assess.se.obj = TRUE,
  assess.variables = TRUE,
  cat.cor.coef = c(0.7, 0.7),
  cont.cor.coef = c(0.7, 0.7),
  remove.na = "sample.annotation",
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A summarized experiment object.}

\item{uv.variables}{Symbol. Indicate of the columns names of unwanted variation variables to specify major groups
with respect to unwanted variation.}

\item{nb.clusters}{Numeric. A value to specify the number of groups for continuous sources of biological variation.
The default is 2. This means each continuous sources will be divided inot 2 groups using the 'clustering.method'.}

\item{clustering.method}{A clustering method to group each continuous sources of biological variation.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment class object.}

\item{assess.variables}{Logical. Indicates whether to assess association between the unwanted variables. For more
details refer to the 'variableCorrelation' function.}

\item{cat.cor.coef}{Numeric. Correlation coefficients (Cramer's V, Pearson's contingency coefficient) cut off for
assessing association between categorical sources of biological variation.}

\item{cont.cor.coef}{Spearman coefficients cut off for assessing association between continuous sources of biological
variation.}

\item{remove.na}{To remove NA or missing values from either the assays or sample annotation or both.}

\item{save.se.obj}{Logical. Indicates whether to save the results to SummarizedExperiment object or not.}

\item{verbose}{Logical. Whether to show the messages of the functions or not.}
}
\value{
SummarizedExperiment A SummarizedExperiment object containing the a set of sutable negatve control genes
seleced by the two-way ANOVA approach.
}
\description{
This is used to create all possible groups with respect to unwanted variation variables in the SummarizedExperiment objct.
}
