% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLogAssays.R
\name{createLogAssays}
\alias{createLogAssays}
\title{Apply log-transformation to the assay(s) of a SummarizedExperiment object.}
\usage{
createLogAssays(
  se.obj,
  assay.names = "all",
  pseudo.count = 1,
  replace.assays = FALSE,
  apply.round = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Symbol. A symbol or vector of symbols to specify the name(s) of the assay(s) in the
SummarizedExperiment object. The default is "all," indicating that all assays in the SummarizedExperiment object will
be selected.}

\item{pseudo.count}{Numeric. A value serving as a pseudo count to be added to all measurements in the assay(s) before
applying log-transformation. This helps prevent -Inf values for measurements equal to 0. The default is 1.}

\item{replace.assays}{Logical. When set to 'TRUE', the assay(s) will be substituted with the log transformation.}

\item{apply.round}{If 'TRUE', the measurements of individual assays will be rounded to two decimal points. The default
is 'TRUE'.}

\item{verbose}{Logical. If 'TRUE', shows the messages of different steps of the function.}
}
\description{
Apply log-transformation to the assay(s) of a SummarizedExperiment object.
}
\author{
Marie Trussart
}
