% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLogAssays.R
\name{createLogAssays}
\alias{createLogAssays}
\title{is used to apply a log-transformation to the assay(s) of the summarized experiment object.}
\usage{
createLogAssays(se.obj, assay.names = "All", pseudo.count = 1, verbose = TRUE)
}
\arguments{
\item{se.obj}{A summarized experiment object.}

\item{assay.names}{A name or names of the assays in summarized experiment object to apply a log-transformation to the data.}

\item{pseudo.count}{Numeric. A value as a pseudo count to be added to all measurements before log transformation,
by default it is set to 1.}

\item{verbose}{whether to show the messages or not.}
}
\description{
is used to apply a log-transformation to the assay(s) of the summarized experiment object.
}
