% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPrPsForCategoricalUV.R
\name{createPrPsForCategoricalUV}
\alias{createPrPsForCategoricalUV}
\title{Create PRPS for a categorical source of unwanted variation.}
\usage{
createPrPsForCategoricalUV(
  se.obj,
  assay.name,
  bio.variables,
  main.uv.variable,
  other.uv.variables = NULL,
  min.sample.for.ps = 3,
  bio.clustering.method = "kmeans",
  nb.bio.clusters = 2,
  other.uv.clustering.method = "kmeans",
  nb.other.uv.clusters = 2,
  check.prps.connectedness = TRUE,
  apply.log = TRUE,
  pseudo.count = 1,
  assess.se.obj = TRUE,
  remove.na = "none",
  assess.variables = FALSE,
  cat.cor.coef = c(0.95, 0.95),
  cont.cor.coef = c(0.95, 0.95),
  plot.prps.map = TRUE,
  save.se.obj = TRUE,
  prps.name = NULL,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.name}{Symbol. A symbol specifying the assay name within the SummarizedExperiment object. The selected
assay should be the one that will be used for RUV-III-PRPS normalization.}

\item{bio.variables}{Symbol. A symbol or symbols representing the label of biological variable(s), such as cancer
subtypes, tumour purity, ... within the SummarizedExperiment object. This can comprise a vector containing either
categorical, continuous, or a combination of both variables.}

\item{main.uv.variable}{Symbols. A symbol representing the label of a categorical unwanted variable, such as batch effects,
, ... within the SummarizedExperiment object.}

\item{other.uv.variables}{Symbol. A symbol or symbols representing the label of unwanted variable(s), such as batch effects,
library size, ... within the SummarizedExperiment object. This can comprise a vector containing either categorical,
continuous, or a combination of both variables. These variable(s) will be considered when generating PRPS sets fo the
main.uv.variable. The default is set to 'NULL'.}

\item{min.sample.for.ps}{Numeric. Indicates the minimum number of biologically homogeneous samples to be averaged
to create one pseudo-sample. The default it is set to 3.}

\item{bio.clustering.method}{Symbol. A symbol indicating the clustering method used to group each continuous biological
variable. The option include 'kmeans', cut' and 'quantile'. The default is to 'kmeans'. We refer to the
createHomogeneousBioGroups() function for more details.}

\item{nb.bio.clusters}{Numeric. A numeric value to specify the number of clusters/groups for each continuous biological
variable. The default it set to 3.}

\item{other.uv.clustering.method}{Symbol. A symbol indicating the clustering method used to group each continuous unwanted
variable. The option include 'kmeans', cut' and 'quantile'. The default is to 'kmeans'. We refer to the
createHomogeneousUVGroups() function for more details.}

\item{nb.other.uv.clusters}{Numeric. A numeric value to specify the number of clusters/groups for each continuous biological
variable. The default it set to 3.}

\item{check.prps.connectedness}{Logical. Indicates whether to assess the connectedness between the PRPS sets or not.
The default is set to 'TRUE'. See the details for more information.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data. The default is set to 'TRUE'.}

\item{pseudo.count}{Numeric. A numeric value as a pseudo count to be added to all measurements before log transformation.
The default is set to 1.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment object. If 'TRUE', the checkSeObj
will be applied inside the function. The default is set to 'TRUE'.}

\item{remove.na}{Symbol. A symbol Indicating whether to remove NA or missing values from either the 'assays', the
'sample.annotation', both' or 'none'. If 'assays' is selected, the genes that contains NA or missing values will be
excluded. If sample.annotation' is selected, the samples that contains NA or missing values for any 'bio.variables' and
'uv.variables' will be excluded. The default is set to both'.}

\item{assess.variables}{Logical. Indicates whether to assess the association between the biological or unwanted
variable(s) separately. The default is set to 'FALSE'. We refer to the 'assessVariableAssociation' for more details.}

\item{cat.cor.coef}{Vector of two numerical values. Indicates the cut-off of the correlation coefficient between each
pair of categorical variables. The first one is between each pair of 'uv.variables' and the second one is between each
pair of 'bio.variables'. The correlation is computed by the function ContCoef from the DescTools package. If the correlation
of a pair of variable is higher than the cut-off, then only the variable that has the highest number of factor will be
kept and the other one will be excluded from the remaining analysis. By default they are both set to 0.7.}

\item{cont.cor.coef}{Vector of two numerical values. Indicates the cut-off of the Spearman correlation coefficient
between each pair of continuous variables. The first one is between each pair of 'uv.variables' and the second one is
between each pair of 'bio.variables'. If the correlation of a pair of variable is higher than the cut-off, then only
the variable that has the highest variance will be kept and the other one will be excluded from the remaining analysis.
By default they are both set to 0.7.}

\item{plot.prps.map}{Logical. Indicates whether to generate the PRPS map plot for individual sources of unwanted variation.
The default is 'TRUE'.}

\item{save.se.obj}{Logical. Indicates whether to save the results in the metadata of the SummarizedExperiment object
or to output the result as list. The default by is set to 'TRUE'.}

\item{prps.name}{Symbol. A symbol to specify the name of all PRPS sets that will be created for all specified source(s)
of unwanted variation. The default is set to 'NULL'. The, the function creates a name based on paste0('prps_', uv.variable).}

\item{verbose}{Logical. If 'TRUE', shows the messages of different steps of the function.}
}
\value{
A SummarizedExperiment object that contains all the PRPS data and PPRS map plot in the metadata or a list
that contains all the results.
}
\description{
Distinct group of PRPS are created for each source of unwanted variation defined in the 'main.uv.variable' argument
within an homogeneous group of samples. The grouping of samples are created based on each biological subtype defined
using the 'bio.variables' using the 'bio.clustering.method' selected and it might be also be combined with
'other.uv.variables' if requested. By default 'other.uv.variables' is set up to NULL, meaning that the homogeneous
grouping of samples created is based solely on the biological subtype(s) defined in 'bio.variables'. If
'other.uv.variables' is not set to NULL, the creation of homogeneous groups of samples is based on the biological subtype
defined in 'bio.variables' combined with 'other.uv.variables' using the 'other.uv.clustering.method' selected. A
pseudo-sample will be created for each homogeneous group of samples by averaging all the samples within that group of
samples.For example to correct for batch effect arising related to a batch variable defined in the 'main.uv.variable'
argument, several group of pseudo-samples will be created, one for each homogeneous group of samples and for each batch.
All those pseudo-samples belonging to the same group across batches will be defined as pseudo-replicates which constitutes
a PRPS set.
}
\author{
Ramyar Molania
}
