% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSeObj.R
\name{createSeObj}
\alias{createSeObj}
\title{is used to create a SummarizedExperiment object.}
\usage{
createSeObj(
  assays,
  raw.count.assay.name = NULL,
  remove.lowly.expressed.genes = FALSE,
  count.cutoff = 10,
  biological.group = NULL,
  minimum.proportion = 0.5,
  calculate.library.size = FALSE,
  sample.annotation = NULL,
  create.sample.annotation = FALSE,
  gene.annotation = NULL,
  create.gene.annotation = FALSE,
  add.gene.details = FALSE,
  gene.group = NULL,
  gene.details = NULL,
  add.housekeeping.genes = FALSE,
  add.immunStroma.genes = FALSE,
  metaData = NULL,
  verbose = TRUE
)
}
\arguments{
\item{assays}{A list of assay(s) or expression data. The genes should be in the rows and samples in the columns. The
row names of the assays should be in the same order.}

\item{raw.count.assay.name}{Symbol. The name of raw counts RNA-seq data in the list of assay(s). The raw counts data
must be provided if "remove.lowly.expressed.genes" or "calculate.library.size" are "TRUE".}

\item{remove.lowly.expressed.genes}{Logical. If TRUE the function finds and removes lowly expressed genes from the
assay that is provided in the "raw.count.assay.name" argument. The default is FALSE.}

\item{count.cutoff}{Numeric. Minimum count required for at least some sample groups. If the "biological.group" argument
is equal to NULL, all samples will be considered as one group. Otherwise, the smallest subgroups of the "biological.group"
will be considered.}

\item{biological.group}{Symbol. Indicates a column name in the sample annotation that specifies the biological groups.
The smallest population is considered for removing lowly expressed genes.}

\item{minimum.proportion}{Numeric. In large sample situations, the minimum proportion of samples in a group that a
gene needs to be expressed in.}

\item{calculate.library.size}{Logical. If TRUE then library size is calculated using the raw.count.assay.name. The
library size should be calculated after removing lowly expressed genes.}

\item{sample.annotation}{A data frame, contains information for individual samples in the assay(s). The order of row
names of the sample annotation should be the same as in the assay(s).}

\item{create.sample.annotation}{Logical. If TRUE then a sample annotation the initially contains the column names of
the assay(s) will be created.}

\item{gene.annotation}{A data frame, that contains details e.g. chromosome names,GC content,... for individual genes.}

\item{create.gene.annotation}{Logical. If TRUE then a gene annotation that initially contains row names of the assay(s)
will be created.}

\item{add.gene.details}{Logical. If TRUE then a pr-set or provided gene details in the gene.details argument will be
added to the gene annotation.}

\item{gene.group}{A name of a gene class in the row names of the assay(s). This must be one of the 'entrezgene_id',
'hgnc_symbol', 'ensembl_gene_id'.}

\item{gene.details}{A vector of gene details to be added to the gene annotation.}

\item{add.housekeeping.genes}{Logical. if TRUE then several sets of publicly available "housekeeping" will be added to
the gene annotation. The housekeeping could be potentially used as negative control genes for the RUV normalization.}

\item{add.immunStroma.genes}{Logical. If TRUE, the immune and stromal genes signature from Kosuke Yoshihara et.al will
be added to the gene annotation. These gene signatures, can be used to estimate tumor purity.}

\item{metaData}{Any metadata data. The metadata can be in any format and dimensions.}

\item{verbose}{Logical. If TRUE shows the messages of different steps of the function.}
}
\value{
A summarizedExperiment object that contains assays, gene annotation, samples annotation and metadata.
}
\description{
This function creates a SummarizedExperiment object from tabular expression data and sample annotation (if available).
In addition, the function can remove lowly expressed genes if a raw count data is provided, add a range of annotations
for genes e.g. bio type, chromosome names , and provide several sets of housekeeping genes and a immunStroma gene signature.
}
\details{
The SummarizedExperiment object is a data structure used in the R for representing and manipulating high-dimensional
experimental data. Here are some key features and components of the SummarizedExperiment object:
Assays:
The SummarizedExperiment allows for the incorporation of multiple assays (data). Each assay is a separate matrix of data
associated with the same features and samples. Rows typically represent features (e.g., genes, transcripts), and columns
represent samples or experimental conditions.
Row and Column Metadata:
The SummarizedExperiment object includes metadata associated with both rows and columns. Row metadata can contain
information about the features, such as gene annotations or genomic coordinates. Column metadata may include sample
information, experimental conditions, or other relevant details.
Metadata
The Metadata of the SummarizedExperiment allows for flexibility in terms of data types and
structures. This makes it suitable for saving plots.
}
\author{
Ramyar Molania
}
