% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBioGenes.R
\name{findBioGenes}
\alias{findBioGenes}
\title{find highly variable genes in RNA-seq data.}
\usage{
findBioGenes(
  se.obj,
  assay.name,
  method = "mad",
  nb.bio.genes = 10,
  bio.variables = NULL,
  uv.variables = NULL,
  nb.pcs = 3,
  top.pca.loadings = 1,
  svd.bsparam = bsparam(),
  center.pca = TRUE,
  scale.pca = FALSE,
  normalization = "CPM",
  uv.clustering.method = "kmeans",
  nb.uv.clusters = 3,
  bio.clustering.method = "kmeans",
  nb.bio.clusters = 3,
  apply.log = TRUE,
  pseudo.count = 1,
  assess.se.obj = TRUE,
  remove.na = "both",
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.name}{TTTTT}

\item{method}{TTTTT}

\item{nb.bio.genes}{TTTTT}

\item{bio.variables}{Symbol. A symbol or a vector of symbols specifying the column names of unwanted variables in
the sample annotation of the SummarizedExperiment object. These 'uv.variables' can be either categorical or continuous
variables.}

\item{uv.variables}{Symbol. A symbol or a vector of symbols specifying the column names of unwanted variables in
the sample annotation of the SummarizedExperiment object. These 'uv.variables' can be either categorical or continuous
variables.}

\item{nb.pcs}{Numeric. A value determining the number of first principal components to be selected for clustering
when the approach is set to 'PCA'. The default is 2.}

\item{top.pca.loadings}{TTTTT}

\item{svd.bsparam}{TTTTT}

\item{center.pca}{Logical. Indicates whether to center the data or not before performing PCA. The default is 'TRUE'.}

\item{scale.pca}{Logical. Indicates whether to scale the data or not before performing PCA. The default is 'FASLE'.}

\item{normalization}{TTTTT}

\item{uv.clustering.method}{TTTTT}

\item{nb.uv.clusters}{TTTTT}

\item{bio.clustering.method}{TTTTT}

\item{nb.bio.clusters}{TTTTT}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data, by default it is set to 'TRUE'.
The data must be in log transformation before comouting RLE or PCA.}

\item{pseudo.count}{Numeric. A value as a pseudo count to be added to all measurements before log transformation,
by default it is set to 1.}

\item{assess.se.obj}{Logical. Whether to assess the SummarizedExperiment object or not. If 'TRUE', the function
'checkSeobj' will be applied. The default is 'TRUE'.}

\item{remove.na}{Symbol. Indicates whether to remove missing values from the 'uv.variables'. The options are
'sample.annotation' or 'none'. The default is 'sample.annotation', indicating the missing values from the variables
will be removed.}

\item{save.se.obj}{Logical. Indicates whether to save the results to the metadata of the SummarizedExperiment object
or not. If 'TRUE', all the possible homogeneous groups will be saved into "metadata$HGgroups$UVgroups", otherwise
the results will outputted as a vector. The default is 'TRUE'.}

\item{verbose}{Logical. If 'TRUE', shows the messages of different steps of the function.}
}
\description{
This function uses different approaches to find possible bioloical genes.
}
\author{
Ramyar Molania
}
