% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBioGenes.R
\name{findBioGenes}
\alias{findBioGenes}
\title{Perform several normalization methods for RNA-seq data.}
\usage{
findBioGenes(
  se.obj,
  assay.name,
  approach = "TwoWayAnova",
  nb.bio.genes = 0.1,
  bio.variables,
  uv.variables,
  bio.clustering.method = "kmeans",
  nb.bio.clusters = 3,
  uv.groups = NULL,
  uv.clustering.method = "kmeans",
  nb.uv.clusters = "kmeans",
  apply.log = TRUE,
  pseudo.count = 1,
  normalization = "CPM",
  regress.out.uv.variables = NULL,
  assess.bio.genes = TRUE,
  variables.to.assess.bio.genes = NULL,
  nb.pcs = 5,
  center = TRUE,
  scale = FALSE,
  anova.method = "aov",
  min.sample.for.aov = 3,
  corr.method = "spearman",
  a = 0.05,
  rho = 0,
  min.sample.for.correlation = 10,
  min.sample.for.mad = 3,
  assess.se.obj = TRUE,
  assess.variables = FALSE,
  cat.cor.coef = c(0.9, 0.9),
  cont.cor.coef = c(0.9, 0.9),
  remove.na = "none",
  save.se.obj = TRUE,
  output.name = NULL,
  save.imf = FALSE,
  imf.name = NULL,
  use.imf = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.name}{Symbol. Indicating the name of the assay in the SummarizedExperiment object. This assay should
be the one that will be used for RUV-III-PRPS normalization. We recommend to use raw data.}

\item{approach}{Symbols. Indicats which NCGs selection method should be used. The options are 'AnovaCorr.PerBatchPerBio',
'AnovaCorr.AcrossAllSamples' and'TwoWayAnova'. The default is set to 'TwoWayAnova'. Refer to details for more information.}

\item{nb.bio.genes}{Numeric. Indicating how many genes should be selected as NCG. The value represents the proportion of the
total genes in the SummarizedExperiment object. The default is set to 0.1.}

\item{bio.variables}{Symbols. Indicating the column names that contain known biological variable(s) in the
SummarizedExperiment object. These biological variables can be categorical or continuous. Continuous variables will be
divided into 'nb.bio.clusters' groups based on a clustering method selected in the 'bio.clustering.method' argument.
This argument cannot be empty.}

\item{uv.variables}{Symbols. Indicating the column names that contain unwanted variable(s) in the SummarizedExperiment
object. These unwanted variables can be categorical or continuous. Continuous variables will be divided into
'nb.uv.clusters' groups based on a clustering method selected in the 'uv.clustering.method' argument. This argument
cannot be empty.}

\item{bio.clustering.method}{Symbols. Indicating which clustering methods should be used to group continuous sources
of biological variation. Refer to the 'createHomogeneousBioGroups' function for more details. The default is set to
'kmeans' clustering .}

\item{nb.bio.clusters}{Numeric. Indicating the number of clusters for each continuous sources of biological variation.
The by default it is set to 2. This means individual continuous sources of biological variation will be divided into two
groups.}

\item{uv.groups}{TTT}

\item{uv.clustering.method}{Symbols. Indicates which clustering methods should be used to group continuous sources
of unwanted variation. Refer to the 'createHomogeneousUvGroups' function for more details. The default is set to
'kmeans' cluster.}

\item{nb.uv.clusters}{Numeric. Indicates the number of clusters for each continuous sources of unwanted variation.
The by default it is set to 2. This means individual continuous sources of biological variation will be divided into two
groups.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data before performing any statistical
analysis. The default it is set to 'TRUE'.}

\item{pseudo.count}{Numeric. A value as a pseudo count to be added to all measurements before log transformation. The
default is 1.}

\item{normalization}{Symbol. Indicates which normalization method should be applied to the data before finding genes
that are affected by biological variation. The default is set to 'CPM'. If is 'NULL', no normalization will be applied.
Refer to the 'applyOtherNormalizations' function for more details.}

\item{regress.out.uv.variables}{Symbols. Indicates the columns names that contain unwanted variation variables in the
SummarizedExperiment object. These variables will be regressed out from the data before finding genes that are highly
affected by biological variation. The default is NULL, indicates the regression will not be applied.}

\item{assess.bio.genes}{Logical. Indicates whether to assess the performance of selected NCG or not. This analysis involves
principal component analysis on only the selected NCG and then explore the R^2 or vector correlation between the 'nb.pcs'
first principal components and with the specified variables.}

\item{variables.to.assess.bio.genes}{Symbols. Indicates the column names of the SummarizedExperiment object that contain
variables whose association with the selected genes as NCG needs to be evaluated. The default is 'NULL'. This means all
the variables specified in the 'bio.variables' and 'uv.variables' will be assessed.}

\item{nb.pcs}{Numeric. Indicateing the number of the first principal components on selected NCG to be used to assess
the performance of NCGs. The default is 5.}

\item{center}{Logical. Indicates whether to center the data before applying principal component analysis or not.
Refer to the 'computePCA' function for more details. The default is set to 'TRUE'.}

\item{scale}{Logical. Indicates whether to scale the data before applying principal component analysis.
Refer to the 'computePCA' function for more details. The default is 'FALSE'.}

\item{anova.method}{Indicates which anova method should be used to compute association between gene-level}

\item{min.sample.for.aov}{Numeric. Indicates the minimum number of samples to be present in each group before applying
the ANOVA. The default is 3.}

\item{corr.method}{Symbol. Indicates which correlation method should be used to compute association between gene-level
expression and a continuous variable. The default is 'spearman'.}

\item{a}{The significance level used for the confidence intervals in the correlation, by default it is set to 0.05.}

\item{rho}{The value of the hypothesised correlation to be used in the hypothesis testing, by default it is set to 0.}

\item{min.sample.for.correlation}{Numeric. Indicates the minimum number of samples to be considered in each group before
applying the correlation analysis. The default is 10.}

\item{min.sample.for.mad}{Numeric. Indicates the minimum number of samples to perform Median Absolute Deviation on
each gene within homogeneous sample groups, considering unwanted variables. The default is 3.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment object or not. If 'TRUE', the function
'checkSeObj' will be applied. The default is set to 'TRUE'.}

\item{assess.variables}{Logical. Indicates whether to assess the correlation between biological and unwanted variation
variables separately when creating homogeneous sample groups. Refer to the function 'assessVariablesAssociation' for
more details. The default is set to 'FLASE'}

\item{cat.cor.coef}{Vector. Containing two numerical values that specifies the cut-off for the correlation coefficient
between each pair of categorical variables. The first value applies to correlations between each pair of 'uv.variables',
and the second value applies to correlations between each pair of 'bio.variables'. Correlation is computed using the
ContCoef function from the DescTools R package. If the correlation between a pair of variables exceeds the cut-off,
only the variable with the highest number of factors will be retained, and the other will be excluded from further
analysis. By default, both values are set to 0.9.}

\item{cont.cor.coef}{Vector. Containing two numerical values that specifies the cut-off for the correlation coefficient
between each pair of continuous variables. The first value applies to correlations between each pair of continuous
'uv.variables', and the second value applies to correlations between each pair of continuous 'bio.variables'.
Correlation is computed using the ContCoef function from the DescTools R package. If the correlation between a pair of
variables exceeds the cut-off, only the variable with the highest variance will be retained, and the other will be
excluded from further analysis. By default, both values are set to 0.9.}

\item{remove.na}{Symbol. Indicates whether to remove NA or missing values from either the 'assays', 'sample.annotation',
'both' or 'none'. If 'assays' is selected, the genes that contains NA or missing values will be excluded. If
'sample.annotation' is selected, the samples that contains NA or missing values for any 'bio.variables' and
'uv.variables' will be excluded. The default is set to 'none'.}

\item{save.se.obj}{Logical. Indicating whether to save the result in the metadata of the SummarizedExperiment object or
to output the result as a logical vector. The default it is set to 'TRUE'. The file will be saved in
'se.obj@metadata$NCG$supervised$output.name".}

\item{output.name}{Symbol. A representation for the output s name. If set to 'NULL', the function will choose a name
automatically. In this case, the file name will be constructed as paste0(sum(ncg.selected),'|', paste0(bio.variables,
collapse = '&'), '|', paste0(uv.variables, collapse = '&'),'|TWAnova:', '|', assay.name).}

\item{save.imf}{Logical. Indicating whether to save the intermediate file in the SummarizedExperiment object or not.
If set to 'TRUE', the function saves the results of the two-way ANOVA. Subsequently, if users wish to adjust parameters
such as 'nb.bio.genes', 'ncg.selection.method', 'top.rank.bio.genes', and 'top.rank.uv.genes', the two-way ANOVA will not
be recalculated. This accelerates parameter tuning for NCG selection. The default value is 'FALSE'.}

\item{imf.name}{Symbol. Indicating the name to use when saving the intermediate file. If set to 'NULL', the function
will create a name. In this case, the file name will be constructed as
paste0(assay.name, '|TwoWayAnova|').The default is 'NULL'.}

\item{use.imf}{Logical. Indicating whether to use the intermediate file or not. The default is set to 'FALSE'.}

\item{verbose}{Logical. If 'TRUE', shows the messages of different steps of the function.}
}
\value{
Either the SummarizedExperiment object containing the a set of biological genes in the metadata  or a
logical vector of the selected biological genes
}
\description{
Perform several normalization methods for RNA-seq data.
}
\author{
Ramyar Molania
}
