% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNcgUnSupervised.R
\name{findNcgsUnSupervised}
\alias{findNcgsUnSupervised}
\title{Find a set of negative control genes (NCG) using unsupervised approach.}
\usage{
findNcgsUnSupervised(
  se.obj,
  assay.name,
  uv.variables,
  nb.ncg = 0.1,
  ncg.selection.method = "non.overlap",
  top.rank.bio.genes = 0.5,
  top.rank.uv.genes = 0.5,
  bio.percentile = 0.2,
  uv.percentile = 0.2,
  grid.nb = 20,
  grid.group = "uv",
  grid.direction = "decrease",
  normalization = "CPM",
  regress.out.variables = NULL,
  min.sample.for.mad = 3,
  min.sample.for.aov = 3,
  min.sample.for.correlation = 10,
  corr.method = "spearman",
  a = 0.05,
  rho = 0,
  anova.method = "aov",
  clustering.method = "kmeans",
  nb.clusters = 3,
  assess.ncg = TRUE,
  variables.to.assess.ncg = NULL,
  nb.pcs = 5,
  scale = FALSE,
  center = TRUE,
  apply.log = TRUE,
  pseudo.count = 1,
  assess.se.obj = TRUE,
  assess.variable = FALSE,
  cat.cor.coef = c(0.9, 0.9),
  cont.cor.coef = c(0.9, 0.9),
  save.se.obj = TRUE,
  remove.na = "both",
  output.name = NULL,
  use.imf = FALSE,
  save.imf = FALSE,
  imf.name = NULL,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.name}{Symbol. Indicates a name of an assay in the SummarizedExperiment object. The selected assay should
be the one that will be used for the RUV-III-PRPS normalization.}

\item{uv.variables}{Symbol. A symbol or symbols indicating name(s) of the columns in the sample annotation that contains
unwanted variables in the SummarizedExperiment object. If there are unknown, the 'identifyUnknownUV' function can
estimate them.}

\item{nb.ncg}{Numeric. Specifies the number of genes to be chosen as negative control genes (NCG) when the
ncg.selection.method parameter is set to 'auto'. This value corresponds to a fraction, 'nb.ncg', of the total genes
in the SummarizedExperiment object. By default, this fraction is set to 0.1.}

\item{ncg.selection.method}{Symbol. Indicates how to select a set genes as negative control genes (NCG). For individual
genes F-statistics (from ANOVA analysis) , correlation coefficient and Median Absolute Deviation (MAD) is calculated.
Ideal control genes have should have possible high F-statistics and correlation coefficient,  and possible lowest MAD
values. To summarize these statistic there two options: 'non-overlap' and 'auto'. If 'non-overlap' is selected, the possible
non-overlap genes between 'top.rank.uv.genes' and 'top.rank.uv.genes' genes will be selected as NCG. If 'auto' selected,
the function finds possible 'nb.ncg' genes as NCGs by changes the 'top.rank.uv.genes' and 'top.rank.uv.genes' automatically.}

\item{top.rank.bio.genes}{Numeric. Indicates the fraction of top ranked genes that are highly affected by the biological
variation. These genes have highest MAD values. The default is 0.5.}

\item{top.rank.uv.genes}{Numeric. Indicates the fraction of top ranked genes that are highly affected by the unwanted
variation variables. These genes have highest F-statistics and correlation coefficients. The default is 0.5.}

\item{bio.percentile}{Numeric. The percentile cut off to select genes that are highly affected with the unwanted
variation. The default is 0.8.}

\item{uv.percentile}{Numeric. The percentile cut off to select genes that are highly affected with the unwanted
variation. The default is 0.8.}

\item{grid.nb}{Numeric. Indicates the number of genes for grid search when the 'ncg.selection.method' is auto'. The
default is 20.}

\item{grid.group}{Symbol. Indicates whether the grid search should be performed on biological, unwanted or both factors.
The options are 'bio', 'uv' or 'both'.}

\item{grid.direction}{Symbol. Indicates the grid search should be performed in decreasing or increasing order. The
option are 'increase' or 'decrease'.}

\item{normalization}{Symbols. Indicates which normalization method to use for library size adjustment before fining
genes that are highly affected by biological variation. The default is CPM. Refer to the 'applyOtherNormalization'
function for further details.}

\item{regress.out.variables}{Symbols. TTTT}

\item{min.sample.for.mad}{Numeric. Indicates the minimum number of samples to perform Median Absolute Deviation on
each gene within homogeneous sample groups, considering unwanted variables. The default is 3.}

\item{min.sample.for.aov}{Numeric. Indicates the minimum number of samples to perform ANOVA analyses between categorical
sources of variation with individual gene expression. The default is 3.}

\item{min.sample.for.correlation}{Numeric. Indicates the minimum number of samples to perform correlation analyses
between continuous sources of unwanted variation with individual gene expression. The default is 10.}

\item{corr.method}{Numeric. Indicates which correlation methods should be used for the correlation analyses. The options
are pearson' or 'spearman'. The default is 'spearman'.}

\item{a}{Numeric. the significance level used for the confidence intervals in the correlation, by default it is set to 0.05.}

\item{rho}{The value of the hypothesised correlation to be used in the hypothesis testing, by default it is set to 0.}

\item{anova.method}{Symbols. Indicates which ANOVA methods to use. The options are 'aov' or 'welch'. The default in
'aov'. Refer to the function 'computeGenesVariableAnova' for more details.}

\item{clustering.method}{Symbol. A symbol specifying the clustering method to be applied for grouping each continuous
source of unwanted variables. Options include 'kmeans', 'cut', and 'quantile'. The default is 'kmeans' clustering.}

\item{nb.clusters}{Numeric. A value indicating the number of groups for continuous sources of unwanted variation.
The default is 3. This implies that each continuous source will be split into 3 groups using the specified
'clustering.method' method.}

\item{assess.ncg}{Logical. Indicates whether to assess the performance of selected NCG or not. This involves principal
component analysis on the selected NCG and then explore the R^2 or vector correlation between the 'nb.pcs' first principal
components and with biological and unwanted variables.}

\item{variables.to.assess.ncg}{Symbols. Indicates the column names of the SummarizedExperiment object that
contain variables whose association with the selected genes as NCG needs to be evaluated. The default is NULL. This
means all the variables in the 'uv.variables' will be assessed.}

\item{nb.pcs}{Numeric. Indicates the number of the first principal components on selected NCG to be used to assess the
performance of NCGs. The defaut is 5.}

\item{scale}{Logical. Indicates whether to scale the data or not before applying SVD.  If scale is TRUE, then scaling
is done by dividing the (centered) columns of the assays by their standard deviations if center is TRUE, and the root
mean square otherwise. The default is FALSE.}

\item{center}{Logical. Indicates whether to scale the data or not before applying SVD. If 'center' is TRUE, then
centering is done by subtracting the column means of the assay from their corresponding columns. The default is TRUE.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data, by default it is set to TRUE.}

\item{pseudo.count}{Numeric. A value as a pseudo count to be added to all measurements before log transformation.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment class object or not.}

\item{assess.variable}{Logical.}

\item{cat.cor.coef}{Vector. Containing two numerical values that specifies the cut-off for the correlation coefficient
between each pair of categorical variables. The first value applies to correlations between each pair of 'uv.variables',
and the second value applies to correlations between each pair of 'bio.variables'. Correlation is computed using the
ContCoef function from the DescTools R package. If the correlation between a pair of variables exceeds the cut-off,
only the variable with the highest number of factors will be retained, and the other will be excluded from further
analysis. By default, both values are set to 0.9.}

\item{cont.cor.coef}{Vector. Containing two numerical values that specifies the cut-off for the correlation coefficient
between each pair of continuous variables. The first value applies to correlations between each pair of continuous
'uv.variables', and the second value applies to correlations between each pair of continuous 'bio.variables'.
Correlation is computed using the ContCoef function from the DescTools R package. If the correlation between a pair of
variables exceeds the cut-off, only the variable with the highest variance will be retained, and the other will be
excluded from further analysis. By default, both values are set to 0.9.}

\item{save.se.obj}{Logical. Indicates whether to save the result of the function in the metadata of the SummarizedExperiment object or
to output the result. The default is TRUE.}

\item{remove.na}{Symbol. Indicates whether to remove NA or missing values from either the 'assays', the 'sample.annotation',
'both' or 'none'. If 'assays' is selected, the genes that contains NA or missing values will be excluded. If 'sample.annotation' is selected, the
samples that contains NA or missing values for any 'uv.variables' will be excluded. By default, it is set to both'.}

\item{output.name}{Symbol. Indicates the name of the output in the meta data of the SummarizedExperiment object. The
default is 'NULL'. This means the function will create a name based the specified argument.}

\item{use.imf}{Logical. Indicates whether to use the intermediate file or not. The default is set to 'FALSE'.}

\item{save.imf}{Logical. Indicates whether to save the intermediate file or not. If 'TRUE', the function save the results
of two-way ANOVA. Then, if users want to change the parameters including 'nb.ncg', 'ncg.selection.method',
'top.rank.bio.genes' and 'top.rank.uv.genes', the two-way ANOVA will not be re-calculated.}

\item{imf.name}{Symbol. A name to save the intermediate file. If is 'NULL', the function creates a name.}

\item{verbose}{Logical. If TRUE, displaying process messages is enabled.}
}
\value{
Either the SummarizedExperiment object containing the a set of negative control genes or a logical vector of
the selected negative control genes
}
\description{
This function identifies a set of genes as negative control genes in cases where no biological variations are known.
The function uses the gene-level correlation, ANOVA and MAD  analyses across and between sample groups to find a set
of genes as negative control genes for the RUV-III-PRPS normalization.
}
\details{
The function initially utilizes gene-level correlation and ANOVA to identify genes significantly influenced by continuous
and categorical sources of variation, respectively. Subsequently, it conducts Median Absolute Deviation (MAD) analysis
on each gene within homogeneous sample groups, considering unwanted variables, to pinpoint genes highly variable due
to biological factors. Lastly, various methods are employed to consolidate the statistical findings, ultimately determining
an appropriate set of genes as negative control genes for RUV-III-PRPS normalization.
}
\author{
Ramyar Molania
}
