% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findRepeatingPatterns.R
\name{findRepeatingPatterns}
\alias{findRepeatingPatterns}
\title{find repeating factors ro characters in a vector.}
\usage{
findRepeatingPatterns(vec, n.repeat)
}
\arguments{
\item{vec}{A vector of factors or characters.}

\item{n.repeat}{Numeric. Indicates the minimum repeat of individual factors ro characters in the vector.}
}
\value{
A vec of factors ro characters that are repeated at least "n.repeat" times.
}
\description{
find repeating factors ro characters in a vector.
}
