% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesDEA.R
\name{genesDEA}
\alias{genesDEA}
\title{is used to perform differential gene expression analysis using Wilcoxon test.}
\usage{
genesDEA(
  se.obj,
  assay.names = "All",
  variable,
  apply.log = TRUE,
  pseudo.count = 1,
  assess.se.obj = TRUE,
  remove.na = "both",
  save.se.obj = TRUE,
  plot.output = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Symbol. Symbol or list of symbols for the selection of the name(s) of the assay(s) of the
SummarizedExperiment object to compute the differential gene expression analysis. By default all the assays of the
SummarizedExperiment object will be selected.}

\item{variable}{Symbol. Indicates the columns of the SummarizedExperiment object that contains a categorical variable
such as batches.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data. By default
the log transformation will be selected.}

\item{pseudo.count}{Numeric. A value as a pseudo count to be added to all measurements before log transformation,
by default it is set to 1.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment class object.}

\item{remove.na}{Symbol. Indicates whether to remove missing/NA values from either the 'assays', 'sample.annotation',
'both' or 'none'. If 'assays' is selected, the genes that contains missing/NA values will be excluded. If 'sample.annotation'
is selected, the samples that contains NA or missing values for each 'variables' will be excluded. By default, it is
set to 'both'.}

\item{save.se.obj}{Logical. Indicates whether to save the result in the metadata of the SummarizedExperiment class
object 'se.obj' or to output the result. By default it is set to TRUE.}

\item{plot.output}{Logical. Indicates whether to plot the boxplot of the F-test statistics, by default it is set to TRUE.}

\item{verbose}{Logical. Indicates whether to show or reduce the level of output or messages displayed during the execution
of the functions, by default it is set to TRUE.}
}
\value{
A SummarizedExperiment object or a list containing all the stats of the Wilcoxon test and if requested the
associated p-value histograms.
}
\description{
This function performs differential gene expression analysis using Wilcoxon test between all possible pairs of a
categorical variable in a SummarizedExperiment object.
}
\details{
DE analyses were performed using the Wilcoxon signed-rank test with log -transformed raw counts and normalized data.
To evaluate the effects 2 of the different sources of unwanted variation on the data, DE analyses were performed across
batches. In the absence of any batch effects, the histogram of the resulting unadjusted P values should
}
\author{
Ramyar Molania
}
