% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAssessmentMetrics.R
\name{getAssessmentMetrics}
\alias{getAssessmentMetrics}
\title{is used to find repeating factors ro characters in a vector.}
\usage{
getAssessmentMetrics(se.obj, variables)
}
\arguments{
\item{se.obj}{A vector of factors or characters.}

\item{variables}{Numeric. Indicates the minimum repeat of individual factors ro characters in the vector.}
}
\value{
A vec of factors ro characters that are repeated at least "n.repeat" times.
}
\description{
is used to find repeating factors ro characters in a vector.
}
