% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_unwanted_variation.R
\name{identify_unwanted_variation}
\alias{identify_unwanted_variation}
\title{is used to identify the unwanted variation of a SummarizedExperiment class object.}
\usage{
identify_unwanted_variation(
  se,
  assay_names = NULL,
  apply.log = FALSE,
  cat_var_label = NULL,
  cont_var_label = NULL,
  output_file = NULL,
  n.cores = 5
)
}
\arguments{
\item{se}{A SummarizedExperiment object that will be used to assess the performance of the normalisation of the data.}

\item{assay_names}{String or list of strings for selection of the name
of the assay of the SummarizedExperiment class object.}

\item{apply.log}{Indicates whether to apply a log-transformation to the data. By default
no transformation will be selected.}

\item{cat_var_label}{String or vector of strings of the label of categorical variable(s) such as
sample types or batches from colData(se).}

\item{cont_var_label}{String or vector of strings of the label of continuous variable(s)
such as library size from colData(se).}

\item{output_file}{Path and name of the output file to save the assessments plots in a pdf format.}

\item{n.cores}{is the number of cpus used for mclapply parallelization. Default is set to 5.}
}
\value{
list List of assessments plots and metrics used for the assessment
}
\description{
Several assessment will be performed:
\enumerate{
\item For each categorical variable:
}
\itemize{
\item PCA plot of the categorical variable.
\item Silhouette and ARI computed on the categorical variable.
\item Differential analysis based ANOVA between the gene expression and the categorical variable.
\item Vector correlation between the first cumulative PCs of the gene expression and the categorical variable.
It will output the following plots:
\item PCA plot of each categorical variable.
\item Boxplot of the F-test distribution from ANOVA between the gene expression and each categorical variable.
\item Vector correlation between the first cumulative PCs of the gene expression and each categorical variable.
\item Combined Silhouette plot of the combined pair of all categorical variables.
}
}
\details{
\enumerate{
\item For each continous variable:
}
\itemize{
\item Linear regression between the first cumulative PC and continuous variable.
\item Correlation between gene expression and continuous variable.
It will output the following plots:
\item Linear regression between the first cumulative PC and continuous variable.
\item Boxplot of the correlation between gene expression and continuous variable.
}

It will also output the RLE plot distribution.
}
