% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_assessment.R
\name{norm_assessment}
\alias{norm_assessment}
\title{is used to assess the performance of the normalisation of a SummarizedExperiment class object.
It will first generate two PCA plots: one colored by biology and another one by batch,
each plot will display PCA plot for each assay in the following order: raw, fpkm, fpkm.uq and ruvprps.
It will also compute the regression between library size and PCs,}
\usage{
norm_assessment(
  se,
  apply.log = FALSE,
  biological_subtypes,
  library_size,
  batch,
  output_file = NULL,
  catvar_da_library_size = NULL,
  n.cores = 5
)
}
\arguments{
\item{se}{Dataset that will be used to assess the performance of the normalisation of the data.}

\item{apply.log}{Indicates whether to apply a log-transformation to the data. By default
no transformation will be selected.}

\item{biological_subtypes}{Vector containing the biological subtypes of each sample}

\item{library_size}{Vector containing the library size of each sample}

\item{batch}{Vector containing the batch (plates/years) of each sample}

\item{output_file}{Path and name of the output file to save the assessments plots in a pdf format}

\item{catvar_da_library_size}{Vector containing the categorical variable of high vs low library size of each sample to compute differential analysis}

\item{n.cores}{is the number of cpus used for mclapply parallelization}
}
\value{
plots List of assessments plots
}
\description{
is used to assess the performance of the normalisation of a SummarizedExperiment class object.
It will first generate two PCA plots: one colored by biology and another one by batch,
each plot will display PCA plot for each assay in the following order: raw, fpkm, fpkm.uq and ruvprps.
It will also compute the regression between library size and PCs,
}
