% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{normalise}
\alias{normalise}
\title{is used to normalise the gene expression (assay) of a SummarizedExperiment class object
using RUVIII-PRPS method.}
\usage{
normalise(
  se.obj,
  assay.name = NULL,
  apply.log = TRUE,
  pseudo.count = 1,
  bio.variable.prps,
  uv.variables.prps,
  batch.variable.prps,
  min.sample.prps = 3,
  min.sample.per.batch.prps = 12,
  assess.cor.variables.prps = TRUE,
  norm.assay.name.ncg,
  apply.normalization.ncg = FALSE,
  normalization.ncg = "CPM",
  bio.variables.ncg,
  uv.variables.ncg,
  no.ncg = 1000,
  regress.out.uv.variables.ncg = FALSE,
  regress.out.bio.variables.ncg = FALSE,
  k = NULL,
  eta = NULL,
  include.intercept = TRUE,
  apply.average.rep = FALSE,
  fullalpha = NULL,
  inputcheck = TRUE,
  assess.se.obj = TRUE,
  remove.na = "both",
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object that will be used to computer fastRUV-III}

\item{assay.name}{String for the selection of the name of the assay data
of the SummarizedExperiment class object}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data, by default it is set to TRUE.}

\item{pseudo.count}{Numeric. A value as a pseudo count to be added to all measurements before log transformation,
by default it is set to 1.}

\item{bio.variable.prps}{String of the label of a categorical variable that specifies major biological groups
such as samples types from colData(se) that will be used to define PRPS.}

\item{uv.variables.prps}{String or vector of strings of the label of continuous or categorical variable(s)
such as samples types, batch or library size from colData(se) that will be used to define PRPS.}

\item{batch.variable.prps}{String of the label of a categorical variable that specifies major batch groups
such as plates from colData(se).}

\item{min.sample.prps}{TO BE DEFINED that will be used to define PRPS.}

\item{min.sample.per.batch.prps}{TO BE DEFINED.}

\item{assess.cor.variables.prps}{Logical. Indicates whether to assess the assess the association between variables
using Spearman correlation to define PRPS.}

\item{norm.assay.name.ncg}{String for the selection of the name of the assay of the SummarizedExperiment class object to use
to define NCG. If you don't provide any assay, we recommend to set apply.normalization to TRUE.}

\item{apply.normalization.ncg}{Logical Indicates whether to apply a normalization method to define NCG if the 'norm.assay.name.ncg'
wasn't provided. By default it is set to FALSE.}

\item{normalization.ncg}{String defining the normalization method to use from 'CPM', 'TMM', 'upper', 'full', 'median', 'VST',
and 'Quantile'to define NCG. By default it is set to 'CPM'.}

\item{bio.variables.ncg}{String or vector of strings of the label of a categorical variable that specifies major biological groups
such as samples types from colData(se) that will be used to find the negative controls.}

\item{uv.variables.ncg}{String or vector of strings of the label of continuous or categorical variable(s)
such as samples types, batch or library size from colData(se) that will be used to find the negative controls.}

\item{no.ncg}{Logical, TO BE BETTER DEFINED. if TRUE then a sample annotation the initially contains column names of the assays.???}

\item{regress.out.uv.variables.ncg}{TO BE DEFINED.}

\item{regress.out.bio.variables.ncg}{TO BE DEFINED.}

\item{k}{A single value or a vector of values containing a single k or a range of k - the number of unwanted factors - to be tested.}

\item{eta}{A matrix with n columns, gene-wise (as opposed to sample-wise) covariates. By default is set to NULL.}

\item{include.intercept}{Logical. Add an intercept term to eta if it does not include one already. By default is set to TRUE.}

\item{apply.average.rep}{Average replicates after adjustment. By default is set to FALSE.}

\item{fullalpha}{Can be included to speed up execution. By default is set to NULL.}

\item{inputcheck}{Perform a basic sanity check on the inputs, and issue a warning if there is a problem.
By default is set to TRUE.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment class object.
By default it is set to TRUE.}

\item{remove.na}{TO BE DEFINED.}

\item{verbose}{Logical. Indicates whether to show or reduce the level of output or messages displayed during the execution
of the functions, by default it is set to TRUE.}
}
\value{
SummarizedExperiment A SummarizedExperiment object containing the normalised gene expression
into a new assay (RUVIII_K), as well as the PRPS created and the negative control genes
}
\description{
The steps involves:
\itemize{
\item Creation of the Pseudo-Replicates of Pseudo-Samples (PRPS)
\item Define Negative Controls Genes (NCG)
\item Run RUVIII-PRPS for multiple k values (the dimension of the unwanted variation).
}
}
