% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overrideCheck.R
\name{overrideCheck}
\alias{overrideCheck}
\title{Check whether to override a current metric or not.}
\usage{
overrideCheck(
  se.obj,
  slot = "Metrics",
  assay.names,
  assessment.type,
  assessment,
  method,
  variable,
  file.name,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{slot}{Symbol. A symbol indicating the name of the slots in the 'metadata' of the SummarizedExperiment object.}

\item{assay.names}{Symbol. A symbol or a vector of symbols specifying the name(s) of the assay(s) in the SummarizedExperiment
object. The default is set to 'all'.}

\item{assessment.type}{Symbol. A symbol indicating the type of assessment. Options are 'gene.level' or 'global.level'.}

\item{assessment}{Symbol. A symbol indicating the name of the metric to be checked.}

\item{method}{Symbol. A symbol indicating the method used to calculate the metric.}

\item{variable}{Symbol. A symbol indicating the variable used to calculate the metric.}

\item{file.name}{Symbol. A symbol indicating the file name to which the results of the metric are assigned.}

\item{verbose}{Logical. If 'TRUE', shows the messages of different steps of the function.}
}
\description{
This function checks whether to re-calculate the specified metric or not.
}
\author{
Ramyar Molania
}
