% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCVariableRegression.R
\name{PCVariableRegression}
\alias{PCVariableRegression}
\title{is used to compute the linear regression between the the first cumulative PCs
of the gene expression (assay) of a SummarizedExperiment class object and a continuous variable (i.e. library size)}
\usage{
PCVariableRegression(
  se.obj,
  assay.names = "All",
  variable,
  fast.pca = TRUE,
  nb.pcs = 10,
  save.se.obj = TRUE,
  plot.output = TRUE,
  assess.se.obj = TRUE,
  remove.na = "both",
  apply.round = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object that will be used to compute the PCA.}

\item{assay.names}{Optional string or list of strings for the selection of the name(s)
of the assay(s) of the SummarizedExperiment class object to compute the regression. By default}

\item{variable}{String of the label of a continuous variable such as
library size from colData(se.obj).}

\item{fast.pca}{Logical. Indicates whether to use the PCA calculated using a specific number of PCs instead of the full range
to speed up the process, by default is set to 'TRUE'.}

\item{nb.pcs}{Numeric. The number of few first cumulative PCs, by default is set to 10.}

\item{save.se.obj}{Logical. Indicates whether to save the result in the metadata of the SummarizedExperiment class object 'se.obj' or
to output the result. By default it is set to TRUE.}

\item{plot.output}{Logical. Indicates whether to plot the regression statistics, by default it is set to TRUE.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment class object.}

\item{remove.na}{TO BE DEFINED.}

\item{apply.round}{Logical. Indicates whether to round the ARI results, by default it is set to TRUE.}

\item{verbose}{Logical. Indicates whether to show or reduce the level of output or messages displayed during the execution
of the functions, by default it is set to TRUE.}
}
\value{
SummarizedExperiment A SummarizedExperiment object containing the computed regression for
the continuous variable and if requested the associated plot.
}
\description{
is used to compute the linear regression between the the first cumulative PCs
of the gene expression (assay) of a SummarizedExperiment class object and a continuous variable (i.e. library size)
}
