% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca}
\alias{plot_pca}
\title{is used to plot the pairwise plots of the first PCA of the gene expression (assay)
of a SummarizedExperiment class object.}
\usage{
plot_pca(
  pca,
  assay_names = NULL,
  cat_var,
  cat_var_label,
  nPCs = 3,
  ncol_plot = 4,
  color,
  strokeSize = 0.2,
  pointSize = 1.5,
  strokeColor = "gray30",
  alpha = 0.5
)
}
\arguments{
\item{pca}{PCA components of a SummarizedExperiment variable that will be used in the plot.}

\item{assay_names}{Optional selection of names of the assays to compute the PCA.}

\item{cat_var}{Vector of a categorical variable such as sample types
(i.e. biological subtypes) or batches.}

\item{cat_var_label}{String or vector of strings of the label of categorical variable(s) such as
sample types or batches from colData(se).}

\item{nPCs}{is the number of PCs used to measure the distance, by default it is set to 3.}

\item{ncol_plot}{is the argument of gtable for the layout specifying ncol, by default it is set to 4.}

\item{color}{The color of the variable that will be used on the PCA plot}

\item{strokeSize}{geom_point aesthetics}

\item{pointSize}{geom_point aesthetics}

\item{strokeColor}{geom_point aesthetics}

\item{alpha}{geom_point aesthetics}
}
\value{
p PCA plot of the data colored by one variable
}
\description{
is used to plot the pairwise plots of the first PCA of the gene expression (assay)
of a SummarizedExperiment class object.
}
