% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotARI.R
\name{plotARI}
\alias{plotARI}
\title{is used to plot the computed the adjusted rand index (ARI).}
\usage{
plotARI(
  se.obj,
  assay.names = "all",
  variables,
  plot.type = "single.plot",
  ari.method = "hclust.complete.euclidian",
  plot.output = TRUE,
  save.se.obj = TRUE,
  assess.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Symbol. A symbol or list of symbols for the selection of the name(s) of the assay(s) in the
SummarizedExperiment object to compute PCA. By default all the assays of the SummarizedExperiment object will be selected.}

\item{variables}{Symbol. Indicates the column name in the SummarizedExperiment object that contains a categorical
variable such as sample types or batches.}

\item{plot.type}{Symbol.Indicates what plot types to plot.}

\item{ari.method}{Symbol.Indicates what plot types to plot.}

\item{plot.output}{Logical. Indicates whether to plot the ARI, by default it is set to FALSE.}

\item{save.se.obj}{Logical. Indicates whether to save the result in the metadata of the SummarizedExperiment class
object 'se.obj' or to output the result. By default it is set to TRUE.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment class object, by default it is
set to TRUE.}

\item{verbose}{Logical. Indicates whether to show or reduce the level of output or messages displayed during the
execution of the functions, by default it is set to TRUE.}
}
\value{
A SummarizedExperiment object or a list that containing the computed ARI on the categorical variable.
}
\description{
This functions computes the adjusted rand index for given a categorical variable using the first PCs of the assay(s)
in a SummarizedExperiment object.
}
\author{
Ramyar Molania
}
