% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDGE.R
\name{plotDGE}
\alias{plotDGE}
\title{plot p-values histograms of DEG analysis.}
\usage{
plotDGE(
  se.obj,
  assay.names = "all",
  variable,
  plot.ncol = 1,
  plot.output = TRUE,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Symbol. A symbol or vector of symbols for the selection of the name(s) of the assay(s) in the
SummarizedExperiment object to compute PCA. By default all the assays of the SummarizedExperiment object will be selected.}

\item{variable}{Symbol. Indicates the column name in the SummarizedExperiment object that contains a categorical
variable such as sample types or batches.}

\item{plot.ncol}{Numeric. Indicates number of columns in the plot grid.}

\item{plot.output}{Logical. Indicates whether to plot the ARI, by default it is set to FALSE.}

\item{save.se.obj}{Logical. Indicates whether to save the result in the metadata of the SummarizedExperiment class
object 'se.obj' or to output the result. By default it is set to TRUE.}

\item{verbose}{Logical. Indicates whether to show or reduce the level of output or messages displayed during the
execution of the functions, by default it is set to TRUE.}
}
\value{
A SummarizedExperiment object or a list that containing the computed ARI on the categorical variable.
}
\description{
This function plots the p-values histograms of the DEG analysis.
}
\details{
DE analyses is performed using the Wilcoxon signed-rank test with log-transformed data e.g. raw counts, normalized data, ....
To evaluate the effects of the different sources of unwanted variation on the data, DE analyses is performed across
batches. In the absence of any batch effects, the histogram of the resulting unadjusted P values should be uniformly
distributed
}
\author{
Ramyar Molania
}
