% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneSetScore.R
\name{plotGeneSetScore}
\alias{plotGeneSetScore}
\title{Plot gene set score of a gene set.}
\usage{
plotGeneSetScore(
  se.obj,
  assay.names = "all",
  reference.score = NULL,
  gene.set.name,
  plot.output = TRUE,
  assess.se.obj = TRUE,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Symbol or a vector of symbols specifying the name(s) of the assay(s) in the SummarizedExperiment
object to calculate sample-wise scores. The default is set to "all," indicating all assays in the SummarizedExperiment
object will be selected.}

\item{reference.score}{Symbol.}

\item{gene.set.name}{Symbol. A symbol indicating the name to be used to save the score in the SummarizedExperiment object.
if 'NULL', the function will select a name as follow :
gene.set.name <- paste0('singscore|',length(c(upset.genes, downset.genes)),'genes')}

\item{plot.output}{Logical. Whether to plot the gene set score or not. The default is set to 'TRUE'.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment object or not. The default is set
to 'TRUE'. Refer to the 'checkSeObj' function for more details.}

\item{save.se.obj}{Indicates whether to save the score results in the metadata of the SummarizedExperiment object
or to output the result as list. The default is set to 'TRUE'.}

\item{verbose}{Logical. If 'TRUE', shows the messages of different steps of the function.}
}
\description{
This function computes all possible gene-gene pairwise ordinary and partial correlation of the assays in the
SummarizedExperiment object.
}
\details{
Partial correlation is used to estimate correlation between two variables while controlling for third
variables
}
\author{
Ramyar Molania
}
