% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneVariableAnova.R
\name{plotGenesVariableAnova}
\alias{plotGenesVariableAnova}
\title{is used to compute the adjusted rand index (ARI).}
\usage{
plotGenesVariableAnova(
  se.obj,
  assay.names = "all",
  variable,
  anova.method = "genes.aov.anova",
  plot.output = TRUE,
  boxplot.color = "black",
  geom.hline.color = "gray",
  assess.se.obj = TRUE,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Symbol. A symbol or list of symbols for the selection of the name(s) of the assay(s) in the
SummarizedExperiment object to compute PCA. By default all the assays of the SummarizedExperiment object will be selected.}

\item{variable}{Symbol. Indicates the column name in the SummarizedExperiment object that contains a categorical
variable such as sample types or batches.}

\item{anova.method}{Logical. Indicates whether to use the PCA calculated using a specific number of PCs instead of the
full range to speed up the process, by default is set to 'TRUE'.}

\item{plot.output}{Logical. Indicates whether to plot the ARI, by default it is set to FALSE.}

\item{boxplot.color}{TTTTTT}

\item{geom.hline.color}{TTTTTT}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment class object, by default it is
set to TRUE.}

\item{save.se.obj}{Logical. Indicates whether to save the result in the metadata of the SummarizedExperiment class
object 'se.obj' or to output the result. By default it is set to TRUE.}

\item{verbose}{Logical. Indicates whether to show or reduce the level of output or messages displayed during the
execution of the functions, by default it is set to TRUE.}
}
\value{
A SummarizedExperiment object or a list that containing the computed ARI on the categorical variable.
}
\description{
This functions computes the adjusted rand index for given a categorical variable using the first PCs of the assay(s)
in a SummarizedExperiment object.
}
\details{
The ARI64 is the corrected-for-chance version of the Rand index. The ARI measures the percentage of matches between
two label lists. We used the ARI to assess the performance of normalization methods in terms of sample subtype
separation and batch mixing. We first calculated PCs and used the first three PCs to perform ARI.
}
\author{
Ramyar Molania
}
