% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneVariableAnova.R
\name{plotGenesVariableAnova}
\alias{plotGenesVariableAnova}
\title{is used to plot F-stattistics of ANOVA.}
\usage{
plotGenesVariableAnova(
  se.obj,
  assay.names = "all",
  variable,
  anova.method = "genes.aov.anova",
  plot.output = TRUE,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Symbol. A symbol or list of symbols for the selection of the name(s) of the assay(s) in the
SummarizedExperiment object to compute PCA. By default all the assays of the SummarizedExperiment object will be selected.}

\item{variable}{Symbol. Indicates the column name in the SummarizedExperiment object that contains a categorical
variable such as sample types or batches.}

\item{anova.method}{Logical. Indicates whether to use the PCA calculated using a specific number of PCs instead of the
full range to speed up the process, by default is set to 'TRUE'.}

\item{plot.output}{Logical. Indicates whether to plot the ARI, by default it is set to FALSE.}

\item{save.se.obj}{Logical. Indicates whether to save the result in the metadata of the SummarizedExperiment class
object 'se.obj' or to output the result. By default it is set to TRUE.}

\item{verbose}{Logical. If TRUE, displaying process messages is enabled.}
}
\value{
A SummarizedExperiment object or a list that containing the computed ARI on the categorical variable.
}
\description{
This functions computes the adjusted rand index for given a categorical variable using the first PCs of the assay(s)
in a SummarizedExperiment object.
}
\author{
Ramyar Molania
}
