% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneVariableCorrelation.R
\name{plotGenesVariableCorrelation}
\alias{plotGenesVariableCorrelation}
\title{Plot Spearman or Pearson correlations coefficients.}
\usage{
plotGenesVariableCorrelation(
  se.obj,
  assay.names = "all",
  variable,
  correlation.method = "gene.spearman.corr",
  plot.output = TRUE,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Symbol. A symbol or vector of symbols for the selection of the name(s) of the assay(s) of the
SummarizedExperiment object to compute the correlation. By default all the assays of the SummarizedExperiment class
object will be selected.}

\item{variable}{Symbol. Indicates a name in the columns of the sample annotation in the SummarizedExperiment object
that contains a continuous variable such as library size, tumor purity, ... .}

\item{correlation.method}{Symbol. Indicates which computed correlation coefficient should be used for plotting. The
default is 'gene.spearman.corr'. We refer to the 'computeGenesVariableCorrelation' function for more details.}

\item{plot.output}{Logical. Indicates whether to plot the ARI, by default it is set to FALSE.}

\item{save.se.obj}{Logical. Indicates whether to save the result in the metadata of the SummarizedExperiment class
object 'se.obj' or to output the result. By default it is set to TRUE.}

\item{verbose}{Logical. If TRUE, displaying process messages is enabled.}
}
\value{
A SummarizedExperiment object or a list that containing the boxplots of the Spearman or Pearson correlations
coefficients for individual assays.
}
\description{
This function generates boxplots of computed Spearman or Pearson correlations coefficients of indivdial assays in a
SummarizedExperiment object
}
\author{
Ramyar Molania
}
