% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCVariableRegression.R
\name{plotPCVariableRegression}
\alias{plotPCVariableRegression}
\title{is used to plot the regression analysis.}
\usage{
plotPCVariableRegression(
  se.obj,
  assay.names = "all",
  variable,
  fast.pca = TRUE,
  nb.pcs = 10,
  save.se.obj = TRUE,
  plot.output = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Optional string or list of strings for the selection of the name(s)
of the assay(s) of the SummarizedExperiment class object to compute the correlation. By default}

\item{variable}{String of the label of a categorical variable such as
sample types or batches from colData(se.obj).}

\item{fast.pca}{Logical. Indicates whether to use the PCA calculated using a specific number of PCs instead of the
full range to speed up the process, by default is set to 'TRUE'.}

\item{nb.pcs}{Numeric. The number of few first cumulative PCs, by default is set to 10.}

\item{save.se.obj}{Logical. Indicates whether to save the result in the metadata of the SummarizedExperiment class
object 'se.obj' or to output the result. By default it is set to TRUE.}

\item{plot.output}{Logical. Indicates whether to plot the correlation statistics, by default it is set to TRUE.}

\item{verbose}{Logical. If TRUE, displaying process messages is enabled.}
}
\value{
A SummarizedExperiment object containing the computed correlation for
the continuous variable and if requested the associated plot.
}
\description{
This function generate a dot-line plot between the first cumulative PCs and the regression R squared obtained
from the regression analysis. An ideal normalization should results a low R squared with unwanted variation
variables and high R squared with known biology.
}
\author{
Ramyar Molania
}
