% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRLE.R
\name{plotRLE}
\alias{plotRLE}
\title{is used to plot the RLE (Relative Log Expression) of a SummarizedExperiment class object.}
\usage{
plotRLE(
  se.obj,
  assay.names = "All",
  apply.log = FALSE,
  pseudo.count = 1,
  save.se.obj = TRUE,
  assess.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object that will be used to plot the RLE.}

\item{assay.names}{Optional string or list of strings for the selection of the name(s)
of the assay(s) of the SummarizedExperiment class object to plot RLE. By default}

\item{apply.log}{Indicates whether to apply a log-transformation to the data. By default
no transformation will be selected.}

\item{pseudo.count}{Numeric. A value as a pseudo count to be added to all measurements before log transformation,
by default it is set to 1.}

\item{save.se.obj}{Indicates whether to save the result in the metadata of the SummarizedExperiment class object 'se.obj' or
to output the result. By default it is set to TRUE.}

\item{assess.se.obj}{Indicates whether to assess the SummarizedExperiment class object.}

\item{verbose}{Indicates whether to show or reduce the level of output or messages displayed during the execution
of the functions, by default it is set to TRUE.}
}
\value{
list List of the computed RLE and the associated plot
}
\description{
is used to plot the RLE (Relative Log Expression) of a SummarizedExperiment class object.
}
