% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRleVariable.R
\name{plotRleVariable}
\alias{plotRleVariable}
\title{is used to plot relative log expression (RLE) of RNA-seq data.}
\usage{
plotRleVariable(
  se.obj,
  assay.names = "all",
  variable,
  rle.data.type = "both",
  ylim.rle.med.plot = NULL,
  ylim.rle.iqr.plot = NULL,
  points.size = 1,
  plot.ncol = 1,
  plot.output = TRUE,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Symbol. A symbol or list of symbols for the selection of the name(s) of the assay(s) in the
SummarizedExperiment object to plot RLE. By default all the assays of the SummarizedExperiment object
will be selected.}

\item{variable}{Symbol. Indicates a name of column in the sample annotation for the SummarizedExperiment object to color
the boxplots of the RLE plots. The variable should be a categorical variable. The default is NULL.}

\item{rle.data.type}{TTTTT}

\item{ylim.rle.med.plot}{TTTTTTTT}

\item{ylim.rle.iqr.plot}{Numeric. A vector of two values to specify the ylim of the RLE plot.}

\item{points.size}{Numeric. Indicates the size of the points of the RLE medians in the plot.}

\item{plot.ncol}{Numeric. Indicates number of columns in the plot grid.}

\item{plot.output}{Logical. If TRUE, individual RLE plot(s) will be printed.}

\item{save.se.obj}{Logical. Indicates whether to save the result in the metadata of the SummarizedExperiment object or
to output the result. By default it is set to TRUE.}

\item{verbose}{Logical. If TRUE, displaying process messages is enabled.}
}
\value{
A SummarizedExperiment object or a list that contains all the RLE plot(s).
}
\description{
This functions generate a boxplot of individual RLE data. The RLE data can be obtained using the computeRLE function.
}
\author{
Ramyar Molania
}
