% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStudyOutline.R
\name{plotStudyOutline}
\alias{plotStudyOutline}
\title{Generates study outline of a SummarizedExperiment object.}
\usage{
plotStudyOutline(
  se.obj,
  variables,
  sort.variable = NULL,
  assess.se.obj = TRUE,
  remove.na = "none",
  plot.output = TRUE,
  legend.font.size = 14,
  legend.ncol = 4,
  legend.direction = "horizontal",
  heatmap.legend.side = "bottom",
  column.names.rot = 25,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{variables}{Symbol. A symbol or symbols representing the label of variable(s) within the SummarizedExperiment
object. This can comprise a vector containing either categorical, continuous, or a combination of both variables.}

\item{sort.variable}{Symbol. A symbol indicating the label of variable that is used to sort the sample information.
The default is 'NULL'.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment object or not. See the 'checkSeObj()'
function for more details.}

\item{remove.na}{Symbol. To remove NA or missing values from the assays or not. The options are 'assays' and 'none'.
The default is "assays", so all the NA or missing values from the assay(s) will be removed before computing RLE. See
the checkSeObj function for more details.}

\item{plot.output}{Logical. Determines whether to plot the study outline. The default is set to 'TRUE'.}

\item{legend.font.size}{Numeric. A numeric value indicating the size of font in the plot.}

\item{legend.ncol}{Numeric. A numeric value indicating the numer of columns of the legend of the heatmap. The default is set
to 4.}

\item{legend.direction}{Symbol. A symbol specifying the direction of the legend in the heatmap. The default is set to
'horizontal'.}

\item{heatmap.legend.side}{Symbol. A symbol indicating the side of the legend in the heatmap.  The default is set to
'bottom'.}

\item{column.names.rot}{Numeric. A numeric value showing the angel of the columns of the heatmap. The default is set to
25.}

\item{save.se.obj}{Logical. Indicates whether to save the study outline plot in the metadata of the SummarizedExperiment
object or to output the result as a plot. The default is set to 'TRUE'.}

\item{verbose}{Logical. If 'TRUE', shows the messages of different steps of the function.}
}
\description{
This function generates a heatmap showing the study outline.
}
\author{
Ramyar Molania
}
