% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prpsForContinuousUV.R
\name{prpsForContinuousUV}
\alias{prpsForContinuousUV}
\title{is used to create PRPS for a continuous sources of unwanted variation.}
\usage{
prpsForContinuousUV(
  se.obj,
  assay.name,
  apply.log = TRUE,
  pseudo.count = 1,
  uv.variable,
  bio.variable,
  min.sample.prps = 3,
  assess.se.obj = TRUE,
  remove.na = "both",
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A summarized experiment object.}

\item{assay.name}{A name of the assays in summarized experiment object.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data.}

\item{pseudo.count}{Numeric. A pseudo count to add to each gene before applying log.}

\item{uv.variable}{String of the label of a continuous or categorical variable.}

\item{bio.variable}{String of the label of a categorical variable that specifies major biological groups
such as samples types from colData(se).}

\item{min.sample.prps}{Numeric. The minimum number of homogeneous biological groups to create pseudo-sample.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment class object.}

\item{remove.na}{TO BE DEFINED.}

\item{save.se.obj}{Logical. Indicates whether to save the result in the metadata of the SummarizedExperiment class
object 'se.obj' or to output the result, by default it is set to TRUE.}

\item{verbose}{Logical. Indicates whether to show or reduce the level of output or
messages displayed during the execution of the functions, by default it is set to TRUE.}
}
\value{
SummarizedExperiment A SummarizedExperiment object containing the PRPS data or just PRPS data.
}
\description{
is used to create PRPS for a continuous sources of unwanted variation.
}
