% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_pc.R
\name{regression_pc}
\alias{regression_pc}
\title{is used to compute the linear regression of a variable to the PCA of the data}
\usage{
regression_pc(pca, normalization, regression_var, nb_pca_comp = 10)
}
\arguments{
\item{pca}{PCs of the dataset that will be used in the plot}

\item{normalization}{All the available assays for the data (i.e. normalizations methods)}

\item{regression_var}{The regression variable that will be computed to the PCA of the data (library size)}

\item{nb_pca_comp}{The number of components of the PCA used to compute the regression}
}
\value{
list List containing the association plot and the computed regression
}
\description{
is used to compute the linear regression of a variable to the PCA of the data
}
