% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUVIII.PRPS.R
\name{RUVIII.PRPS}
\alias{RUVIII.PRPS}
\title{RUVIII-PRPS normalization.}
\usage{
RUVIII.PRPS(
  se.obj,
  assay.name,
  prps.group = "supervised",
  prps.set.names = NULL,
  ncg.group = "supervised",
  ncg.set.names = NULL,
  k = 1,
  apply.log = "assay",
  pseudo.count = 1,
  alpha = NULL,
  eta = NULL,
  include.intercept = TRUE,
  return.info = FALSE,
  assess.se.obj = TRUE,
  remove.na = "none",
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object that will be used to computer fastRUV-III}

\item{assay.name}{String for the selection of the name of the assay data of the SummarizedExperiment class object}

\item{prps.group}{TTTTT}

\item{prps.set.names}{TTTTT}

\item{ncg.group}{TTTT}

\item{ncg.set.names}{TTTT}

\item{k}{The number of unwanted factors to use. Can be 0, in which case no adjustment is made. Can also be NULL (the
default value), in which case the maximum possible value of k is used; note that in this case no singular value
decomposition is necessary and execution is faster.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data, by default it is set to TRUE.}

\item{pseudo.count}{Numeric. A value as a pseudo count to be added to all measurements before log transformation,
by default it is set to 1.}

\item{alpha}{TTTT}

\item{eta}{Gene-wise (as opposed to sample-wise) covariates. These covariates are adjusted for by RUV-1 before any
further analysis proceeds. Can be either (1) a matrix with n columns, (2) a matrix with n rows, (3) a dataframe with
n rows, (4) a vector or factor of length n, or (5) simply 1, for an intercept term.}

\item{include.intercept}{When eta is specified (not NULL) but does not already include an intercept term, this will
automatically include one.}

\item{return.info}{If FALSE, only the adjusted data matrix is returned. If TRUE, additional information is returned
(see below).}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment object or not. See the checkSeObj
function for more details.}

\item{remove.na}{Symbol. To remove NA or missing values from the assays or not. The options are 'assays' and 'none'.
The default is "assays", so all the NA or missing values from the assay(s) will be removed before computing RLE. See
the checkSeObj function for more details.}

\item{save.se.obj}{Logical. Indicates whether to save the RUV-III-PRPS normalized data as assay(s) in the
SummarizedExperiment object or to output the result as list. By default it is set to 'TRUE'.}

\item{verbose}{Logical. If 'TRUE', shows the messages of different steps of the function.}
}
\value{
SummarizedExperiment A SummarizedExperiment object containing the normalised gene expression
into a new assay (RUVIII_K).
}
\description{
This function applies the RUV-III-PRPS on RNA-seq data.
}
\details{
Additional details...
}
\references{
Molania R., ..., Speed, T. P., A new normalization for Nanostring nCounter gene expression data, Nucleic Acids Research,
2019.
Molania R., ..., Speed, T. P., Removing unwanted variation from large-scale RNA sequencing data with PRPS,
Nature Biotechnology, 2023
}
\author{
Ramyar Molania
}
