% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUVIII.PRPS.R
\name{RUVIII.PRPS}
\alias{RUVIII.PRPS}
\title{is used to run the RUVIII-PRPS method for a given assay and given k.}
\usage{
RUVIII.PRPS(
  se.obj,
  assay.name,
  prps.data = NULL,
  prps.group = "supervised",
  ncg.sets = NULL,
  ncg.approach = "supervised",
  k = 1,
  apply.log = "both",
  pseudo.count = 1,
  alpha = NULL,
  eta = NULL,
  include.intercept = TRUE,
  return.info = FALSE,
  assess.se.obj = TRUE,
  remove.na = "measurements",
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object that will be used to computer fastRUV-III}

\item{assay.name}{String for the selection of the name of the assay data of the SummarizedExperiment class object}

\item{prps.data}{TTTTT}

\item{prps.group}{TTTTT}

\item{ncg.sets}{TTTT}

\item{ncg.approach}{TTTT}

\item{k}{TTTT}

\item{apply.log}{TTTT}

\item{pseudo.count}{TTTT}

\item{alpha}{TTTT}

\item{eta}{TTTT}

\item{include.intercept}{TTTT}

\item{return.info}{TTTT}

\item{assess.se.obj}{TTTT}

\item{remove.na}{TTTT}

\item{save.se.obj}{TTTT}

\item{verbose}{TTTT}
}
\value{
SummarizedExperiment A SummarizedExperiment object containing the normalised gene expression
into a new assay (RUVIII_K).
}
\description{
is used to run the RUVIII-PRPS method for a given assay and given k.
}
\author{
Ramyar Molania
}
