% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUVIII.PRPS.R
\name{RUVIII.PRPS}
\alias{RUVIII.PRPS}
\title{RUVIII with technical replicates or PRPS normalization.}
\usage{
RUVIII.PRPS(
  se.obj,
  assay.name,
  prps.group = "supervised",
  prps.set.names = NULL,
  ncg.group = "supervised",
  ncg.set.names = NULL,
  technical.replicate = NULL,
  k = 1,
  apply.log = "assay",
  pseudo.count = 1,
  eta = NULL,
  include.intercept = TRUE,
  return.info = FALSE,
  assess.se.obj = TRUE,
  remove.na = "none",
  output.name = NULL,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.name}{Symbol. Indicates a name of an assay in the SummarizedExperiment object. The selected assay should
be the one that was used for PRPS and NCGs selection.}

\item{prps.group}{Symbol. Denotes which groups of the generated PRPS should be utilized for normalization. The
options include 'supervised,' 'unsupervised,' or 'both.' When 'supervised' is chosen, the function employs the PRPS
set(s) generated through supervised functions. Opting for 'unsupervised' utilizes the PRPS set(s) created through
unsupervised functions. If 'both' is selected, the function utilizes the PRPS set(s) generated by both supervised and
unsupervised functions. The default setting is 'supervised.'}

\item{prps.set.names}{Symbol. sample names/ids or a name of column in sample annotation.}

\item{ncg.group}{Symbol. Denotes which groups of the selected NCGs should be utilized for normalization. The
options include 'supervised,' 'unsupervised,' or 'both.' When 'supervised' is chosen, the function employs the PRPS
set(s) generated through supervised functions. Opting for 'unsupervised' utilizes the PRPS set(s) created through
unsupervised functions. If 'both' is selected, the function utilizes the PRPS set(s) generated by both supervised and
unsupervised functions. The default setting is 'supervised.'}

\item{ncg.set.names}{Symbol. A symbol or symbols that indicate which specif set(s) of NCG should be used for normalization.
If is 'NULL', the function uses all available NCGs of the specified group by "ncg.group".}

\item{k}{The number of unwanted factors to use. Can be 0, in which case no adjustment is made. Can also be NULL (the
default value), in which case the maximum possible value of k is used; note that in this case no singular value
decomposition is necessary and execution is faster.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data, by default it is set to TRUE.}

\item{pseudo.count}{Numeric. A value as a pseudo count to be added to all measurements before log transformation,
by default it is set to 1.}

\item{eta}{Gene-wise (as opposed to sample-wise) covariates. These covariates are adjusted for by RUV-1 before any
further analysis proceeds. Can be either (1) a matrix with n columns, (2) a matrix with n rows, (3) a dataframe with
n rows, (4) a vector or factor of length n, or (5) simply 1, for an intercept term.}

\item{include.intercept}{When eta is specified (not NULL) but does not already include an intercept term, this will
automatically include one.}

\item{return.info}{If FALSE, only the adjusted data matrix is returned. If TRUE, additional information is returned
(see below).}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment object or not. See the checkSeObj
function for more details.}

\item{remove.na}{Symbol. To remove NA or missing values from the assays or not. The options are 'assays' and 'none'.
The default is "assays", so all the NA or missing values from the assay(s) will be removed before computing RLE. See
the checkSeObj function for more details.}

\item{save.se.obj}{Logical. Indicates whether to save the RUV-III-PRPS normalized data as assay(s) in the
SummarizedExperiment object or to output the result as list. By default it is set to 'TRUE'.}

\item{verbose}{Logical. If 'TRUE', shows the messages of different steps of the function.}

\item{technical.replicates}{Symbol. Indicates}
}
\value{
SummarizedExperiment A SummarizedExperiment object containing the normalised gene expression
into a new assay (RUVIII_K).
}
\description{
This function applies the RUV-III with technical replicates or PRPS on transcriptions data.
}
\details{
RUV-III normalization. Before we can describe the linear model underlying RUV-III, we need to introduce the m × m1
mapping matrix M connecting assays to distinct samples, which captures the pattern of replication in our assays. Here,
m is the number of assays, and m1 is the number of distinct samples being assayed. M(i,h) = 1if assay i is on sample
h and M(i,h) = 0 otherwise. Each row of M sums to 1, and the columns sum to the distinct sample replication numbers, the
elements of MTM. We also define an m1 × p design matrix X to capture the biological factor(s) of interest indexed by
sample rather than assay. There are no constraints on p; indeed, X could be the m1 × m1 identity matrix. Our goal
here is to remove unwanted variation, not to estimate regression parameters.
The linear model we use is: Y = 1μ + MXβ + Wα + ε
where the data Y = (yij)and unobserved errors ε = (εij)are m × n; the matrices X and M have just been defined; μ is
the 1 × n row of gene means; β is p × n; the matrix W whose columns capture the unwanted variation is m × k; and α
is k × n. 1 = 1m is the m × 1 column vector of 1s. Here, m = number of assays, n = number of genes and p is the
dimension of the wanted variation X and k that of the unwanted variation W. Assume that W⊥1. Also, we suppose that we
have a subset of nc of negative control genes whose m × nc submatrix Yc satisfies Yc = 1μc + Wαc + εc, where we have
assumed that βc = 0—that is, that there is no true association between these genes and the biology of interest.
The projection PM = M(M M) M replaces each entry yij of Y by the simple average of the entries yi′j over all i′ for
which M (i, h) = M (i′ , h) = 1—that is, over all i′ such that i′ and i label replicate assays of the same unique
sample (or pseudo-sample) labeled h. Write RM = I − PM for the corresponding residual projector. This is our source
of information on the unwanted variation that we will remove. If the replication is technical at some level, then RMY
mainly contains information about unwanted variation in the system after the technical replicates were created.
Depending on the study details, technical replicates could be created immediately beforethe assay was run, in parallel
with or immediately after sample was collected or somewhere in between. The earlier the creation of technical replicates,
the more unwanted variation will be captured in their differences. The use of pseudo-replicates of suitable pseudo-samples
enables us to start to deal with pre-technical unwanted variation. Write the spectral decomposition of RMYY RM = UDU ,
where U is an m × m orthogonal matrix and D is an m × n diagonal matrix with entries ordered from largest to smallest
eigenvalue. Let P1 be the orthogonal projection onto 1m. For a chosen k, 1 ≤ k ≤ m−m1,
I- Define αˆ(k) = U(k)TY, where U(k) is the first k columns of U
II- Estimate W by regressing the centered negative controls (I − P1)Yc on αˆ(k)T c
III-
}
\references{
Molania R., ..., Speed, T. P., A new normalization for Nanostring nCounter gene expression data, Nucleic Acids Research,
2019.
Molania R., ..., Speed, T. P., Removing unwanted variation from large-scale RNA sequencing data with PRPS,
Nature Biotechnology, 2023
}
\author{
Ramyar Molania
}
