% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervisedFindNCG.R
\name{supervisedFindNCG}
\alias{supervisedFindNCG}
\title{is used to find a set of negative control genes of a SummarizedExperiment class object
using RUVIII-PRPS method with a supervised approach using the 'bio.variables' and
'uv.variables' given.}
\usage{
supervisedFindNCG(
  se.obj,
  assay.name,
  bio.variables,
  uv.variables,
  no.ncg = 1000,
  apply.log = TRUE,
  pseudo.count = 1,
  regress.out.uv.variables = FALSE,
  regress.out.bio.variables = FALSE,
  apply.normalization = FALSE,
  normalization = "CPM",
  corr.method = "pearson",
  a = 0.05,
  rho = 0,
  anova.method = "aov",
  assess.se.obj = TRUE,
  assess.variables = FALSE,
  remove.na = "both",
  plot.output = TRUE,
  fast.pca = TRUE,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A summarized experiment object.}

\item{assay.name}{String for the selection of the name of the assay of the SummarizedExperiment class object used to define NCG.
If you provide the raw data assay, we recommend to set apply.normalization to TRUE.}

\item{bio.variables}{String or vector of strings of the label of a categorical variable that specifies major biological groups
such as samples types from colData(se) that will be used to find the negative controls.}

\item{uv.variables}{String or vector of strings of the label of continuous or categorical variable(s)
such as samples types, batch or library size from colData(se) that will be used to find the negative controls.}

\item{no.ncg}{Logical, TO BE BETTER DEFINED. if TRUE then a sample annotation the initially contains column names of the assays.???}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data.}

\item{pseudo.count}{Numeric. A value as a pseudo count to be added to all measurements before log transformation,
by default it is set to 1.}

\item{regress.out.uv.variables}{TO BE DEFINED.}

\item{regress.out.bio.variables}{TO BE DEFINED.}

\item{apply.normalization}{Logical Indicates whether to apply a normalization method when providing the raw data assay.
By default it is set to FALSE.}

\item{normalization}{String defining the normalization method to use from 'CPM', 'TMM', 'upper', 'full', 'median', 'VST',
and 'Quantile'. By default it is set to 'CPM'.}

\item{corr.method}{TO BE DEFINED.}

\item{a}{TO BE DEFINED.}

\item{rho}{TO BE DEFINED.}

\item{anova.method}{TO BE DEFINED.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment class object.
By default it is set to TRUE.}

\item{assess.variables}{Logical. TO BE DEFINED.}

\item{remove.na}{TO BE DEFINED.}

\item{plot.output}{Logical. Indicates whether to plot the PCA of the defined NCG colored by the categorical variables, by default it is set to TRUE.}

\item{fast.pca}{Logical. Indicates whether to calculate a specific number of PCs instead of the full range
to speed up the process, by default is set to 'TRUE'.}

\item{save.se.obj}{Logical. Indicates whether to save the result in the metadata of the SummarizedExperiment
class object 'se.obj' or to output the result, by default it is set to TRUE.}

\item{verbose}{Logical. Indicates whether to show or reduce the level of output or messages displayed
during the execution of the functions, by default it is set to TRUE.}
}
\value{
SummarizedExperiment or a List A SummarizedExperiment object or a list containing
all the negative control genes defined.
}
\description{
is used to find a set of negative control genes of a SummarizedExperiment class object
using RUVIII-PRPS method with a supervised approach using the 'bio.variables' and
'uv.variables' given.
}
