% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervisedPRPS.R
\name{supervisedPRPS}
\alias{supervisedPRPS}
\title{is used to create pseudo-replicates of pseudo samples of a SummarizedExperiment class object
using RUVIII-PRPS method with a supervised approach using the 'uv.variables' and 'bio.variable' given.}
\usage{
supervisedPRPS(
  se.obj,
  assay.name,
  bio.variable,
  uv.variables,
  batch.variable,
  min.sample.prps = 3,
  min.sample.per.batch = 10,
  apply.log = TRUE,
  pseudo.count = 1,
  assess.se.obj = TRUE,
  assess.cor.variables = FALSE,
  cont.coef = c(0.7, 0.7),
  spearman.coef = c(0.7, 0.7),
  remove.na = "both",
  save.se.obj = TRUE,
  plot.output = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A summarized experiment object.}

\item{assay.name}{String for the selection of the name of the assay of the SummarizedExperiment class object used to define PRPS.
We recommend to use the raw data assay.}

\item{bio.variable}{String of the label of a categorical variable that specifies major biological groups
such as samples types from colData(se).}

\item{uv.variables}{String or vector of strings of the label of continuous or categorical variable(s)
such as samples types, batch or library size from colData(se) that will be used to define PRPS.}

\item{batch.variable}{String of the label of a categorical variable that specifies major batch groups
such as plates from colData(se).}

\item{min.sample.prps}{TO BE DEFINED.}

\item{min.sample.per.batch}{TO BE DEFINED.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data, by default it is set to TRUE.}

\item{pseudo.count}{Numeric. A value as a pseudo count to be added to all measurements before log transformation,
by default it is set to 1.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment class object.
By default it is set to TRUE.}

\item{assess.cor.variables}{Logical. Indicates whether to assess the assess the association between variables
using Spearman correlation.}

\item{cont.coef}{TO BE DEFINED.}

\item{spearman.coef}{TO BE DEFINED.}

\item{remove.na}{TO BE DEFINED.}

\item{save.se.obj}{Logical. Indicates whether to save the result in the metadata of the SummarizedExperiment
class object 'se.obj' or to output the result, by default it is set to TRUE.}

\item{plot.output}{Logical. Indicates whether to plot the PRPS map define by the 'bio.variable'
colored by the categorical variables, by default it is set to TRUE.}

\item{verbose}{Logical. Indicates whether to show or reduce the level of output or messages displayed
during the execution of the functions, by default it is set to TRUE.}
}
\value{
SummarizedExperiment or a List A SummarizedExperiment object or a list containing
all the prps defined.
}
\description{
We will create distinct group of pseudo-replicates for each source of unwanted variation defined in the 'uv.variables' argument.
For example to correct for batch effect if defined in the 'uv.variables' argument, several group of pseudo-samples
will be created by averaging the samples of the same biological subtype defined in 'bio.variable' in each batch. Then those
pseudo-samples will be defined as pseudo-replicates.
For example to correct for library size if defined in the 'uv.variables' argument, several group of pseudo-samples
will be created by averaging the top and bottom-ranked samples by library size of the same biological subtype in each batch.
Then those pseudo-samples will be defined as pseudo-replicates.
Similarly to correct for purity if defined in the 'purity' argument, several group of pseudo-samples
will be created by averaging the top and bottom-ranked samples by purity of the same biological subtype in each batch.
Then those pseudo-samples will be defined as pseudo-replicates.
}
