% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervisedPRPS.R
\name{supervisedPRPS}
\alias{supervisedPRPS}
\title{is used to create pseudo-replicates of pseudo samples of a SummarizedExperiment class object
using RUVIII-PRPS method with a supervised approach using the 'uv.variables', 'bio.variable' and 'batch.variable' given.}
\usage{
supervisedPRPS(
  se.obj,
  assay.name,
  bio.variable,
  uv.variables,
  batch.variable,
  min.sample.for.prps = 3,
  min.sample.per.batch = 6,
  apply.log = TRUE,
  pseudo.count = 1,
  assess.se.obj = TRUE,
  assess.cor.variables = FALSE,
  cat.cor.coef = c(0.85, 0.85),
  cont.cor.coef = c(0.85, 0.85),
  remove.na = "both",
  save.se.obj = TRUE,
  plot.output = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object that will be used to create PRPS.}

\item{assay.name}{String for the selection of the name of the assay of the SummarizedExperiment class object.}

\item{bio.variable}{String of the label of a categorical variable that specifies major biological groups
such as samples types from colData(se).}

\item{uv.variables}{String or vector of strings of the label of continuous or categorical variable(s)
such as samples types, batch or library size from colData(se) that will be used to define PRPS.}

\item{batch.variable}{String of the label of a categorical variable that specifies major batch groups
such as plates from colData(se).}

\item{min.sample.for.prps}{Numeric. Indicates the minimum number of samples to create one pseudo-sample,
by default it is set to 3.}

\item{min.sample.per.batch}{Numeric. Indicates the minimum number of homogeneous biological samples within each batch
to create a PRPS set for each continuous variable. The minimum should be '2*min.sample.for.prps'. By default it is set to 6.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data, by default it is set to TRUE.}

\item{pseudo.count}{Numeric. A value as a pseudo count to be added to all measurements before log transformation,
by default it is set to 1.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment class object.
By default it is set to TRUE.}

\item{assess.cor.variables}{Logical. Indicates whether to assess the association between pairs of categorical variables
and/or pairs of continuous variables in order to select only one of the variable of a pair of highly correlated
variables.}

\item{cat.cor.coef}{Vector of two numerical values. Indicates the cut-off of the correlation coefficient between each pair of
categorical variables. The first one is between each pair of 'uv.variables' and the second one is between each pair of 'bio.variables'.
The correlation is computed by the function ContCoef from the DescTools package. If the correlation of a pair of variable is higher than
the cut-off, then only the variable that has the highest number of factor will be kept and the other one will be excluded from the
remaining analysis. By default they are both set to 0.85.}

\item{cont.cor.coef}{Vector of two numerical values. Indicates the cut-off of the Spearman correlation coefficient between each pair of
continuous variables. The first one is between each pair of 'uv.variables' and the second one is between each pair of 'bio.variables'.
If the correlation of a pair of variable is higher than the cut-off, then only the variable that has the highest variance will
be kept and the other one will be excluded from the remaining analysis. By default they are both set to 0.85.}

\item{remove.na}{String. Indicates whether to remove NA or missing values from either the 'assays', the 'sample.annotation',
'both' or 'none'. If 'assays' is selected, the genes that contains NA or missing values will be excluded. If 'sample.annotation' is selected, the
samples that contains NA or missing values for any 'bio.variables' and 'uv.variables' will be excluded. By default, it is set to
'both'.}

\item{save.se.obj}{Logical. Indicates whether to save the result in the metadata of the SummarizedExperiment
class object 'se.obj' or to output the result, by default it is set to TRUE.}

\item{plot.output}{Logical. Indicates whether to plot the PRPS map of the PRPS defined by the 'bio.variable'
and the categorical variables, by default it is set to TRUE.}

\item{verbose}{Logical. Indicates whether to show or reduce the level of output or messages displayed
during the execution of the functions, by default it is set to TRUE.}
}
\value{
SummarizedExperiment or a List A SummarizedExperiment object or a list containing all the PRPS created.
}
\description{
We will create distinct group of pseudo-replicates for each source of unwanted variation defined in the 'uv.variables' argument.
For example to correct for batch effect if defined in the 'uv.variables' argument, several group of pseudo-samples
will be created by averaging the samples of the same biological subtype defined in 'bio.variable' in each batch. Then those
pseudo-samples will be defined as pseudo-replicates which constitutes a PRPS set.
For example to correct for library size if defined in the 'uv.variables' argument, several group of pseudo-samples
will be created by averaging the top and bottom-ranked samples by library size of the same biological subtype in each batch.
Then those pseudo-samples will be defined as pseudo-replicates which constitutes a PRPS set.
Similarly to correct for purity if defined in the 'purity' argument, several group of pseudo-samples
will be created by averaging the top and bottom-ranked samples by purity of the same biological subtype in each batch.
Then those pseudo-samples will be defined as pseudo-replicates which constitutes a PRPS set.
}
