% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variablesCorrelation.R
\name{variablesCorrelation}
\alias{variablesCorrelation}
\title{is used to assess the association between variables of a SummarizedExperiment object using Spearman correlation.}
\usage{
variablesCorrelation(
  se.obj,
  assay.name = NULL,
  bio.variables,
  uv.variables,
  cont.coef = c(0.7, 0.7),
  spearman.coef = c(0.7, 0.7),
  assess.se.obj = TRUE,
  remove.na = "sample.annotation",
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A summarized experiment object.}

\item{assay.name}{String for the selection of the name of the assay of the SummarizedExperiment class object used to define PRPS.}

\item{bio.variables}{String of the label of (a) categorical variable(s) that specifies major biological groups
such as samples types from colData(se).}

\item{uv.variables}{String or vector of strings of the label of continuous or categorical variable(s)
such as samples types, batch or library size from colData(se) that will be used to define PRPS.}

\item{cont.coef}{TO BE DEFINED.}

\item{spearman.coef}{Vector of two numeric values. Each value is the cut-off to consider there is an association
between the variables based on the Spearman correlation coefficient. The first one is used for the 'uv.variables' and the second for 'bio.variables'.}

\item{assess.se.obj}{Logical. Whether to assess the SummarizedExperiment object or not.}

\item{remove.na}{To remove NA or missing values from either the assays or sample annotation or both.}

\item{verbose}{Logical. Whether to show the messages of the functions or not.}
}
\value{
a SummarizedExperiment object and the selected 'uv.variables' and 'bio.variables'.
}
\description{
is used to assess the association between variables of a SummarizedExperiment object using Spearman correlation.
}
